Extract forecasts from a summarised posterior

Extract forecasts from a summarised posterior

extract_forecast(posterior, forecast_dates = NULL)

Arguments

posterior
A list of posterior output as produced by summarise_posterior().

forecast_dates
A named vector of dates to use to identify when output is a forecast. Must contain a "Cases" date and a "Sequence" date. Default is to infer these dates from the summarised posterior.

Value

A list containing a forecast for each parameter

Examples

if (FALSE) {
obs <- latest_obs(germany_covid19_delta_obs)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
p <- summarise_posterior(fit)
extract_forecast(p)
}