Filter data based on availability and forecast date

Filter data based on availability and forecast date

filter_by_availability(
  obs,
  date = max(obs$date),
  seq_date = date,
  case_date = date
)

Arguments

obs
A data frame with the following variables: date, cases, seq_voc, and seq_total, cases_available, and seq_available. seq_available and case_available must be uniquely define data rows but other rows can be duplicated based on data availability.

date
Date at which to filter. Defaults to the maximum date in obs.

seq_date
Date from which to use available sequence data. Defaults to the forecast date.

case_date
Date from which to use available case data. Defaults to the forecast date.

Examples

latest_dt <- latest_obs(germany_covid19_delta_obs)
dt <- rbind(
  update_obs_availability(latest_dt, seq_lag = 3),
  update_obs_availability(latest_dt, seq_lag = 1)
)
# filter out duplicates and up to the present date
filter_by_availability(dt)
#>           date location_name location  cases cases_available seq_total seq_voc
#>  1: 2021-03-20       Germany       DE  87328      2021-03-20        NA      NA
#>  2: 2021-03-27       Germany       DE 109442      2021-03-27        NA      NA
#>  3: 2021-04-03       Germany       DE 117965      2021-04-03        NA      NA
#>  4: 2021-04-10       Germany       DE 107223      2021-04-10        NA      NA
#>  5: 2021-04-17       Germany       DE 142664      2021-04-17      4108       6
#>  6: 2021-04-24       Germany       DE 145568      2021-04-24      4539      34
#>  7: 2021-05-01       Germany       DE 131887      2021-05-01      3667      54
#>  8: 2021-05-08       Germany       DE 107141      2021-05-08      4547      89
#>  9: 2021-05-15       Germany       DE  77261      2021-05-15      3688     101
#> 10: 2021-05-22       Germany       DE  57310      2021-05-22      3817     119
#> 11: 2021-05-29       Germany       DE  33052      2021-05-29      2781     102
#> 12: 2021-06-05       Germany       DE  22631      2021-06-05      2197     181
#> 13: 2021-06-12       Germany       DE  15553      2021-06-12      1566     309
#> 14: 2021-06-19       Germany       DE   7659      2021-06-19       991     414
#> 15: 2021-06-26       Germany       DE   5033      2021-06-26       706     451
#> 16: 2021-07-03       Germany       DE   4181      2021-07-03       803     631
#> 17: 2021-07-10       Germany       DE   5067      2021-07-10       735     643
#> 18: 2021-07-17       Germany       DE   7969      2021-07-17       564     521
#> 19: 2021-07-24       Germany       DE  11122      2021-07-24        NA      NA
#> 20: 2021-07-31       Germany       DE  14654      2021-07-31        NA      NA
#> 21: 2021-08-07       Germany       DE  18474      2021-08-07        NA      NA
#> 22: 2021-08-14       Germany       DE  28646      2021-08-14        NA      NA
#> 23: 2021-08-21       Germany       DE  44862      2021-08-21        NA      NA
#> 24: 2021-08-28       Germany       DE  62984      2021-08-28        NA      NA
#> 25: 2021-09-04       Germany       DE  71057      2021-09-04        NA      NA
#> 26: 2021-09-11       Germany       DE  75107      2021-09-11        NA      NA
#>           date location_name location  cases cases_available seq_total seq_voc
#>       share_voc seq_available
#>  1:          NA          <NA>
#>  2:          NA          <NA>
#>  3:          NA          <NA>
#>  4:          NA          <NA>
#>  5: 0.001460565    2021-05-08
#>  6: 0.007490637    2021-05-15
#>  7: 0.014725934    2021-05-22
#>  8: 0.019573345    2021-05-29
#>  9: 0.027386117    2021-06-05
#> 10: 0.031176316    2021-06-12
#> 11: 0.036677454    2021-06-19
#> 12: 0.082385071    2021-06-26
#> 13: 0.197318008    2021-07-03
#> 14: 0.417759839    2021-07-10
#> 15: 0.638810198    2021-07-17
#> 16: 0.785803238    2021-07-24
#> 17: 0.874829932    2021-07-31
#> 18: 0.923758865    2021-08-07
#> 19:          NA          <NA>
#> 20:          NA          <NA>
#> 21:          NA          <NA>
#> 22:          NA          <NA>
#> 23:          NA          <NA>
#> 24:          NA          <NA>
#> 25:          NA          <NA>
#> 26:          NA          <NA>
#>       share_voc seq_available

# filter to only use sequence data up the the 12th of June
filter_by_availability(dt, seq_date = "2021-06-12")
#>           date location_name location  cases cases_available seq_total seq_voc
#>  1: 2021-03-20       Germany       DE  87328      2021-03-20        NA      NA
#>  2: 2021-03-27       Germany       DE 109442      2021-03-27        NA      NA
#>  3: 2021-04-03       Germany       DE 117965      2021-04-03        NA      NA
#>  4: 2021-04-10       Germany       DE 107223      2021-04-10        NA      NA
#>  5: 2021-04-17       Germany       DE 142664      2021-04-17      4108       6
#>  6: 2021-04-24       Germany       DE 145568      2021-04-24      4539      34
#>  7: 2021-05-01       Germany       DE 131887      2021-05-01      3667      54
#>  8: 2021-05-08       Germany       DE 107141      2021-05-08      4547      89
#>  9: 2021-05-15       Germany       DE  77261      2021-05-15      3688     101
#> 10: 2021-05-22       Germany       DE  57310      2021-05-22      3817     119
#> 11: 2021-05-29       Germany       DE  33052      2021-05-29      2781     102
#> 12: 2021-06-05       Germany       DE  22631      2021-06-05      2197     181
#> 13: 2021-06-12       Germany       DE  15553      2021-06-12        NA      NA
#> 14: 2021-06-19       Germany       DE   7659      2021-06-19        NA      NA
#> 15: 2021-06-26       Germany       DE   5033      2021-06-26        NA      NA
#> 16: 2021-07-03       Germany       DE   4181      2021-07-03        NA      NA
#> 17: 2021-07-10       Germany       DE   5067      2021-07-10        NA      NA
#> 18: 2021-07-17       Germany       DE   7969      2021-07-17        NA      NA
#> 19: 2021-07-24       Germany       DE  11122      2021-07-24        NA      NA
#> 20: 2021-07-31       Germany       DE  14654      2021-07-31        NA      NA
#> 21: 2021-08-07       Germany       DE  18474      2021-08-07        NA      NA
#> 22: 2021-08-14       Germany       DE  28646      2021-08-14        NA      NA
#> 23: 2021-08-21       Germany       DE  44862      2021-08-21        NA      NA
#> 24: 2021-08-28       Germany       DE  62984      2021-08-28        NA      NA
#> 25: 2021-09-04       Germany       DE  71057      2021-09-04        NA      NA
#> 26: 2021-09-11       Germany       DE  75107      2021-09-11        NA      NA
#>           date location_name location  cases cases_available seq_total seq_voc
#>       share_voc seq_available
#>  1:          NA          <NA>
#>  2:          NA          <NA>
#>  3:          NA          <NA>
#>  4:          NA          <NA>
#>  5: 0.001460565    2021-05-08
#>  6: 0.007490637    2021-05-15
#>  7: 0.014725934    2021-05-22
#>  8: 0.019573345    2021-05-29
#>  9: 0.027386117    2021-06-05
#> 10: 0.031176316    2021-06-12
#> 11: 0.036677454    2021-06-05
#> 12: 0.082385071    2021-06-12
#> 13:          NA    2021-07-03
#> 14:          NA    2021-07-10
#> 15:          NA    2021-07-17
#> 16:          NA    2021-07-24
#> 17:          NA    2021-07-31
#> 18:          NA    2021-08-07
#> 19:          NA          <NA>
#> 20:          NA          <NA>
#> 21:          NA          <NA>
#> 22:          NA          <NA>
#> 23:          NA          <NA>
#> 24:          NA          <NA>
#> 25:          NA          <NA>
#> 26:          NA          <NA>
#>       share_voc seq_available

# as above but only use
filter_by_availability(dt,
  seq_date = "2021-06-12",
  case_date = "2021-07-01"
)
#>           date location_name location  cases cases_available seq_total seq_voc
#>  1: 2021-03-20       Germany       DE  87328      2021-03-20        NA      NA
#>  2: 2021-03-27       Germany       DE 109442      2021-03-27        NA      NA
#>  3: 2021-04-03       Germany       DE 117965      2021-04-03        NA      NA
#>  4: 2021-04-10       Germany       DE 107223      2021-04-10        NA      NA
#>  5: 2021-04-17       Germany       DE 142664      2021-04-17      4108       6
#>  6: 2021-04-24       Germany       DE 145568      2021-04-24      4539      34
#>  7: 2021-05-01       Germany       DE 131887      2021-05-01      3667      54
#>  8: 2021-05-08       Germany       DE 107141      2021-05-08      4547      89
#>  9: 2021-05-15       Germany       DE  77261      2021-05-15      3688     101
#> 10: 2021-05-22       Germany       DE  57310      2021-05-22      3817     119
#> 11: 2021-05-29       Germany       DE  33052      2021-05-29      2781     102
#> 12: 2021-06-05       Germany       DE  22631      2021-06-05      2197     181
#> 13: 2021-06-12       Germany       DE  15553      2021-06-12        NA      NA
#> 14: 2021-06-19       Germany       DE   7659      2021-06-19        NA      NA
#> 15: 2021-06-26       Germany       DE   5033      2021-06-26        NA      NA
#> 16: 2021-07-03       Germany       DE     NA      2021-07-03        NA      NA
#> 17: 2021-07-10       Germany       DE     NA      2021-07-10        NA      NA
#> 18: 2021-07-17       Germany       DE     NA      2021-07-17        NA      NA
#> 19: 2021-07-24       Germany       DE     NA      2021-07-24        NA      NA
#> 20: 2021-07-31       Germany       DE     NA      2021-07-31        NA      NA
#> 21: 2021-08-07       Germany       DE     NA      2021-08-07        NA      NA
#> 22: 2021-08-14       Germany       DE     NA      2021-08-14        NA      NA
#> 23: 2021-08-21       Germany       DE     NA      2021-08-21        NA      NA
#> 24: 2021-08-28       Germany       DE     NA      2021-08-28        NA      NA
#> 25: 2021-09-04       Germany       DE     NA      2021-09-04        NA      NA
#> 26: 2021-09-11       Germany       DE     NA      2021-09-11        NA      NA
#>           date location_name location  cases cases_available seq_total seq_voc
#>       share_voc seq_available
#>  1:          NA          <NA>
#>  2:          NA          <NA>
#>  3:          NA          <NA>
#>  4:          NA          <NA>
#>  5: 0.001460565    2021-05-08
#>  6: 0.007490637    2021-05-15
#>  7: 0.014725934    2021-05-22
#>  8: 0.019573345    2021-05-29
#>  9: 0.027386117    2021-06-05
#> 10: 0.031176316    2021-06-12
#> 11: 0.036677454    2021-06-05
#> 12: 0.082385071    2021-06-12
#> 13:          NA    2021-07-03
#> 14:          NA    2021-07-10
#> 15:          NA    2021-07-17
#> 16:          NA    2021-07-24
#> 17:          NA    2021-07-31
#> 18:          NA    2021-08-07
#> 19:          NA          <NA>
#> 20:          NA          <NA>
#> 21:          NA          <NA>
#> 22:          NA          <NA>
#> 23:          NA          <NA>
#> 24:          NA          <NA>
#> 25:          NA          <NA>
#> 26:          NA          <NA>
#>       share_voc seq_available