Forecast using branching processes at a target date returning a summary data frame

Designed to streamline workflows with a large number of indepedent model fits this function returns a data frame summarising a forecast for a single model (rather than the multiple models that may be used directly with forecast_dt()). Unlike forecast() it is also error tolerant and returns the error if forecasting fails.

forecast_dt(
  obs,
  forecast_date = max(obs$date),
  strains = 2,
  overdispersion = TRUE,
  variant_relationship = "pooled",
  model = forecast.vocs::load_model(strains = strains),
  keep_forecast = FALSE,
  keep_fit = TRUE,
  id = 0,
  ...
)

Arguments

obs
A data frame with the following variables: date, cases, seq_voc, and seq_total, cases_available, and seq_available. seq_available and case_available must be uniquely define data rows but other rows can be duplicated based on data availability.

forecast_date
Date at which to forecast. Defaults to the maximum date in obs.

strains
Integer number of strains. Defaults to 2. Current maximum is 2.

overdispersion
Logical, defaults to TRUE. Should the observations used include overdispersion.

variant_relationship
Character string, defaulting to "pooled". Controls the relationship of strains with options being "pooled" (dependence determined from the data), "scaled" (a fixed scaling between strains), and "independent" (fully independent strains after initial scaling).

model
A model as loaded by load_model().

keep_forecast
Logical, defaults to FALSE. Should the forecast() output be kept.

keep_fit
Logical, defaults to TRUE. Should the stan model fit be kept and returned. Dropping this can substantially reduce memory usage in situtations where multiple models are being fit.

id
An integer, defaulting to 0 to identify this forecast.

...
Additional arguments passed to forecast()

Value

A dataframe containing the output of forecast() in each row as well as summary information about the forecast

Examples

if (FALSE) {
options(mc.cores = 4)
dt <- forecast_dt(latest_obs(germany_covid19_delta_obs),
  max_treedepth = 15, adapt_delta = 0.95
)
print(dt)
}