Extract forecasts from a summarised posterior

Uses the observed variable returned by fv_tidy_posterior() to return posterior predictions for forecast dates only.

Usage

fv_extract_forecast(posterior)

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

Value

A data.frame of forecasts in the format returned by fv_tidy_posterior() but with fitting variables dropped.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")

fv_extract_forecast(p)
#>     value_type     type       date horizon forecast_start          mean
#>  1:      cases Combined 2021-07-03       1          FALSE 3429.24600000
#>  2:      cases Combined 2021-07-10       2          FALSE 2588.25250000
#>  3:      cases Combined 2021-07-17       3          FALSE 2401.54200000
#>  4:      cases Combined 2021-07-24       4          FALSE 3449.94100000
#>  5:      cases      VOC 2021-06-19       1          FALSE 2098.75500000
#>  6:      cases      VOC 2021-06-26       2          FALSE 1951.21250000
#>  7:      cases      VOC 2021-07-03       3          FALSE 1793.53600000
#>  8:      cases      VOC 2021-07-10       4          FALSE 1712.20950000
#>  9:      cases      VOC 2021-07-17       5          FALSE 1875.48200000
#> 10:      cases      VOC 2021-07-24       6          FALSE 2967.00100000
#> 11:      cases  non-VOC 2021-06-19       1          FALSE 6203.23450000
#> 12:      cases  non-VOC 2021-06-26       2          FALSE 3187.24000000
#> 13:      cases  non-VOC 2021-07-03       3          FALSE 1635.71000000
#> 14:      cases  non-VOC 2021-07-10       4          FALSE  876.04300000
#> 15:      cases  non-VOC 2021-07-17       5          FALSE  526.06000000
#> 16:      cases  non-VOC 2021-07-24       6          FALSE  482.94100000
#> 17:        voc      VOC 2021-06-19       1          FALSE    0.25325522
#> 18:        voc      VOC 2021-06-26       2          FALSE    0.37860417
#> 19:        voc      VOC 2021-07-03       3          FALSE    0.52155912
#> 20:        voc      VOC 2021-07-10       4          FALSE    0.66026949
#> 21:        voc      VOC 2021-07-17       5          FALSE    0.77564376
#> 22:        voc      VOC 2021-07-24       6          FALSE    0.86002722
#> 23:     growth Combined 2021-06-26       1          FALSE   -0.30251201
#> 24:     growth Combined 2021-07-03       2          FALSE   -0.24626492
#> 25:     growth Combined 2021-07-10       3          FALSE   -0.19733734
#> 26:     growth Combined 2021-07-17       4          FALSE   -0.16751240
#> 27:     growth      VOC 2021-06-12       1          FALSE   -0.05034430
#> 28:     growth      VOC 2021-06-19       2          FALSE   -0.06464587
#> 29:     growth      VOC 2021-06-26       3          FALSE   -0.08196106
#> 30:     growth      VOC 2021-07-03       4          FALSE   -0.09005981
#> 31:     growth      VOC 2021-07-10       5          FALSE   -0.09445468
#> 32:     growth      VOC 2021-07-17       6          FALSE   -0.10348861
#> 33:     growth  non-VOC 2021-06-12       1          FALSE   -0.51425415
#> 34:     growth  non-VOC 2021-06-19       2          FALSE   -0.52855571
#> 35:     growth  non-VOC 2021-06-26       3          FALSE   -0.54587090
#> 36:     growth  non-VOC 2021-07-03       4          FALSE   -0.55396964
#> 37:     growth  non-VOC 2021-07-10       5          FALSE   -0.55836450
#> 38:     growth  non-VOC 2021-07-17       6          FALSE   -0.56739846
#> 39:         rt Combined 2021-06-26       1          FALSE    0.73895962
#> 40:         rt Combined 2021-07-03       2          FALSE    0.78171510
#> 41:         rt Combined 2021-07-10       3          FALSE    0.82091366
#> 42:         rt Combined 2021-07-17       4          FALSE    0.84576613
#> 43:         rt      VOC 2021-06-12       1          FALSE    0.95090197
#> 44:         rt      VOC 2021-06-19       2          FALSE    0.93739936
#> 45:         rt      VOC 2021-06-26       3          FALSE    0.92130783
#> 46:         rt      VOC 2021-07-03       4          FALSE    0.91387652
#> 47:         rt      VOC 2021-07-10       5          FALSE    0.90986897
#> 48:         rt      VOC 2021-07-17       6          FALSE    0.90168629
#> 49:         rt  non-VOC 2021-06-12       1          FALSE    0.59794641
#> 50:         rt  non-VOC 2021-06-19       2          FALSE    0.58945570
#> 51:         rt  non-VOC 2021-06-26       3          FALSE    0.57933702
#> 52:         rt  non-VOC 2021-07-03       4          FALSE    0.57466406
#> 53:         rt  non-VOC 2021-07-10       5          FALSE    0.57214404
#> 54:         rt  non-VOC 2021-07-17       6          FALSE    0.56699859
#>     value_type     type       date horizon forecast_start          mean
#>            median           sd          mad           q5           q20
#>  1: 3296.00000000 1.001209e+03 7.872606e+02 2141.0000000 2670.00000000
#>  2: 2279.50000000 1.448296e+03 9.644313e+02 1109.9000000 1609.00000000
#>  3: 1706.50000000 4.534740e+03 1.076368e+03  525.9000000  983.00000000
#>  4: 1355.50000000 3.075086e+04 1.177926e+03  260.0000000  599.60000000
#>  5: 2058.00000000 4.517657e+02 4.329192e+02 1431.0000000 1727.80000000
#>  6: 1905.50000000 4.597504e+02 4.158693e+02 1298.9500000 1574.00000000
#>  7: 1694.50000000 6.112130e+02 5.018601e+02 1012.9500000 1322.80000000
#>  8: 1492.00000000 9.742778e+02 6.716178e+02  685.9500000 1015.00000000
#>  9: 1314.00000000 3.573439e+03 8.658384e+02  396.9500000  746.60000000
#> 10: 1167.00000000 2.593159e+04 1.028183e+03  223.8500000  514.60000000
#> 11: 6141.50000000 1.008974e+03 8.977143e+02 4683.9500000 5418.00000000
#> 12: 3131.50000000 6.230724e+02 5.656119e+02 2291.6000000 2685.80000000
#> 13: 1571.50000000 5.363618e+02 4.588647e+02  953.0000000 1209.80000000
#> 14:  767.00000000 5.564496e+02 3.409980e+02  338.0000000  523.00000000
#> 15:  376.00000000 9.875049e+02 2.401812e+02  111.0000000  206.00000000
#> 16:  187.00000000 4.844803e+03 1.601208e+02   31.0000000   79.00000000
#> 17:    0.25022550 4.128863e-02 3.843863e-02    0.1940648    0.22025600
#> 18:    0.37543400 5.664506e-02 5.366864e-02    0.2944653    0.33303260
#> 19:    0.51962900 6.588415e-02 6.362949e-02    0.4196759    0.46830080
#> 20:    0.66141100 6.500086e-02 6.270212e-02    0.5539496    0.60841360
#> 21:    0.77836850 5.552442e-02 5.238545e-02    0.6804697    0.73277000
#> 22:    0.86398600 4.236880e-02 3.836524e-02    0.7866587    0.82870060
#> 23:   -0.30435350 1.749757e-01 1.594180e-01   -0.5138655   -0.40969814
#> 24:   -0.25116418 2.375012e-01 2.150889e-01   -0.5408795   -0.39256250
#> 25:   -0.20213089 3.027830e-01 2.561703e-01   -0.5815155   -0.37631597
#> 26:   -0.16941100 3.601256e-01 2.980068e-01   -0.6166901   -0.36961713
#> 27:   -0.04720850 7.286879e-02 6.092433e-02   -0.1500362   -0.09199646
#> 28:   -0.06584034 9.793869e-02 8.574543e-02   -0.1866830   -0.12390699
#> 29:   -0.08494750 1.735136e-01 1.552123e-01   -0.2927158   -0.18954949
#> 30:   -0.09163864 2.360269e-01 2.143995e-01   -0.3852124   -0.23445871
#> 31:   -0.09590507 3.018693e-01 2.588064e-01   -0.4763217   -0.27169937
#> 32:   -0.10277064 3.597775e-01 2.972860e-01   -0.5539901   -0.30884779
#> 33:   -0.51256582 7.698416e-02 6.571031e-02   -0.6213916   -0.55927960
#> 34:   -0.53001182 1.012330e-01 9.031332e-02   -0.6539449   -0.59201764
#> 35:   -0.54815043 1.755339e-01 1.575633e-01   -0.7639700   -0.65529483
#> 36:   -0.55424639 2.374524e-01 2.134114e-01   -0.8522537   -0.70461663
#> 37:   -0.56193382 3.032762e-01 2.584668e-01   -0.9418290   -0.73827050
#> 38:   -0.56841086 3.614116e-01 2.978380e-01   -1.0206790   -0.76961641
#> 39:    0.73760008 1.749757e-01 1.594180e-01    0.5981789    0.66385061
#> 40:    0.77789465 2.375012e-01 2.150889e-01    0.5822359    0.67532414
#> 41:    0.81698798 3.027830e-01 2.561703e-01    0.5590505    0.68638542
#> 42:    0.84416188 3.601256e-01 2.980068e-01    0.5397279    0.69099884
#> 43:    0.95388849 7.286879e-02 6.092433e-02    0.8606769    0.91210838
#> 44:    0.93628034 9.793869e-02 8.574543e-02    0.8297067    0.88346201
#> 45:    0.91856051 1.735136e-01 1.552123e-01    0.7462342    0.82733177
#> 46:    0.91243480 2.360269e-01 2.143995e-01    0.6803061    0.79099889
#> 47:    0.90855026 3.018693e-01 2.588064e-01    0.6210637    0.76208333
#> 48:    0.90233391 3.597775e-01 2.972860e-01    0.5746523    0.73429253
#> 49:    0.59895679 7.698416e-02 6.571031e-02    0.5371963    0.57162071
#> 50:    0.58859801 1.012330e-01 9.031332e-02    0.5199904    0.55320998
#> 51:    0.57801791 1.755339e-01 1.575633e-01    0.4658135    0.51928894
#> 52:    0.57450505 2.374524e-01 2.134114e-01    0.4264528    0.49429804
#> 53:    0.57010551 3.032762e-01 2.584668e-01    0.3899140    0.47793980
#> 54:    0.56642485 3.614116e-01 2.978380e-01    0.3603502    0.46319071
#>            median           sd          mad           q5           q20
#>               q80           q95
#>  1:  4.053400e+03 5099.25000000
#>  2:  3.313000e+03 4882.10000000
#>  3:  3.071000e+03 5946.05000000
#>  4:  2.970400e+03 7898.60000000
#>  5:  2.434000e+03 2912.05000000
#>  6:  2.285200e+03 2742.05000000
#>  7:  2.209600e+03 2860.05000000
#>  8:  2.232200e+03 3372.00000000
#>  9:  2.407000e+03 4733.85000000
#> 10:  2.565000e+03 6839.00000000
#> 11:  6.954200e+03 7939.25000000
#> 12:  3.636200e+03 4299.90000000
#> 13:  1.977200e+03 2516.10000000
#> 14:  1.131200e+03 1689.95000000
#> 15:  6.660000e+02 1354.30000000
#> 16:  4.090000e+02 1067.25000000
#> 17:  2.849324e-01    0.32067325
#> 18:  4.232296e-01    0.47110610
#> 19:  5.745284e-01    0.62755680
#> 20:  7.140604e-01    0.76101130
#> 21:  8.217192e-01    0.85734490
#> 22:  8.945640e-01    0.91952600
#> 23: -1.953094e-01   -0.07420857
#> 24: -1.066186e-01    0.06081701
#> 25: -2.388700e-02    0.18918916
#> 26:  2.207108e-02    0.29103101
#> 27: -8.905679e-03    0.03542019
#> 28: -6.473509e-03    0.06004202
#> 29:  2.229543e-02    0.14784746
#> 30:  5.016352e-02    0.21281221
#> 31:  7.704944e-02    0.29192888
#> 32:  8.521904e-02    0.34726018
#> 33: -4.699099e-01   -0.41819937
#> 34: -4.660493e-01   -0.39885210
#> 35: -4.402726e-01   -0.30736911
#> 36: -4.167730e-01   -0.24578137
#> 37: -3.804664e-01   -0.16630252
#> 38: -3.762457e-01   -0.11241972
#> 39:  8.225801e-01    0.92847803
#> 40:  8.988684e-01    1.06270443
#> 41:  9.763960e-01    1.20826948
#> 42:  1.022316e+00    1.33780606
#> 43:  9.911339e-01    1.03605496
#> 44:  9.935474e-01    1.06188117
#> 45:  1.022546e+00    1.15933604
#> 46:  1.051443e+00    1.23715230
#> 47:  1.080095e+00    1.33900779
#> 48:  1.088956e+00    1.41518488
#> 49:  6.250586e-01    0.65823098
#> 50:  6.274764e-01    0.67108995
#> 51:  6.438609e-01    0.73537911
#> 52:  6.591705e-01    0.78209320
#> 53:  6.835425e-01    0.84679003
#> 54:  6.864336e-01    0.89366908
#>               q80           q95