Plot the posterior prediction for the growth rate

Plot the posterior prediction for the growth rate

plot_growth(posterior, forecast_dates = NULL, col = NULL)

Arguments

posterior
A dataframe of posterior output as produced by summarise_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

forecast_dates
A data.frame in the format produced by extract_forecast_dates() (with at least a date variable and a Data unavailable variable)). Specifies when date availability should be add to plots. May contain facetting variables.

col
A character string denoting the variable to use to stratify the ribbon plot. Defaults to "type" which indicates the data stream.

Value

A ggplot2 plot.

See also

Examples

if (FALSE) { # interactive()
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_delta = 0.99, max_treedepth = 15)
posterior <- summarise_posterior(fit)
plot_growth(posterior)
}