Pairs plot of parameters of interest and fitting diagnostics

Pairs plot of parameters of interest and fitting diagnostics

plot_pairs(
  fit,
  pars = c("r_init", "r_noise", "beta", "voc_noise[1]", "nvoc_noise[1]", "init_cases",
    "init_cases[1]", "init_cases[2]", "eta[1]", "voc_eta[1]", "nvoc_eta[1]",
    "sqrt_phi[1]", "sqrt_phi[2]", "sqrt_phi"),
  diagnostics = TRUE,
  ...
)

Arguments

fit
List of output as returned by stan_fit().

pars
Character vector of parameters to try and include in the plot. Will only be included if present in the fitted model.

diagnostics
Logical, defaults to TRUE. Should fitting diagnostics be returned as a data.frame.

...
Additional parameters passed to bayesplot::mcmc_pairs().

Value

A ggplot2 based pairs plot of parameters of interest

See also

Examples

if (FALSE) { # interactive()
obs <- filter_by_availability(
  germany_covid19_delta_obs,
  date = as.Date("2021-06-12"),
)
dt <- stan_data(obs)
inits <- stan_inits(dt)
fit <- stan_fit(dt, init = inits, adapt_voc = 0.99, max_treedepth = 15)
plot_pairs(fit)
}