Format data for use with stan

Format data for use with stan

stan_data(
  obs,
  horizon = 4,
  r_init = c(0, 0.25),
  voc_scale = c(0, 0.2),
  variant_relationship = "pooled",
  overdispersion = TRUE,
  likelihood = TRUE,
  output_loglik = TRUE,
  debug = FALSE
)

Arguments

obs
A data frame with the following variables: date, cases, seq_voc, and seq_total.

horizon
Integer forecast horizon. Defaults to 4.

r_init
Numeric vector of length 2. Prior mean and standard deviation for the initial growht rate.

voc_scale
Numeric vector of length 2. Prior mean and standard deviation for the initial growth rate modifier due to the variant of concern.

variant_relationship
Character string, defaulting to "pooled". Controls the relationship of strains with options being "pooled" (dependence determined from the data), "scaled" (a fixed scaling between strains), and "independent" (fully independent strains after initial scaling).

overdispersion
Logical, defaults to TRUE. Should the observations used include overdispersion.

likelihood
Logical, defaults to TRUE. Should the likelihood be included in the model

output_loglik
Logical, defaults to FALSE. Should the log-likelihood be output. Disabling this will speed up fitting if evaluating the model fit is not required.

debug
Logical, defaults to FALSE. Should within model debug information be returned.

Value

A list as required by stan.

See also

Examples

stan_data(latest_obs(germany_covid19_delta_obs))
#> $t
#> [1] 30
#> 
#> $t_nots
#> [1] 26
#> 
#> $t_nseq
#> [1] 4
#> 
#> $t_seq
#> [1] 14
#> 
#> $t_seqf
#> [1] 26
#> 
#> $X
#>  [1]  87328 109442 117965 107223 142664 145568 131887 107141  77261  57310
#> [11]  33052  22631  15553   7659   5033   4181   5067   7969  11122  14654
#> [21]  18474  28646  44862  62984  71057  75107
#> 
#> $N
#>  [1] 4108 4539 3667 4547 3688 3817 2781 2197 1566  991  706  803  735  564
#> 
#> $Y
#>  [1]   6  34  54  89 101 119 102 181 309 414 451 631 643 521
#> 
#> $likelihood
#> [1] 1
#> 
#> $output_loglik
#> [1] 1
#> 
#> $start_date
#> [1] "2021-03-20"
#> 
#> $seq_start_date
#> [1] "2021-04-17"
#> 
#> $r_init_mean
#> [1] 0
#> 
#> $r_init_sd
#> [1] 0.25
#> 
#> $voc_mean
#> [1] 0
#> 
#> $voc_sd
#> [1] 0.2
#> 
#> $relat
#> [1] 1
#> 
#> $overdisp
#> [1] 1
#> 
#> $debug
#> [1] 0
#> 
#> $t_dep
#> [1] 28
#>