Summary method for fv_tidy_posterior

summary method for class "fv_tidy_posterior". Can be used to filter the posterior for variables of interest, to return forecasts only, and to summarise using the data.table method

Usage

# S3 method for fv_posterior
summary(object, type = "model", forecast = FALSE, as_dt = FALSE, ...)

Arguments

object

An object of the class fv_posterior as returned by fv_tidy_posterior() .

type

A character string used to filter the summarised output and defaulting to "model". Current options are: "model" which returns a summary of key model parameters, "cases" which returns summarised cases, "voc_frac" which returns summarised estimates of the fraction of cases that have the variant of concern, "voc_advantage" that returns summarised estimates of the the transmission advantage of the variant of concern, "growth" which returns summarised variant specific and overall growth rates, "rt" which returns summarised variant specific and overall reproduction number estimates, "raw" which returns a raw posterior summary, and "all" which returns all tidied posterior estimates.

forecast

Logical defaults to FALSE. Should fv_extract_forecast() be used to return only forecasts rather than complete posterior.

as_dt

Logical defaults to FALSE. Once any filtering has been applied should summary() fall back to using the default data.table method.

...

Additional summary arguments.

Value

A summary data.table table unless type "all" is used in which case the output is still of type "fv_posterior".

See also

Examples

posterior <- fv_example(strains = 2, type = "posterior")

# case summary
summary(posterior, type = "cases")
#>               variable       date     type    obs observed forecast_start
#>  1:       sim_cases[1] 2021-03-20 Combined  87328     TRUE          FALSE
#>  2:       sim_cases[2] 2021-03-27 Combined 109442     TRUE          FALSE
#>  3:       sim_cases[3] 2021-04-03 Combined 117965     TRUE          FALSE
#>  4:       sim_cases[4] 2021-04-10 Combined 107223     TRUE          FALSE
#>  5:       sim_cases[5] 2021-04-17 Combined 142664     TRUE          FALSE
#>  6:       sim_cases[6] 2021-04-24 Combined 145568     TRUE          FALSE
#>  7:       sim_cases[7] 2021-05-01 Combined 131887     TRUE          FALSE
#>  8:       sim_cases[8] 2021-05-08 Combined 107141     TRUE          FALSE
#>  9:       sim_cases[9] 2021-05-15 Combined  77261     TRUE          FALSE
#> 10:      sim_cases[10] 2021-05-22 Combined  57310     TRUE          FALSE
#> 11:      sim_cases[11] 2021-05-29 Combined  33052     TRUE          FALSE
#> 12:      sim_cases[12] 2021-06-05 Combined  22631     TRUE          FALSE
#> 13:      sim_cases[13] 2021-06-12 Combined  15553     TRUE          FALSE
#> 14:      sim_cases[14] 2021-06-19 Combined   7659     TRUE          FALSE
#> 15:      sim_cases[15] 2021-06-26 Combined   5033     TRUE           TRUE
#> 16:      sim_cases[16] 2021-07-03 Combined     NA    FALSE          FALSE
#> 17:      sim_cases[17] 2021-07-10 Combined     NA    FALSE          FALSE
#> 18:      sim_cases[18] 2021-07-17 Combined     NA    FALSE          FALSE
#> 19:      sim_cases[19] 2021-07-24 Combined     NA    FALSE          FALSE
#> 20:   sim_voc_cases[1] 2021-04-17      VOC     NA     TRUE          FALSE
#> 21:   sim_voc_cases[2] 2021-04-24      VOC     NA     TRUE          FALSE
#> 22:   sim_voc_cases[3] 2021-05-01      VOC     NA     TRUE          FALSE
#> 23:   sim_voc_cases[4] 2021-05-08      VOC     NA     TRUE          FALSE
#> 24:   sim_voc_cases[5] 2021-05-15      VOC     NA     TRUE          FALSE
#> 25:   sim_voc_cases[6] 2021-05-22      VOC     NA     TRUE          FALSE
#> 26:   sim_voc_cases[7] 2021-05-29      VOC     NA     TRUE          FALSE
#> 27:   sim_voc_cases[8] 2021-06-05      VOC     NA     TRUE          FALSE
#> 28:   sim_voc_cases[9] 2021-06-12      VOC     NA     TRUE           TRUE
#> 29:  sim_voc_cases[10] 2021-06-19      VOC     NA    FALSE          FALSE
#> 30:  sim_voc_cases[11] 2021-06-26      VOC     NA    FALSE          FALSE
#> 31:  sim_voc_cases[12] 2021-07-03      VOC     NA    FALSE          FALSE
#> 32:  sim_voc_cases[13] 2021-07-10      VOC     NA    FALSE          FALSE
#> 33:  sim_voc_cases[14] 2021-07-17      VOC     NA    FALSE          FALSE
#> 34:  sim_voc_cases[15] 2021-07-24      VOC     NA    FALSE          FALSE
#> 35:  sim_nvoc_cases[1] 2021-03-20  non-VOC     NA     TRUE          FALSE
#> 36:  sim_nvoc_cases[2] 2021-03-27  non-VOC     NA     TRUE          FALSE
#> 37:  sim_nvoc_cases[3] 2021-04-03  non-VOC     NA     TRUE          FALSE
#> 38:  sim_nvoc_cases[4] 2021-04-10  non-VOC     NA     TRUE          FALSE
#> 39:  sim_nvoc_cases[5] 2021-04-17  non-VOC     NA     TRUE          FALSE
#> 40:  sim_nvoc_cases[6] 2021-04-24  non-VOC     NA     TRUE          FALSE
#> 41:  sim_nvoc_cases[7] 2021-05-01  non-VOC     NA     TRUE          FALSE
#> 42:  sim_nvoc_cases[8] 2021-05-08  non-VOC     NA     TRUE          FALSE
#> 43:  sim_nvoc_cases[9] 2021-05-15  non-VOC     NA     TRUE          FALSE
#> 44: sim_nvoc_cases[10] 2021-05-22  non-VOC     NA     TRUE          FALSE
#> 45: sim_nvoc_cases[11] 2021-05-29  non-VOC     NA     TRUE          FALSE
#> 46: sim_nvoc_cases[12] 2021-06-05  non-VOC     NA     TRUE          FALSE
#> 47: sim_nvoc_cases[13] 2021-06-12  non-VOC     NA     TRUE           TRUE
#> 48: sim_nvoc_cases[14] 2021-06-19  non-VOC     NA    FALSE          FALSE
#> 49: sim_nvoc_cases[15] 2021-06-26  non-VOC     NA    FALSE          FALSE
#> 50: sim_nvoc_cases[16] 2021-07-03  non-VOC     NA    FALSE          FALSE
#> 51: sim_nvoc_cases[17] 2021-07-10  non-VOC     NA    FALSE          FALSE
#> 52: sim_nvoc_cases[18] 2021-07-17  non-VOC     NA    FALSE          FALSE
#> 53: sim_nvoc_cases[19] 2021-07-24  non-VOC     NA    FALSE          FALSE
#>               variable       date     type    obs observed forecast_start
#>       mean median      sd     mad     q5    q20    q80    q95  rhat ess_bulk
#>  1:  90000  89200 13600.0 11800.0  70300  79500  99900 113000 1.000     1700
#>  2: 103000 103000 15300.0 13200.0  78900  91000 114000 127000 1.000     1820
#>  3: 113000 113000 16300.0 14500.0  88600 101000 126000 141000 1.000     2090
#>  4: 121000 120000 18000.0 15300.0  93300 108000 133000 152000 1.000     1320
#>  5: 133000 133000 20600.0 17500.0 102000 118000 147000 167000 1.000     2040
#>  6: 140000 139000 20900.0 18500.0 107000 123000 155000 174000 1.000     1980
#>  7: 130000 129000 19500.0 16200.0  99600 115000 143000 162000 1.000     2050
#>  8: 107000 106000 16900.0 14000.0  81900  94400 119000 135000 1.000     2090
#>  9:  79600  78700 12100.0 10500.0  61500  70100  88000  99300 1.000     1920
#> 10:  55300  55000  8030.0  6940.0  42900  49200  61000  69000 1.000     2120
#> 11:  35700  35100  5400.0  4360.0  28300  31800  39200  44900 1.000     1980
#> 12:  22900  22700  3240.0  2840.0  17900  20400  25300  28500 1.000     1810
#> 13:  14100  14100  1900.0  1660.0  11100  12700  15500  17200 1.000     1710
#> 14:   8260   8180  1120.0   937.0   6600   7410   9040  10100 1.000     2240
#> 15:   5110   5010   826.0   678.0   3980   4500   5660   6490 1.000     2070
#> 16:   3470   3340   962.0   830.0   2200   2710   4130   5160 1.000     2070
#> 17:   2660   2390  1440.0  1000.0   1150   1650   3420   5000 1.000     1950
#> 18:   2370   1840  2110.0  1180.0    591   1040   3240   5780 1.000     1990
#> 19:   2540   1490  4390.0  1260.0    285    676   3140   7810 1.000     2080
#> 20:    220    219    37.9    35.6    161    190    251    284 1.000     2160
#> 21:    413    410    72.5    69.7    302    352    469    532 1.000     1900
#> 22:    690    680   121.0   113.0    506    593    782    891 1.000     2050
#> 23:   1020   1010   186.0   164.0    764    878   1150   1350 0.999     2020
#> 24:   1370   1360   254.0   234.0    990   1160   1560   1800 0.999     2430
#> 25:   1700   1670   341.0   309.0   1210   1410   1940   2250 1.000     1790
#> 26:   1930   1900   404.0   354.0   1340   1610   2220   2620 1.000     1520
#> 27:   2130   2090   459.0   411.0   1470   1760   2450   2930 1.000     1840
#> 28:   2210   2170   476.0   441.0   1480   1820   2570   3060 1.000     1600
#> 29:   2070   2030   442.0   412.0   1420   1710   2420   2860 1.000     2270
#> 30:   1920   1880   460.0   388.0   1270   1560   2230   2670 1.000     2030
#> 31:   1800   1710   585.0   491.0   1040   1320   2190   2850 1.000     2190
#> 32:   1750   1570   977.0   681.0    723   1080   2260   3290 1.000     1970
#> 33:   1840   1420  1650.0   927.0    447    786   2490   4440 1.000     2100
#> 34:   2180   1280  3730.0  1070.0    251    571   2700   6610 1.000     2070
#> 35:  90000  89200 13600.0 11800.0  70300  79500  99900 113000 1.000     1700
#> 36: 103000 103000 15300.0 13200.0  78900  91000 114000 127000 1.000     1820
#> 37: 113000 113000 16300.0 14500.0  88600 101000 126000 141000 1.000     2090
#> 38: 121000 120000 18000.0 15300.0  93300 108000 133000 152000 1.000     1320
#> 39: 133000 132000 20600.0 17500.0 102000 118000 147000 167000 1.000     2040
#> 40: 139000 139000 20900.0 18400.0 107000 123000 154000 174000 1.000     1970
#> 41: 129000 128000 19500.0 16200.0  99000 115000 142000 161000 1.000     2050
#> 42: 106000 105000 16800.0 14100.0  81000  93400 118000 134000 1.000     2080
#> 43:  78200  77400 12000.0 10400.0  60300  68700  86600  97900 1.000     1920
#> 44:  53600  53400  7960.0  6830.0  41300  47600  59200  67300 1.000     2120
#> 45:  33800  33300  5310.0  4330.0  26500  29900  37300  43100 1.000     1990
#> 46:  20700  20600  3160.0  2810.0  16000  18200  23100  26300 1.000     1870
#> 47:  11900  11900  1830.0  1620.0   9070  10600  13300  14900 1.000     1760
#> 48:   6190   6120  1030.0   892.0   4690   5380   6940   7910 1.000     2310
#> 49:   3190   3120   648.0   566.0   2280   2680   3640   4340 1.000     2160
#> 50:   1670   1590   525.0   449.0    964   1260   2050   2620 1.000     1800
#> 51:    910    808   538.0   368.0    354    542   1180   1780 1.000     1790
#> 52:    534    402   515.0   265.0    118    226    725   1340 1.000     1810
#> 53:    361    204   710.0   176.0     37     91    444   1090 1.000     1910
#>       mean median      sd     mad     q5    q20    q80    q95  rhat ess_bulk
#>     ess_tail
#>  1:     1610
#>  2:     1620
#>  3:     1900
#>  4:     1550
#>  5:     1710
#>  6:     1730
#>  7:     1750
#>  8:     1890
#>  9:     1950
#> 10:     1800
#> 11:     1700
#> 12:     1530
#> 13:     1730
#> 14:     1780
#> 15:     1820
#> 16:     1740
#> 17:     1590
#> 18:     1530
#> 19:     1340
#> 20:     1750
#> 21:     1850
#> 22:     1810
#> 23:     1620
#> 24:     1940
#> 25:     1930
#> 26:     1480
#> 27:     1670
#> 28:     1480
#> 29:     1750
#> 30:     1920
#> 31:     1710
#> 32:     1700
#> 33:     1580
#> 34:     1540
#> 35:     1610
#> 36:     1620
#> 37:     1900
#> 38:     1550
#> 39:     1710
#> 40:     1730
#> 41:     1750
#> 42:     1890
#> 43:     1990
#> 44:     1810
#> 45:     1690
#> 46:     1530
#> 47:     1780
#> 48:     1930
#> 49:     1380
#> 50:     1810
#> 51:     1390
#> 52:     1430
#> 53:     1400
#>     ess_tail

# summary of the case summary
summary(posterior, type = "cases", as_dt = TRUE)
#>    variable              date                type                obs        
#>  Length:53          Min.   :2021-03-20   Length:53          Min.   :  5033  
#>  Class :character   1st Qu.:2021-04-24   Class :character   1st Qu.: 27842  
#>  Mode  :character   Median :2021-05-29   Mode  :character   Median : 87328  
#>                     Mean   :2021-05-25                      Mean   : 77848  
#>                     3rd Qu.:2021-06-26                      3rd Qu.:113704  
#>                     Max.   :2021-07-24                      Max.   :145568  
#>                                                             NA's   :38      
#>   observed       forecast_start       mean            median      
#>  Mode :logical   Mode :logical   Min.   :   220   Min.   :   204  
#>  FALSE:16        FALSE:50        1st Qu.:  1920   1st Qu.:  1670  
#>  TRUE :37        TRUE :3         Median :  8260   Median :  8180  
#>                                  Mean   : 44100   Mean   : 43764  
#>                                  3rd Qu.:103000   3rd Qu.:103000  
#>                                  Max.   :140000   Max.   :139000  
#>                                                                   
#>        sd               mad                q5              q20        
#>  Min.   :   37.9   Min.   :   35.6   Min.   :    37   Min.   :    91  
#>  1st Qu.:  538.0   1st Qu.:  441.0   1st Qu.:   990   1st Qu.:  1260  
#>  Median : 2110.0   Median : 1260.0   Median :  6600   Median :  7410  
#>  Mean   : 6944.1   Mean   : 5830.0   Mean   : 33834   Mean   : 38924  
#>  3rd Qu.:15300.0   3rd Qu.:13200.0   3rd Qu.: 78900   3rd Qu.: 91000  
#>  Max.   :20900.0   Max.   :18500.0   Max.   :107000   Max.   :123000  
#>                                                                       
#>       q80              q95              rhat          ess_bulk   
#>  Min.   :   251   Min.   :   284   Min.   :0.999   Min.   :1320  
#>  1st Qu.:  2260   1st Qu.:  2860   1st Qu.:1.000   1st Qu.:1820  
#>  Median :  9040   Median : 10100   Median :1.000   Median :1990  
#>  Mean   : 48879   Mean   : 55590   Mean   :1.000   Mean   :1952  
#>  3rd Qu.:114000   3rd Qu.:127000   3rd Qu.:1.000   3rd Qu.:2080  
#>  Max.   :155000   Max.   :174000   Max.   :1.000   Max.   :2430  
#>                                                                  
#>     ess_tail   
#>  Min.   :1340  
#>  1st Qu.:1580  
#>  Median :1710  
#>  Mean   :1696  
#>  3rd Qu.:1810  
#>  Max.   :1990  
#>                

# case forecast only
summary(posterior, type = "cases", forecast = TRUE)
#>         type       date horizon forecast_start mean median   sd  mad   q5  q20
#>  1: Combined 2021-07-03       1          FALSE 3470   3340  962  830 2200 2710
#>  2: Combined 2021-07-10       2          FALSE 2660   2390 1440 1000 1150 1650
#>  3: Combined 2021-07-17       3          FALSE 2370   1840 2110 1180  591 1040
#>  4: Combined 2021-07-24       4          FALSE 2540   1490 4390 1260  285  676
#>  5:      VOC 2021-06-19       1          FALSE 2070   2030  442  412 1420 1710
#>  6:      VOC 2021-06-26       2          FALSE 1920   1880  460  388 1270 1560
#>  7:      VOC 2021-07-03       3          FALSE 1800   1710  585  491 1040 1320
#>  8:      VOC 2021-07-10       4          FALSE 1750   1570  977  681  723 1080
#>  9:      VOC 2021-07-17       5          FALSE 1840   1420 1650  927  447  786
#> 10:      VOC 2021-07-24       6          FALSE 2180   1280 3730 1070  251  571
#> 11:  non-VOC 2021-06-19       1          FALSE 6190   6120 1030  892 4690 5380
#> 12:  non-VOC 2021-06-26       2          FALSE 3190   3120  648  566 2280 2680
#> 13:  non-VOC 2021-07-03       3          FALSE 1670   1590  525  449  964 1260
#> 14:  non-VOC 2021-07-10       4          FALSE  910    808  538  368  354  542
#> 15:  non-VOC 2021-07-17       5          FALSE  534    402  515  265  118  226
#> 16:  non-VOC 2021-07-24       6          FALSE  361    204  710  176   37   91
#>      q80  q95
#>  1: 4130 5160
#>  2: 3420 5000
#>  3: 3240 5780
#>  4: 3140 7810
#>  5: 2420 2860
#>  6: 2230 2670
#>  7: 2190 2850
#>  8: 2260 3290
#>  9: 2490 4440
#> 10: 2700 6610
#> 11: 6940 7910
#> 12: 3640 4340
#> 13: 2050 2620
#> 14: 1180 1780
#> 15:  725 1340
#> 16:  444 1090

# voc fraction summary
summary(posterior, type = "voc_frac")
#>         variable       date type         obs observed forecast_start    mean
#>  1:  frac_voc[1] 2021-04-17  VOC 0.001465917     TRUE          FALSE 0.00166
#>  2:  frac_voc[2] 2021-04-24  VOC 0.007085917     TRUE          FALSE 0.00299
#>  3:  frac_voc[3] 2021-05-01  VOC 0.015085025     TRUE          FALSE 0.00537
#>  4:  frac_voc[4] 2021-05-08  VOC 0.019443217     TRUE          FALSE 0.00963
#>  5:  frac_voc[5] 2021-05-15  VOC 0.026459592     TRUE          FALSE 0.01720
#>  6:  frac_voc[6] 2021-05-22  VOC 0.032900194     TRUE          FALSE 0.03060
#>  7:  frac_voc[7] 2021-05-29  VOC 0.038084020     TRUE          FALSE 0.05400
#>  8:  frac_voc[8] 2021-06-05  VOC 0.085600907     TRUE          FALSE 0.09330
#>  9:  frac_voc[9] 2021-06-12  VOC 0.166666667     TRUE           TRUE 0.15700
#> 10: frac_voc[10] 2021-06-19  VOC          NA    FALSE          FALSE 0.25100
#> 11: frac_voc[11] 2021-06-26  VOC          NA    FALSE          FALSE 0.37500
#> 12: frac_voc[12] 2021-07-03  VOC          NA    FALSE          FALSE 0.51700
#> 13: frac_voc[13] 2021-07-10  VOC          NA    FALSE          FALSE 0.65600
#> 14: frac_voc[14] 2021-07-17  VOC          NA    FALSE          FALSE 0.77200
#> 15: frac_voc[15] 2021-07-24  VOC          NA    FALSE          FALSE 0.85800
#>      median       sd      mad      q5     q20     q80     q95 rhat ess_bulk
#>  1: 0.00165 0.000208 0.000200 0.00136 0.00149 0.00183 0.00203    1     2090
#>  2: 0.00296 0.000328 0.000323 0.00248 0.00272 0.00326 0.00357    1     2210
#>  3: 0.00533 0.000540 0.000524 0.00449 0.00491 0.00582 0.00630    1     2390
#>  4: 0.00960 0.000953 0.000953 0.00809 0.00884 0.01040 0.01120    1     2510
#>  5: 0.01720 0.001810 0.001750 0.01430 0.01580 0.01870 0.02020    1     2320
#>  6: 0.03050 0.003580 0.003400 0.02510 0.02780 0.03350 0.03660    1     2160
#>  7: 0.05370 0.007090 0.006590 0.04330 0.04830 0.05940 0.06610    1     2010
#>  8: 0.09250 0.013600 0.012200 0.07270 0.08270 0.10400 0.11700    1     1830
#>  9: 0.15500 0.024200 0.021700 0.12000 0.13800 0.17500 0.19800    1     1720
#> 10: 0.24800 0.038800 0.035100 0.19100 0.22000 0.28000 0.31800    1     1640
#> 11: 0.37300 0.054000 0.048500 0.28900 0.33200 0.41700 0.46800    1     1600
#> 12: 0.51700 0.063700 0.057400 0.41300 0.46900 0.56700 0.62400    1     1600
#> 13: 0.65800 0.063500 0.057300 0.54700 0.60900 0.70800 0.75800    1     1610
#> 14: 0.77500 0.054700 0.048900 0.67600 0.73300 0.81700 0.85500    1     1610
#> 15: 0.86100 0.041900 0.036700 0.78300 0.82900 0.89200 0.91700    1     1620
#>     ess_tail
#>  1:     1790
#>  2:     1640
#>  3:     1790
#>  4:     1780
#>  5:     1640
#>  6:     1530
#>  7:     1420
#>  8:     1340
#>  9:     1370
#> 10:     1440
#> 11:     1380
#> 12:     1440
#> 13:     1440
#> 14:     1420
#> 15:     1490

# voc advantage summary
summary(posterior, type = "voc_advantage")
#>              variable       date type observed forecast_start     mean   median
#>  1:  voc_advantage[1] 2021-04-17  VOC     TRUE          FALSE 1.588493 1.587245
#>  2:  voc_advantage[2] 2021-04-24  VOC     TRUE          FALSE 1.588493 1.587245
#>  3:  voc_advantage[3] 2021-05-01  VOC     TRUE          FALSE 1.588493 1.587245
#>  4:  voc_advantage[4] 2021-05-08  VOC     TRUE          FALSE 1.588493 1.587245
#>  5:  voc_advantage[5] 2021-05-15  VOC     TRUE          FALSE 1.588493 1.587245
#>  6:  voc_advantage[6] 2021-05-22  VOC     TRUE          FALSE 1.588493 1.587245
#>  7:  voc_advantage[7] 2021-05-29  VOC     TRUE          FALSE 1.588493 1.587245
#>  8:  voc_advantage[8] 2021-06-05  VOC     TRUE           TRUE 1.588493 1.587245
#>  9:  voc_advantage[9] 2021-06-12  VOC    FALSE          FALSE 1.588493 1.587245
#> 10: voc_advantage[10] 2021-06-19  VOC    FALSE          FALSE 1.588493 1.587245
#> 11: voc_advantage[11] 2021-06-26  VOC    FALSE          FALSE 1.588493 1.587245
#> 12: voc_advantage[12] 2021-07-03  VOC    FALSE          FALSE 1.588493 1.587245
#> 13: voc_advantage[13] 2021-07-10  VOC    FALSE          FALSE 1.588493 1.587245
#> 14: voc_advantage[14] 2021-07-17  VOC    FALSE          FALSE 1.588493 1.587245
#>         sd    mad      q5      q20     q80     q95 rhat ess_bulk ess_tail
#>  1: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  2: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  3: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  4: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  5: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  6: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  7: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  8: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#>  9: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#> 10: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#> 11: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#> 12: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#> 13: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510
#> 14: 0.0287 0.0273 1.53331 1.560044 1.61619 1.64825    1     1670     1510

# growth summary
summary(posterior, type = "growth")
#>      variable       date     type observed forecast_start        mean
#>  1:  com_r[1] 2021-03-20 Combined     TRUE          FALSE  0.10607143
#>  2:  com_r[2] 2021-03-27 Combined     TRUE          FALSE  0.07935714
#>  3:  com_r[3] 2021-04-03 Combined     TRUE          FALSE  0.04950000
#>  4:  com_r[4] 2021-04-10 Combined     TRUE          FALSE  0.07472143
#>  5:  com_r[5] 2021-04-17 Combined     TRUE          FALSE  0.03150714
#>  6:  com_r[6] 2021-04-24 Combined     TRUE          FALSE -0.05775000
#>  7:  com_r[7] 2021-05-01 Combined     TRUE          FALSE -0.14771429
#>  8:  com_r[8] 2021-05-08 Combined     TRUE          FALSE -0.22707143
#>  9:  com_r[9] 2021-05-15 Combined     TRUE          FALSE -0.28207143
#> 10: com_r[10] 2021-05-22 Combined     TRUE          FALSE -0.33785714
#> 11: com_r[11] 2021-05-29 Combined     TRUE          FALSE -0.34178571
#> 12: com_r[12] 2021-06-05 Combined     TRUE          FALSE -0.36142857
#> 13: com_r[13] 2021-06-12 Combined     TRUE          FALSE -0.39992857
#> 14: com_r[14] 2021-06-19 Combined     TRUE           TRUE -0.34650000
#> 15: com_r[15] 2021-06-26 Combined    FALSE          FALSE -0.29307143
#> 16: com_r[16] 2021-07-03 Combined    FALSE          FALSE -0.23100000
#> 17: com_r[17] 2021-07-10 Combined    FALSE          FALSE -0.18700000
#> 18: com_r[18] 2021-07-17 Combined    FALSE          FALSE -0.15164286
#> 19:  voc_r[1] 2021-04-17      VOC     TRUE          FALSE  0.49264286
#> 20:  voc_r[2] 2021-04-24      VOC     TRUE          FALSE  0.40228571
#> 21:  voc_r[3] 2021-05-01      VOC     TRUE          FALSE  0.31035714
#> 22:  voc_r[4] 2021-05-08      VOC     TRUE          FALSE  0.22707143
#> 23:  voc_r[5] 2021-05-15      VOC     TRUE          FALSE  0.16578571
#> 24:  voc_r[6] 2021-05-22      VOC     TRUE          FALSE  0.09900000
#> 25:  voc_r[7] 2021-05-29      VOC     TRUE          FALSE  0.07770714
#> 26:  voc_r[8] 2021-06-05      VOC     TRUE           TRUE  0.02860000
#> 27:  voc_r[9] 2021-06-12      VOC    FALSE          FALSE -0.05397857
#> 28: voc_r[10] 2021-06-19      VOC    FALSE          FALSE -0.05830000
#> 29: voc_r[11] 2021-06-26      VOC    FALSE          FALSE -0.07142143
#> 30: voc_r[12] 2021-07-03      VOC    FALSE          FALSE -0.07377857
#> 31: voc_r[13] 2021-07-10      VOC    FALSE          FALSE -0.08250000
#> 32: voc_r[14] 2021-07-17      VOC    FALSE          FALSE -0.08721429
#> 33:      r[1] 2021-03-20  non-VOC     TRUE          FALSE  0.10607143
#> 34:      r[2] 2021-03-27  non-VOC     TRUE          FALSE  0.07935714
#> 35:      r[3] 2021-04-03  non-VOC     TRUE          FALSE  0.04950000
#> 36:      r[4] 2021-04-10  non-VOC     TRUE          FALSE  0.07472143
#> 37:      r[5] 2021-04-17  non-VOC     TRUE          FALSE  0.03009286
#> 38:      r[6] 2021-04-24  non-VOC     TRUE          FALSE -0.06018571
#> 39:      r[7] 2021-05-01  non-VOC     TRUE          FALSE -0.15242857
#> 40:      r[8] 2021-05-08  non-VOC     TRUE          FALSE -0.23571429
#> 41:      r[9] 2021-05-15  non-VOC     TRUE          FALSE -0.29700000
#> 42:     r[10] 2021-05-22  non-VOC     TRUE          FALSE -0.36300000
#> 43:     r[11] 2021-05-29  non-VOC     TRUE          FALSE -0.38500000
#> 44:     r[12] 2021-06-05  non-VOC     TRUE           TRUE -0.43371429
#> 45:     r[13] 2021-06-12  non-VOC    FALSE          FALSE -0.51621429
#> 46:     r[14] 2021-06-19  non-VOC    FALSE          FALSE -0.52092857
#> 47:     r[15] 2021-06-26  non-VOC    FALSE          FALSE -0.53428571
#> 48:     r[16] 2021-07-03  non-VOC    FALSE          FALSE -0.53664286
#> 49:     r[17] 2021-07-10  non-VOC    FALSE          FALSE -0.54528571
#> 50:     r[18] 2021-07-17  non-VOC    FALSE          FALSE -0.54921429
#>      variable       date     type observed forecast_start        mean
#>          median     sd    mad           q5          q20          q80
#>  1:  0.10764286 0.0759 0.0750  0.007857143  0.057200000  0.154785714
#>  2:  0.08014286 0.0656 0.0561 -0.005971429  0.041328571  0.116285714
#>  3:  0.05500000 0.0710 0.0618 -0.048321429  0.007566429  0.091928571
#>  4:  0.07157857 0.0711 0.0632 -0.008485714  0.031978571  0.117857143
#>  5:  0.03001429 0.0706 0.0635 -0.052957143 -0.011471429  0.074171429
#>  6: -0.05892857 0.0676 0.0586 -0.143785714 -0.098214286 -0.019250000
#>  7: -0.14771429 0.0699 0.0612 -0.235714286 -0.189357143 -0.108428571
#>  8: -0.22550000 0.0696 0.0589 -0.319785714 -0.266357143 -0.188571429
#>  9: -0.28285714 0.0709 0.0596 -0.372428571 -0.322928571 -0.242785714
#> 10: -0.33471429 0.0679 0.0591 -0.428214286 -0.377142857 -0.298571429
#> 11: -0.34335714 0.0691 0.0621 -0.425071429 -0.381857143 -0.299357143
#> 12: -0.36300000 0.0745 0.0661 -0.457285714 -0.404642857 -0.317428571
#> 13: -0.39835714 0.0803 0.0697 -0.509142857 -0.445500000 -0.352000000
#> 14: -0.34807143 0.0975 0.0886 -0.465142857 -0.407785714 -0.287571429
#> 15: -0.29464286 0.1760 0.1590 -0.511500000 -0.399142857 -0.185428571
#> 16: -0.23021429 0.2340 0.2120 -0.527214286 -0.375571429 -0.094285714
#> 17: -0.18464286 0.2950 0.2620 -0.553142857 -0.364571429 -0.012885714
#> 18: -0.14614286 0.3450 0.3020 -0.576714286 -0.358285714  0.059242857
#> 19:  0.48950000 0.0723 0.0664  0.403857143  0.449428571  0.536642857
#> 20:  0.40307143 0.0699 0.0630  0.315071429  0.360642857  0.443928571
#> 21:  0.31114286 0.0728 0.0665  0.214500000  0.266357143  0.353571429
#> 22:  0.22942857 0.0740 0.0635  0.129642857  0.183857143  0.271071429
#> 23:  0.16657143 0.0741 0.0624  0.068357143  0.122571429  0.209000000
#> 24:  0.10214286 0.0729 0.0659  0.002993571  0.054764286  0.142214286
#> 25:  0.07739286 0.0728 0.0665 -0.011785714  0.034021429  0.123357143
#> 26:  0.02805000 0.0768 0.0672 -0.069928571 -0.016342857  0.074800000
#> 27: -0.05280000 0.0811 0.0715 -0.162642857 -0.101357143 -0.005570714
#> 28: -0.06097143 0.0981 0.0905 -0.176000000 -0.117857143  0.000269500
#> 29: -0.07322857 0.1770 0.1590 -0.291500000 -0.179142857  0.034650000
#> 30: -0.07315000 0.2350 0.2110 -0.367714286 -0.214500000  0.066550000
#> 31: -0.08171429 0.2950 0.2630 -0.446285714 -0.258500000  0.091142857
#> 32: -0.08328571 0.3450 0.3030 -0.520928571 -0.293071429  0.121785714
#> 33:  0.10764286 0.0759 0.0750  0.007857143  0.057200000  0.154785714
#> 34:  0.08014286 0.0656 0.0561 -0.005971429  0.041328571  0.116285714
#> 35:  0.05500000 0.0710 0.0618 -0.048321429  0.007566429  0.091928571
#> 36:  0.07157857 0.0711 0.0632 -0.008485714  0.031978571  0.117857143
#> 37:  0.02860000 0.0706 0.0635 -0.054371429 -0.012964286  0.072757143
#> 38: -0.06136429 0.0676 0.0585 -0.146142857 -0.100571429 -0.021764286
#> 39: -0.15242857 0.0699 0.0610 -0.239642857 -0.194071429 -0.113142857
#> 40: -0.23414286 0.0695 0.0592 -0.327642857 -0.274214286 -0.196428571
#> 41: -0.29700000 0.0710 0.0597 -0.385785714 -0.337857143 -0.256928571
#> 42: -0.35985714 0.0678 0.0588 -0.454142857 -0.403071429 -0.323714286
#> 43: -0.38735714 0.0696 0.0623 -0.469857143 -0.426642857 -0.342571429
#> 44: -0.43528571 0.0757 0.0658 -0.528000000 -0.478500000 -0.389714286
#> 45: -0.51542857 0.0832 0.0735 -0.629357143 -0.564142857 -0.463571429
#> 46: -0.52092857 0.1030 0.0948 -0.649000000 -0.585357143 -0.458071429
#> 47: -0.53428571 0.1810 0.1620 -0.763714286 -0.645857143 -0.426642857
#> 48: -0.53821429 0.2380 0.2160 -0.840714286 -0.677285714 -0.392071429
#> 49: -0.54292857 0.2980 0.2670 -0.919285714 -0.722857143 -0.370071429
#> 50: -0.54292857 0.3470 0.3080 -0.990000000 -0.754285714 -0.346500000
#>          median     sd    mad           q5          q20          q80
#>             q95 rhat ess_bulk ess_tail
#>  1:  0.20114286 1.00      952     1190
#>  2:  0.16185714 1.00     1840     1450
#>  3:  0.13435714 1.01     1190     1230
#>  4:  0.17050000 1.00     1170     1220
#>  5:  0.12492857 1.00     1910     1440
#>  6:  0.02954286 1.00     2310     1580
#>  7: -0.05782857 1.00     2290     1630
#>  8: -0.14300000 1.00     2660     1870
#>  9: -0.19407143 1.00     2410     1850
#> 10: -0.26085714 1.00     1740     1800
#> 11: -0.25142857 1.00     2270     1920
#> 12: -0.26635714 1.00     1870     1710
#> 13: -0.30328571 1.00     1380     1240
#> 14: -0.22157143 1.00     2870     1750
#> 15: -0.06435000 1.00     2000     1240
#> 16:  0.06309286 1.00     2200     1620
#> 17:  0.19407143 1.00     2000     1390
#> 18:  0.30014286 1.00     2190     1510
#> 19:  0.58850000 1.00     1910     1700
#> 20:  0.49185714 1.00     2200     1650
#> 21:  0.40150000 1.00     2210     1540
#> 22:  0.31821429 1.00     2410     1550
#> 23:  0.25771429 1.00     2300     1720
#> 24:  0.18228571 1.00     1750     1780
#> 25:  0.17442857 1.00     2260     1890
#> 26:  0.12492857 1.00     1810     1710
#> 27:  0.04541429 1.00     1440     1250
#> 28:  0.06906429 1.00     2880     1640
#> 29:  0.16028571 1.00     1980     1120
#> 30:  0.23335714 1.00     2210     1510
#> 31:  0.30171429 1.00     1960     1450
#> 32:  0.36535714 1.00     2190     1480
#> 33:  0.20114286 1.00      952     1190
#> 34:  0.16185714 1.00     1840     1450
#> 35:  0.13435714 1.01     1190     1230
#> 36:  0.17050000 1.00     1170     1220
#> 37:  0.12335714 1.00     1910     1410
#> 38:  0.02710714 1.00     2310     1580
#> 39: -0.06167857 1.00     2290     1630
#> 40: -0.15085714 1.00     2660     1840
#> 41: -0.20742857 1.00     2410     1810
#> 42: -0.28521429 1.00     1700     1830
#> 43: -0.29385714 1.00     2240     1890
#> 44: -0.33392857 1.00     1760     1760
#> 45: -0.41800000 1.00     1380     1360
#> 46: -0.38657143 1.00     2650     1670
#> 47: -0.29228571 1.00     1930     1090
#> 48: -0.23650000 1.00     2170     1550
#> 49: -0.15714286 1.00     1910     1490
#> 50: -0.09821429 1.00     2160     1440
#>             q95 rhat ess_bulk ess_tail

# Rt summary
summary(posterior, type = "rt")
#>      variable       date     type observed forecast_start      mean    median
#>  1:  com_r[1] 2021-03-20 Combined     TRUE          FALSE 1.1119013 1.1136499
#>  2:  com_r[2] 2021-03-27 Combined     TRUE          FALSE 1.0825909 1.0834418
#>  3:  com_r[3] 2021-04-03 Combined     TRUE          FALSE 1.0507456 1.0565406
#>  4:  com_r[4] 2021-04-10 Combined     TRUE          FALSE 1.0775839 1.0742025
#>  5:  com_r[5] 2021-04-17 Combined     TRUE          FALSE 1.0320087 1.0304693
#>  6:  com_r[6] 2021-04-24 Combined     TRUE          FALSE 0.9438859 0.9427741
#>  7:  com_r[7] 2021-05-01 Combined     TRUE          FALSE 0.8626776 0.8626776
#>  8:  com_r[8] 2021-05-08 Combined     TRUE          FALSE 0.7968639 0.7981171
#>  9:  com_r[9] 2021-05-15 Combined     TRUE          FALSE 0.7542198 0.7536274
#> 10: com_r[10] 2021-05-22 Combined     TRUE          FALSE 0.7132972 0.7155425
#> 11: com_r[11] 2021-05-29 Combined     TRUE          FALSE 0.7105004 0.7093848
#> 12: com_r[12] 2021-06-05 Combined     TRUE          FALSE 0.6966804 0.6955864
#> 13: com_r[13] 2021-06-12 Combined     TRUE          FALSE 0.6703679 0.6714222
#> 14: com_r[14] 2021-06-19 Combined     TRUE           TRUE 0.7071588 0.7060484
#> 15: com_r[15] 2021-06-26 Combined    FALSE          FALSE 0.7459689 0.7447975
#> 16: com_r[16] 2021-07-03 Combined    FALSE          FALSE 0.7937395 0.7943634
#> 17: com_r[17] 2021-07-10 Combined    FALSE          FALSE 0.8294437 0.8314012
#> 18: com_r[18] 2021-07-17 Combined    FALSE          FALSE 0.8592951 0.8640343
#> 19:  voc_r[1] 2021-04-17      VOC     TRUE          FALSE 1.6366359 1.6315003
#> 20:  voc_r[2] 2021-04-24      VOC     TRUE          FALSE 1.4952385 1.4964138
#> 21:  voc_r[3] 2021-05-01      VOC     TRUE          FALSE 1.3639121 1.3649842
#> 22:  voc_r[4] 2021-05-08      VOC     TRUE          FALSE 1.2549195 1.2578810
#> 23:  voc_r[5] 2021-05-15      VOC     TRUE          FALSE 1.1803201 1.1812479
#> 24:  voc_r[6] 2021-05-22      VOC     TRUE          FALSE 1.1040663 1.1075417
#> 25:  voc_r[7] 2021-05-29      VOC     TRUE          FALSE 1.0808061 1.0804665
#> 26:  voc_r[8] 2021-06-05      VOC     TRUE           TRUE 1.0290129 1.0284471
#> 27:  voc_r[9] 2021-06-12      VOC    FALSE          FALSE 0.9474524 0.9485697
#> 28: voc_r[10] 2021-06-19      VOC    FALSE          FALSE 0.9433669 0.9408501
#> 29: voc_r[11] 2021-06-26      VOC    FALSE          FALSE 0.9310694 0.9293884
#> 30: voc_r[12] 2021-07-03      VOC    FALSE          FALSE 0.9288774 0.9294614
#> 31: voc_r[13] 2021-07-10      VOC    FALSE          FALSE 0.9208114 0.9215352
#> 32: voc_r[14] 2021-07-17      VOC    FALSE          FALSE 0.9164807 0.9200882
#> 33:      r[1] 2021-03-20  non-VOC     TRUE          FALSE 1.1119013 1.1136499
#> 34:      r[2] 2021-03-27  non-VOC     TRUE          FALSE 1.0825909 1.0834418
#> 35:      r[3] 2021-04-03  non-VOC     TRUE          FALSE 1.0507456 1.0565406
#> 36:      r[4] 2021-04-10  non-VOC     TRUE          FALSE 1.0775839 1.0742025
#> 37:      r[5] 2021-04-17  non-VOC     TRUE          FALSE 1.0305502 1.0290129
#> 38:      r[6] 2021-04-24  non-VOC     TRUE          FALSE 0.9415897 0.9404806
#> 39:      r[7] 2021-05-01  non-VOC     TRUE          FALSE 0.8586202 0.8586202
#> 40:      r[8] 2021-05-08  non-VOC     TRUE          FALSE 0.7900064 0.7912488
#> 41:      r[9] 2021-05-15  non-VOC     TRUE          FALSE 0.7430440 0.7430440
#> 42:     r[10] 2021-05-22  non-VOC     TRUE          FALSE 0.6955864 0.6977760
#> 43:     r[11] 2021-05-29  non-VOC     TRUE          FALSE 0.6804506 0.6788486
#> 44:     r[12] 2021-06-05  non-VOC     TRUE           TRUE 0.6480974 0.6470798
#> 45:     r[13] 2021-06-12  non-VOC    FALSE          FALSE 0.5967755 0.5972446
#> 46:     r[14] 2021-06-19  non-VOC    FALSE          FALSE 0.5939687 0.5939687
#> 47:     r[15] 2021-06-26  non-VOC    FALSE          FALSE 0.5860878 0.5860878
#> 48:     r[16] 2021-07-03  non-VOC    FALSE          FALSE 0.5847079 0.5837898
#> 49:     r[17] 2021-07-10  non-VOC    FALSE          FALSE 0.5796761 0.5810441
#> 50:     r[18] 2021-07-17  non-VOC    FALSE          FALSE 0.5774033 0.5810441
#>      variable       date     type observed forecast_start      mean    median
#>         sd    mad        q5       q20       q80       q95 rhat ess_bulk
#>  1: 0.0759 0.0750 1.0078881 1.0588676 1.1674078 1.2227994 1.00      952
#>  2: 0.0656 0.0561 0.9940464 1.0421945 1.1233168 1.1756923 1.00     1840
#>  3: 0.0710 0.0618 0.9528275 1.0075951 1.0962865 1.1438012 1.01     1190
#>  4: 0.0711 0.0632 0.9915502 1.0324954 1.1250834 1.1858977 1.00     1170
#>  5: 0.0706 0.0635 0.9484207 0.9885941 1.0769914 1.1330675 1.00     1910
#>  6: 0.0676 0.0586 0.8660733 0.9064546 0.9809341 1.0299836 1.00     2310
#>  7: 0.0699 0.0612 0.7900064 0.8274909 0.8972430 0.9438117 1.00     2290
#>  8: 0.0696 0.0589 0.7263047 0.7661654 0.8281413 0.8667541 1.00     2660
#>  9: 0.0709 0.0596 0.6890589 0.7240256 0.7844396 0.8235991 1.00     2410
#> 10: 0.0679 0.0591 0.6516718 0.6858181 0.7418773 0.7703910 1.00     1740
#> 11: 0.0691 0.0621 0.6537231 0.6825926 0.7412946 0.7776890 1.00     2270
#> 12: 0.0745 0.0661 0.6329995 0.6672151 0.7280187 0.7661654 1.00     1870
#> 13: 0.0803 0.0697 0.6010105 0.6405039 0.7032801 0.7383881 1.00     1380
#> 14: 0.0975 0.0886 0.6280454 0.6651214 0.7500830 0.8012587 1.00     2870
#> 15: 0.1760 0.1590 0.5995955 0.6708949 0.8307482 0.9376768 1.00     2000
#> 16: 0.2340 0.2120 0.5902469 0.6868967 0.9100227 1.0651257 1.00     2200
#> 17: 0.2950 0.2620 0.5751394 0.6944942 0.9871970 1.2141830 1.00     2000
#> 18: 0.3450 0.3020 0.5617411 0.6988734 1.0610329 1.3500517 1.00     2190
#> 19: 0.0723 0.0664 1.4975900 1.5674163 1.7102556 1.8012845 1.00     1910
#> 20: 0.0699 0.0630 1.3703572 1.4342511 1.5588191 1.6353505 1.00     2200
#> 21: 0.0728 0.0665 1.2392421 1.3052011 1.4241447 1.4940641 1.00     2210
#> 22: 0.0740 0.0635 1.1384217 1.2018441 1.3113687 1.3746708 1.00     2410
#> 23: 0.0741 0.0624 1.0707477 1.1303999 1.2324450 1.2939691 1.00     2300
#> 24: 0.0729 0.0659 1.0029981 1.0562916 1.1528237 1.1999570 1.00     1750
#> 25: 0.0728 0.0665 0.9882835 1.0346068 1.1312884 1.1905657 1.00     2260
#> 26: 0.0768 0.0672 0.9324604 0.9837900 1.0776686 1.1330675 1.00     1810
#> 27: 0.0811 0.0715 0.8498947 0.9036103 0.9944448 1.0464613 1.00     1440
#> 28: 0.0981 0.0905 0.8386180 0.8888230 1.0002695 1.0715051 1.00     2880
#> 29: 0.1770 0.1590 0.7471420 0.8359865 1.0352573 1.1738462 1.00     1980
#> 30: 0.2350 0.2110 0.6923150 0.8069448 1.0688144 1.2628324 1.00     2210
#> 31: 0.2950 0.2630 0.6400009 0.7722090 1.0954255 1.3521748 1.00     1960
#> 32: 0.3450 0.3030 0.5939687 0.7459689 1.1295120 1.4410286 1.00     2190
#> 33: 0.0759 0.0750 1.0078881 1.0588676 1.1674078 1.2227994 1.00      952
#> 34: 0.0656 0.0561 0.9940464 1.0421945 1.1233168 1.1756923 1.00     1840
#> 35: 0.0710 0.0618 0.9528275 1.0075951 1.0962865 1.1438012 1.01     1190
#> 36: 0.0711 0.0632 0.9915502 1.0324954 1.1250834 1.1858977 1.00     1170
#> 37: 0.0706 0.0635 0.9470803 0.9871194 1.0754693 1.1312884 1.00     1910
#> 38: 0.0676 0.0585 0.8640343 0.9043205 0.9784708 1.0274779 1.00     2310
#> 39: 0.0699 0.0610 0.7869088 0.8235991 0.8930231 0.9401850 1.00     2290
#> 40: 0.0695 0.0592 0.7206203 0.7601692 0.8216600 0.8599705 1.00     2660
#> 41: 0.0710 0.0597 0.6799162 0.7132972 0.7734235 0.8126713 1.00     2410
#> 42: 0.0678 0.0588 0.6349920 0.6682644 0.7234569 0.7518531 1.00     1700
#> 43: 0.0696 0.0623 0.6250916 0.6526966 0.7099424 0.7453830 1.00     2240
#> 44: 0.0757 0.0658 0.5897834 0.6197123 0.6772503 0.7161049 1.00     1760
#> 45: 0.0832 0.0735 0.5329343 0.5688475 0.6290331 0.6583622 1.00     1380
#> 46: 0.1030 0.0948 0.5225681 0.5569069 0.6325023 0.6793822 1.00     2650
#> 47: 0.1810 0.1620 0.4659326 0.5242130 0.6526966 0.7465552 1.00     1930
#> 48: 0.2380 0.2160 0.4314023 0.5079940 0.6756559 0.7893859 1.00     2170
#> 49: 0.2980 0.2670 0.3988038 0.4853635 0.6906850 0.8545820 1.00     1910
#> 50: 0.3470 0.3080 0.3715767 0.4703465 0.7071588 0.9064546 1.00     2160
#>         sd    mad        q5       q20       q80       q95 rhat ess_bulk
#>     ess_tail
#>  1:     1190
#>  2:     1450
#>  3:     1230
#>  4:     1220
#>  5:     1440
#>  6:     1580
#>  7:     1630
#>  8:     1870
#>  9:     1850
#> 10:     1800
#> 11:     1920
#> 12:     1710
#> 13:     1240
#> 14:     1750
#> 15:     1240
#> 16:     1620
#> 17:     1390
#> 18:     1510
#> 19:     1700
#> 20:     1650
#> 21:     1540
#> 22:     1550
#> 23:     1720
#> 24:     1780
#> 25:     1890
#> 26:     1710
#> 27:     1250
#> 28:     1640
#> 29:     1120
#> 30:     1510
#> 31:     1450
#> 32:     1480
#> 33:     1190
#> 34:     1450
#> 35:     1230
#> 36:     1220
#> 37:     1410
#> 38:     1580
#> 39:     1630
#> 40:     1840
#> 41:     1810
#> 42:     1830
#> 43:     1890
#> 44:     1760
#> 45:     1360
#> 46:     1670
#> 47:     1090
#> 48:     1550
#> 49:     1490
#> 50:     1440
#>     ess_tail

# model parameter summary
summary(posterior, type = "model")
#>             variable                  clean_name exponentiated         mean
#> 1: avg_voc_advantage          Average VOC effect         FALSE     0.589000
#> 2:              beta                        Beta         FALSE     0.203000
#> 3:     init_cases[1]               Initial cases          TRUE 89321.723361
#> 4:     init_cases[2]           Initial VOC cases          TRUE   219.203386
#> 5:            phi[1] Notification overdispersion         FALSE   103.000000
#> 6:            phi[2]   Sequencing overdispersion         FALSE   188.000000
#> 7:            r_init              Initial growth         FALSE     0.135000
#> 8:           r_scale                 Growth (sd)         FALSE     0.122000
#> 9:           voc_mod          Initial VOC effect          TRUE     1.802185
#>          median       sd     mad           q5          q20          q80
#> 1:     0.588000   0.0287  0.0273     0.544000     0.566000     0.611000
#> 2:     0.209000   0.4080  0.4480    -0.499000    -0.160000     0.582000
#> 3: 89321.723361   0.0670  0.0652 80821.637540 80821.637540 98715.771011
#> 4:   219.203386   0.0982  0.0948   186.792804   202.350228   237.460193
#> 5:    79.100000 121.0000 52.4000    23.500000    44.200000   137.000000
#> 6:   165.000000 111.0000 94.0000    55.400000    95.000000   266.000000
#> 7:     0.137000   0.0759  0.0750     0.010000     0.072800     0.197000
#> 8:     0.117000   0.0577  0.0573     0.040300     0.070800     0.170000
#> 9:     1.800384   0.0287  0.0273     1.722885     1.761208     1.842273
#>            q95 rhat ess_bulk ess_tail
#> 1:     0.63600    1     1670     1510
#> 2:     0.85300    1      892     1010
#> 3: 98715.77101    1     1570     1320
#> 4:   259.82284    1     2530     1550
#> 5:   232.00000    1      675      729
#> 6:   409.00000    1     2540     1510
#> 7:     0.25600    1      952     1190
#> 8:     0.22100    1      749      773
#> 9:     1.88891    1     1670     1510

# raw posterior values
summary(posterior, type = "raw")
#>         variable      mean    median     sd    mad        q5       q20
#>   1:        lp__ -3340.000 -3340.000 4.8700 4.8500 -3.35e+03 -3.35e+03
#>   2:      r_init     0.135     0.137 0.0759 0.0750  1.00e-02  7.28e-02
#>   3:     r_scale     0.122     0.117 0.0577 0.0573  4.03e-02  7.08e-02
#>   4:        beta     0.203     0.209 0.4080 0.4480 -4.99e-01 -1.60e-01
#>   5:      eta[1]    -0.304    -0.309 0.8030 0.7920 -1.58e+00 -9.73e-01
#>  ---                                                                  
#> 259: log_lik[11]   -14.700   -14.700 0.7240 0.6410 -1.60e+01 -1.52e+01
#> 260: log_lik[12]   -14.100   -14.000 0.7040 0.6900 -1.53e+01 -1.46e+01
#> 261: log_lik[13]   -13.600   -13.600 0.9010 0.7840 -1.52e+01 -1.43e+01
#> 262: log_lik[14]    -8.130    -8.040 0.5760 0.4690 -9.22e+00 -8.51e+00
#> 263: log_lik[15]    -7.710    -7.570 0.6990 0.5280 -8.96e+00 -8.12e+00
#>            q80       q95 rhat ess_bulk ess_tail
#>   1: -3340.000 -3330.000    1      402      684
#>   2:     0.197     0.256    1      952     1190
#>   3:     0.170     0.221    1      749      773
#>   4:     0.582     0.853    1      892     1010
#>   5:     0.361     1.000    1     1220     1310
#>  ---                                           
#> 259:   -14.100   -13.700    1     1110     1300
#> 260:   -13.500   -13.000    1      649      779
#> 261:   -12.900   -12.300    1      996     1190
#> 262:    -7.680    -7.410    1     1130      951
#> 263:    -7.190    -6.890    1      863      785