Label the Variant of Concern

Assign a custom label to the variant of concern in the output from fv_tidy_posterior().

Usage

update_voc_label(posterior, label, target_label = "VOC")

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

label

Character string indicating the new label to use for the variant of concern.

target_label

A character string defaulting to "VOC". Indicates the current label for the variant of concern.

Value

A list of data frames as returned by `fv_tidy_posterior() but with updated labels.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")
p <- update_voc_label(p, "Delta")
p[value_type == "model"]
#>    value_type      variable                  clean_name date type obs observed
#> 1:      model   avg_voc_mod        Average Delta effect <NA>       NA       NA
#> 2:      model          beta                        Beta <NA>       NA       NA
#> 3:      model init_cases[1]               Initial cases <NA>       NA       NA
#> 4:      model init_cases[2]         Initial Delta cases <NA>       NA       NA
#> 5:      model        phi[1] Notification overdispersion <NA>       NA       NA
#> 6:      model        phi[2]   Sequencing overdispersion <NA>       NA       NA
#> 7:      model        r_init              Initial growth <NA>       NA       NA
#> 8:      model       r_noise                 Growth (sd) <NA>       NA       NA
#> 9:      model       voc_mod        Initial Delta effect <NA>       NA       NA
#>    forecast_start exponentiated         mean       median           sd
#> 1:             NA          TRUE 1.804766e+00 1.804291e+00   0.02921097
#> 2:             NA         FALSE 3.209816e-01 3.984735e-01   0.46752521
#> 3:             NA          TRUE 8.959069e+04 8.965731e+04   0.06936367
#> 4:             NA          TRUE 2.200103e+02 2.201898e+02   0.09745460
#> 5:             NA         FALSE 1.056693e+02 8.444825e+01 106.02242241
#> 6:             NA         FALSE 1.891149e+02 1.615045e+02 124.32716962
#> 7:             NA         FALSE 1.325174e-01 1.341870e-01   0.07180019
#> 8:             NA         FALSE 1.092370e-01 9.991730e-02   0.05998836
#> 9:             NA          TRUE 1.804766e+00 1.804291e+00   0.02921097
#>            mad            q5           q20          q80          q95      rhat
#> 1:  0.02825984  1.719788e+00  1.762777e+00 1.847177e+00 1.895543e+00 0.9999955
#> 2:  0.51877731 -5.182967e-01 -1.379756e-01 7.690168e-01 9.354680e-01 1.0000555
#> 3:  0.06382593  7.974545e+04  8.474313e+04 9.463664e+04 1.002407e+05 1.0035276
#> 4:  0.09427853  1.877616e+02  2.027711e+02 2.384415e+02 2.594063e+02 1.0000300
#> 5: 53.64839991  2.663438e+01  4.717452e+01 1.411910e+02 2.330759e+02 1.0006074
#> 6: 93.93219864  5.013444e+01  9.281310e+01 2.663616e+02 4.393338e+02 1.0042635
#> 7:  0.06461401  1.351774e-02  7.735250e-02 1.884312e-01 2.496943e-01 1.0001805
#> 8:  0.06087993  3.048898e-02  5.322236e-02 1.579574e-01 2.195497e-01 0.9999600
#> 9:  0.02825984  1.719788e+00  1.762777e+00 1.847177e+00 1.895543e+00 0.9999955
#>     ess_bulk  ess_tail
#> 1: 2121.0506 1419.0800
#> 2:  521.4538 1072.1236
#> 3: 2338.2868 1108.6765
#> 4: 2524.9633 1311.7136
#> 5:  651.3741  910.9929
#> 6: 2584.5556 1247.1047
#> 7: 1327.3538  997.1388
#> 8:  574.6457 1183.9425
#> 9: 2121.0506 1419.0800