Forecast using branching processes at a target date

Forecast using branching processes at a target date

forecast(
  obs,
  forecast_date = max(obs$date),
  seq_date = forecast_date,
  case_date = forecast_date,
  horizon = 4,
  r_init = c(0, 0.25),
  voc_scale = c(0, 0.2),
  voc_label = "VOC",
  strains = 2,
  variant_relationship = "pooled",
  overdispersion = TRUE,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  debug = FALSE,
  keep_fit = TRUE,
  scale_r = 1,
  probs = c(0.05, 0.2, 0.8, 0.95),
  id = 0,
  ...
)

Arguments

obs
A data.frame with the following variables: date, cases, seq_voc, and seq_total, cases_available, and seq_available. seq_available and case_available must be uniquely define data rows but other rows can be duplicated based on data availability.

forecast_date
Date at which to forecast. Defaults to the maximum date in obs.

seq_date
Date from which to use available sequence data. Defaults to the date.

case_date
Date from which to use available case data. Defaults to the date.

horizon
Integer forecast horizon. Defaults to 4.

r_init
Numeric vector of length 2. Prior mean and standard deviation for the initial growht rate.

voc_scale
Numeric vector of length 2. Prior mean and standard deviation for the initial growth rate modifier due to the variant of concern.

voc_label
A character string, default to "VOC". Defines the label to assign to variant of concern specific parameters. Example usage is to rename parameters to use variant specific terminology.

strains
Integer number of strains to use. Defaults to 2. Current maximum is 2. A numeric vector can be passed if forecasts from multiple strain models are desired.

variant_relationship
Character string, defaulting to "pooled". Controls the relationship of strains with options being "pooled" (dependence determined from the data), "scaled" (a fixed scaling between strains), and "independent" (fully independent strains after initial scaling).

overdispersion
Logical, defaults to TRUE. Should the observations used include overdispersion.

models
A model as supplied by load_model(). If not supplied the default for that strain is used. If multiple strain models are being forecast then models should be a list models.

likelihood
Logical, defaults to TRUE. Should the likelihood be included in the model

output_loglik
Logical, defaults to FALSE. Should the log-likelihood be output. Disabling this will speed up fitting if evaluating the model fit is not required.

debug
Logical, defaults to FALSE. Should within model debug information be returned.

keep_fit
Logical, defaults to TRUE. Should the stan model fit be kept and returned. Dropping this can substantially reduce memory usage in situtations where multiple models are being fit.

scale_r
Numeric, defaults to 1. Rescale the timespan over which the growth rate and reproduction number is calculated. An example use case is rescaling the growth rate from weekly to be scaled by the mean of the generation time (for COVID-19 for example this would be 5.5 / 7.

probs
A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work (such as plot_cases(), plot_rt(), and plot_voc()).

id
ID to assign to this forecast. Defaults to 0.

...
Additional parameters passed to stan_fit().

Value

A data.frame containing the output of stan_fit() in each row as well as the summarised posterior, forecast and information about the parameters specified.

See also

Examples

if (FALSE) { # interactive()
options(mc.cores = 4)

forecasts <- forecast(
  germany_covid19_delta_obs,
  forecast_date = as.Date("2021-06-12"),
  horizon = 4,
  strains = c(1, 2),
  adapt_delta = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)
# inspect forecasts
forecasts

# unnest posteriors
forecasts <- unnest_posterior(forecasts)

# plot case posterior predictions
plot_cases(forecasts, log = TRUE)

# plot voc posterior predictions
plot_voc(forecasts)
}