Forecast using branching processes at a target date

Forecast using branching processes at a target date

forecast(
  obs,
  forecast_date = max(obs$date),
  seq_date = forecast_date,
  case_date = forecast_date,
  horizon = 4,
  r_init = c(0, 0.25),
  voc_scale = c(0, 0.2),
  strains = 2,
  variant_relationship = "pooled",
  overdispersion = TRUE,
  models = NULL,
  likelihood = TRUE,
  output_loglik = FALSE,
  debug = FALSE,
  keep_fit = TRUE,
  probs = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  id = 0,
  ...
)

Arguments

obs
A data frame with the following variables: date, cases, seq_voc, and seq_total, cases_available, and seq_available. seq_available and case_available must be uniquely define data rows but other rows can be duplicated based on data availability.

forecast_date
Date at which to forecast. Defaults to the maximum date in obs.

seq_date
Date from which to use available sequence data. Defaults to the forecast date.

case_date
Date from which to use available case data. Defaults to the forecast date.

horizon
Integer forecast horizon. Defaults to 4.

r_init
Numeric vector of length 2. Prior mean and standard deviation for the initial growht rate.

voc_scale
Numeric vector of length 2. Prior mean and standard deviation for the initial growth rate modifier due to the variant of concern.

strains
Integer number of strains to use. Defaults to 2. Current maximum is 2. A numeric vector can be passed if forecasts from multiple strain models are desired.

variant_relationship
Character string, defaulting to "pooled". Controls the relationship of strains with options being "pooled" (dependence determined from the data), "scaled" (a fixed scaling between strains), and "independent" (fully independent strains after initial scaling).

overdispersion
Logical, defaults to TRUE. Should the observations used include overdispersion.

models
A list of models to use named by strain. If not supplied uses the package default model for that strain.

likelihood
Logical, defaults to TRUE. Should the likelihood be included in the model.extract

output_loglik
Logical, defaults to FALSE. Should the log likelihood be output. Disabling this will speed up fitting if evaluating the model fit is not required.

debug
Logical, defaults to FALSE. Should within model debug information be returned.

keep_fit
Logical, defaults to TRUE. Should the stan model fit be kept and returned. Dropping this can substantially reduce memory usage in situtations where multiple models are being fit.

probs
A vector of numeric probabilities to produce quantile summaries for.

id
ID to assign to this forecast. Defaults to 0.

...
Additional parameters passed to stan_fit().

Value

A dataframe containing the output of stan_fit() in each row as well as the summarised posterior, forecast and information about the parameters specified.

Examples

if (FALSE) {
options(mc.cores = 4)
results <- forecast(
  latest_obs(germany_covid19_delta_obs),
  horizon = 4,
  save_path = tempdir(),
  strains = c(1, 2),
  adapt_delta = 0.99,
  max_treedepth = 15,
  variant_relationship = "scaled"
)
# inspect results
results

# unnest posteriors
results <- unnest_posterior(results)

# plot case posterior predictions
plot_cases(results, log = TRUE)
}