Forecast for a single model and summarise

Forecast for a single model and summarise

forecast_n_strain(
  data,
  model = NULL,
  strains = 2,
  voc_label = "VOC",
  probs = c(0.05, 0.2, 0.8, 0.95),
  scale_r = 1,
  ...
)

Arguments

data
A list of data as produced by stan_data().

model
A cmdstanr model object as loaded by load_model().

strains
Integer number of strains. Defaults to 2. Current maximum is 2.

voc_label
A character string, default to "VOC". Defines the label to assign to variant of concern specific parameters. Example usage is to rename parameters to use variant specific terminology.

probs
A vector of numeric probabilities to produce quantile summaries for. By default these are the 5%, 20%, 80%, and 95% quantiles which are also the minimum set required for plotting functions to work (such as plot_cases(), plot_rt(), and plot_voc()).

scale_r
Numeric, defaults to 1. Rescale the timespan over which the growth rate and reproduction number is calculated. An example use case is rescaling the growth rate from weekly to be scaled by the mean of the generation time (for COVID-19 for example this would be 5.5 / 7.

...
Additional parameters passed to stan_fit().

See also