Load a package example

Loads examples of posterior and forecast summaries produced using example scripts. Used to streamline examples, in package tests and to enable users to explore package functionality without needing to install cmdstanr.

Usage

fv_example(strains = 1, type = "posterior")

Arguments

strains

Integer number of strains. Defaults to 2. Current maximum is 2.

type

A character string indicating the example to load. Supported options are "posterior", "forecast", "observations", and "script" which are the output of fv_tidy_posterior(), fv_extract_forecast(), filter_by_availability (with the date argument set to "2021-08-26" applied to the germany_covid19_delta_obs package dataset), and the script used to generate these examples respectively.

Value

A data.table of summarised output

See also

Examples

# Load the summarised posterior from an example fit of the one strain model
fv_example(strains = 1, type = "posterior")
#> Available value types: model, cases, growth, rt, raw
#>      value_type     variable                  clean_name       date    type
#>   1:      model         beta                        Beta       <NA>        
#>   2:      model       phi[1] Notification overdispersion       <NA>        
#>   3:      model       r_init              Initial growth       <NA>        
#>   4:      model      r_noise                 Growth (sd)       <NA>        
#>   5:      cases sim_cases[1]                             2021-03-20 Overall
#>  ---                                                                       
#> 167:        raw  log_lik[11]                                   <NA>        
#> 168:        raw  log_lik[12]                                   <NA>        
#> 169:        raw  log_lik[13]                                   <NA>        
#> 170:        raw  log_lik[14]                                   <NA>        
#> 171:        raw  log_lik[15]                                   <NA>        
#>        obs observed forecast_start exponentiated      mean    median       sd
#>   1:    NA       NA             NA         FALSE     0.199     0.249 4.78e-01
#>   2:    NA       NA             NA         FALSE   110.000    83.100 1.12e+02
#>   3:    NA       NA             NA         FALSE     0.139     0.139 6.84e-02
#>   4:    NA       NA             NA         FALSE     0.111     0.104 5.52e-02
#>   5: 87328     TRUE          FALSE            NA 89200.000 88800.000 1.30e+04
#>  ---                                                                         
#> 167:    NA       NA             NA            NA    -9.570    -9.510 5.71e-01
#> 168:    NA       NA             NA            NA    -8.950    -8.910 4.61e-01
#> 169:    NA       NA             NA            NA    -8.940    -8.830 6.69e-01
#> 170:    NA       NA             NA            NA    -8.210    -8.100 6.56e-01
#> 171:    NA       NA             NA            NA    -7.640    -7.520 6.51e-01
#>           mad        q5       q20       q80       q95 rhat ess_bulk ess_tail
#>   1: 5.55e-01 -6.47e-01 -2.52e-01     0.669  8.88e-01 1.00      741     1290
#>   2: 5.21e+01  2.67e+01  4.81e+01   145.000  2.64e+02 1.01      537      369
#>   3: 6.55e-02  2.95e-02  8.14e-02     0.194  2.49e-01 1.00     1290     1190
#>   4: 5.84e-02  3.53e-02  5.94e-02     0.158  2.10e-01 1.00      795     1480
#>   5: 1.13e+04  6.93e+04  7.93e+04 98700.000  1.11e+05 1.00     2430     1640
#>  ---                                                                        
#> 167: 4.85e-01 -1.06e+01 -9.97e+00    -9.130 -8.76e+00 1.00      621      481
#> 168: 3.83e-01 -9.79e+00 -9.26e+00    -8.600 -8.32e+00 1.00      586      444
#> 169: 5.22e-01 -1.02e+01 -9.37e+00    -8.460 -8.08e+00 1.00      892      529
#> 170: 5.32e-01 -9.46e+00 -8.64e+00    -7.700 -7.38e+00 1.00      889      601
#> 171: 4.87e-01 -8.86e+00 -8.01e+00    -7.170 -6.86e+00 1.00      750      635

# Load the summarised forecast from this posterior
fv_example(strains = 1, type = "forecast")
#> Available value types: cases, growth, rt
#>     value_type    type       date horizon forecast_start         mean
#>  1:      cases Overall 2021-07-03       1          FALSE 3030.0000000
#>  2:      cases Overall 2021-07-10       2          FALSE 1900.0000000
#>  3:      cases Overall 2021-07-17       3          FALSE 1260.0000000
#>  4:      cases Overall 2021-07-24       4          FALSE  930.0000000
#>  5:     growth Overall 2021-06-26       1          FALSE   -0.4140714
#>  6:     growth Overall 2021-07-03       2          FALSE   -0.4109286
#>  7:     growth Overall 2021-07-10       3          FALSE   -0.4164286
#>  8:     growth Overall 2021-07-17       4          FALSE   -0.4164286
#>  9:         rt Overall 2021-06-26       1          FALSE    0.6609537
#> 10:         rt Overall 2021-07-03       2          FALSE    0.6630343
#> 11:         rt Overall 2021-07-10       3          FALSE    0.6593976
#> 12:         rt Overall 2021-07-17       4          FALSE    0.6593976
#>           median       sd     mad           q5          q20          q80
#>  1: 2910.0000000  793.000 665.000 1910.0000000 2420.0000000 3590.0000000
#>  2: 1750.0000000  870.000 638.000  903.0000000 1270.0000000 2380.0000000
#>  3: 1020.0000000 1000.000 544.000  384.0000000  644.0000000 1630.0000000
#>  4:  597.0000000 1420.000 436.000  156.0000000  306.0000000 1180.0000000
#>  5:   -0.4125000    0.159   0.143   -0.6136429   -0.5075714   -0.3166429
#>  6:   -0.4125000    0.206   0.182   -0.6647143   -0.5342857   -0.2867857
#>  7:   -0.4172143    0.256   0.227   -0.7275714   -0.5665000   -0.2647857
#>  8:   -0.4180000    0.302   0.254   -0.7857143   -0.5869286   -0.2412143
#>  9:    0.6619932    0.159   0.143    0.5413751    0.6019557    0.7285909
#> 10:    0.6619932    0.206   0.182    0.5144205    0.5860878    0.7506726
#> 11:    0.6588797    0.256   0.227    0.4830808    0.5675082    0.7673704
#> 12:    0.6583622    0.302   0.254    0.4557940    0.5560325    0.7856732
#>               q95
#>  1: 4490.00000000
#>  2: 3420.00000000
#>  3: 2990.00000000
#>  4: 2680.00000000
#>  5:   -0.20271429
#>  6:   -0.15321429
#>  7:   -0.09035714
#>  8:   -0.03912857
#>  9:    0.81651150
#> 10:    0.85794586
#> 11:    0.91360484
#> 12:    0.96162706

# Load the script used to generate these examples
# Optionally source this script to regenerate the example
readLines(fv_example(strains = 1, type = "script"))
#>  [1] "options(mc.cores = 2)"                                              
#>  [2] "obs <- filter_by_availability("                                     
#>  [3] "  germany_covid19_delta_obs,"                                       
#>  [4] "  date = \"2021-06-26\""                                            
#>  [5] ")"                                                                  
#>  [6] ""                                                                   
#>  [7] "current_obs <- filter_by_availability("                             
#>  [8] "  germany_covid19_delta_obs,"                                       
#>  [9] "  date = \"2021-08-26\""                                            
#> [10] ")"                                                                  
#> [11] ""                                                                   
#> [12] "dt <- fv_as_data_list("                                             
#> [13] "  obs,"                                                             
#> [14] "  overdispersion = TRUE,"                                           
#> [15] "  variant_relationship = \"scaled\","                               
#> [16] "  voc_scale = c(0.4, 0.2)"                                          
#> [17] ")"                                                                  
#> [18] ""                                                                   
#> [19] "inits <- fv_inits(dt, strains = 1)"                                 
#> [20] ""                                                                   
#> [21] "model <- suppressMessages(fv_model(strains = 1))"                   
#> [22] ""                                                                   
#> [23] "fit <- fv_sample("                                                  
#> [24] "  data = dt, model = model, init = inits,"                          
#> [25] "  adapt_delta = 0.99, max_treedepth = 15, chains = 2"               
#> [26] ")"                                                                  
#> [27] ""                                                                   
#> [28] "# summarise posterior assuming a mean generation time of  5.5 days."
#> [29] "posterior <- fv_tidy_posterior(fit, scale_r = 5.5 / 7)"             
#> [30] "forecast <- fv_extract_forecast(posterior)"