Extract forecasts from a summarised posterior

Uses the observed variable returned by fv_tidy_posterior() to return posterior predictions for forecast dates only.

Usage

fv_extract_forecast(posterior)

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

Value

A data.frame of forecasts in the format returned by fv_tidy_posterior() but with fitting variables dropped.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")

fv_extract_forecast(p)
#> Available value types: cases, voc_frac, voc_advantage, growth, rt
#>        value_type     type       date horizon forecast_start          mean
#>  1:         cases Combined 2021-07-03       1          FALSE 3470.00000000
#>  2:         cases Combined 2021-07-10       2          FALSE 2620.00000000
#>  3:         cases Combined 2021-07-17       3          FALSE 2300.00000000
#>  4:         cases Combined 2021-07-24       4          FALSE 2330.00000000
#>  5:         cases      VOC 2021-06-19       1          FALSE 2100.00000000
#>  6:         cases      VOC 2021-06-26       2          FALSE 1950.00000000
#>  7:         cases      VOC 2021-07-03       3          FALSE 1810.00000000
#>  8:         cases      VOC 2021-07-10       4          FALSE 1730.00000000
#>  9:         cases      VOC 2021-07-17       5          FALSE 1780.00000000
#> 10:         cases      VOC 2021-07-24       6          FALSE 2010.00000000
#> 11:         cases  non-VOC 2021-06-19       1          FALSE 6180.00000000
#> 12:         cases  non-VOC 2021-06-26       2          FALSE 3180.00000000
#> 13:         cases  non-VOC 2021-07-03       3          FALSE 1660.00000000
#> 14:         cases  non-VOC 2021-07-10       4          FALSE  895.00000000
#> 15:         cases  non-VOC 2021-07-17       5          FALSE  514.00000000
#> 16:         cases  non-VOC 2021-07-24       6          FALSE  328.00000000
#> 17:      voc_frac      VOC 2021-06-19       1          FALSE    0.25300000
#> 18:      voc_frac      VOC 2021-06-26       2          FALSE    0.37800000
#> 19:      voc_frac      VOC 2021-07-03       3          FALSE    0.52000000
#> 20:      voc_frac      VOC 2021-07-10       4          FALSE    0.65800000
#> 21:      voc_frac      VOC 2021-07-17       5          FALSE    0.77400000
#> 22:      voc_frac      VOC 2021-07-24       6          FALSE    0.85800000
#> 23: voc_advantage      VOC 2021-06-12       1          FALSE    1.58849291
#> 24: voc_advantage      VOC 2021-06-19       2          FALSE    1.58849291
#> 25: voc_advantage      VOC 2021-06-26       3          FALSE    1.58849291
#> 26: voc_advantage      VOC 2021-07-03       4          FALSE    1.58849291
#> 27: voc_advantage      VOC 2021-07-10       5          FALSE    1.58849291
#> 28: voc_advantage      VOC 2021-07-17       6          FALSE    1.58849291
#> 29:        growth Combined 2021-06-26       1          FALSE   -0.29700000
#> 30:        growth Combined 2021-07-03       2          FALSE   -0.24121429
#> 31:        growth Combined 2021-07-10       3          FALSE   -0.19014286
#> 32:        growth Combined 2021-07-17       4          FALSE   -0.15635714
#> 33:        growth      VOC 2021-06-12       1          FALSE   -0.05272143
#> 34:        growth      VOC 2021-06-19       2          FALSE   -0.06222857
#> 35:        growth      VOC 2021-06-26       3          FALSE   -0.07613571
#> 36:        growth      VOC 2021-07-03       4          FALSE   -0.08485714
#> 37:        growth      VOC 2021-07-10       5          FALSE   -0.08721429
#> 38:        growth      VOC 2021-07-17       6          FALSE   -0.09192857
#> 39:        growth  non-VOC 2021-06-12       1          FALSE   -0.51542857
#> 40:        growth  non-VOC 2021-06-19       2          FALSE   -0.52485714
#> 41:        growth  non-VOC 2021-06-26       3          FALSE   -0.53900000
#> 42:        growth  non-VOC 2021-07-03       4          FALSE   -0.54764286
#> 43:        growth  non-VOC 2021-07-10       5          FALSE   -0.54921429
#> 44:        growth  non-VOC 2021-07-17       6          FALSE   -0.55471429
#> 45:            rt Combined 2021-06-26       1          FALSE    0.74304401
#> 46:            rt Combined 2021-07-03       2          FALSE    0.78567325
#> 47:            rt Combined 2021-07-10       3          FALSE    0.82684101
#> 48:            rt Combined 2021-07-17       4          FALSE    0.85525369
#> 49:            rt      VOC 2021-06-12       1          FALSE    0.94864424
#> 50:            rt      VOC 2021-06-19       2          FALSE    0.93966808
#> 51:            rt      VOC 2021-06-26       3          FALSE    0.92669043
#> 52:            rt      VOC 2021-07-03       4          FALSE    0.91864351
#> 53:            rt      VOC 2021-07-10       5          FALSE    0.91648069
#> 54:            rt      VOC 2021-07-17       6          FALSE    0.91217030
#> 55:            rt  non-VOC 2021-06-12       1          FALSE    0.59724458
#> 56:            rt  non-VOC 2021-06-19       2          FALSE    0.59163988
#> 57:            rt  non-VOC 2021-06-26       3          FALSE    0.58333129
#> 58:            rt  non-VOC 2021-07-03       4          FALSE    0.57831137
#> 59:            rt  non-VOC 2021-07-10       5          FALSE    0.57740331
#> 60:            rt  non-VOC 2021-07-17       6          FALSE    0.57423631
#>        value_type     type       date horizon forecast_start          mean
#>            median       sd      mad           q5          q20           q80
#>  1: 3320.00000000 9.90e+02 8.30e+02 2160.0000000 2710.0000000  4.130000e+03
#>  2: 2360.00000000 1.31e+03 1.00e+03 1130.0000000 1630.0000000  3.380000e+03
#>  3: 1790.00000000 1.89e+03 1.10e+03  529.0000000 1050.0000000  3.090000e+03
#>  4: 1450.00000000 2.93e+03 1.19e+03  259.0000000  673.0000000  3.150000e+03
#>  5: 2050.00000000 5.04e+02 4.50e+02 1380.0000000 1700.0000000  2.450000e+03
#>  6: 1880.00000000 5.05e+02 4.29e+02 1260.0000000 1560.0000000  2.290000e+03
#>  7: 1720.00000000 6.18e+02 5.24e+02  980.0000000 1320.0000000  2.240000e+03
#>  8: 1550.00000000 9.07e+02 6.96e+02  687.0000000 1030.0000000  2.290000e+03
#>  9: 1380.00000000 1.50e+03 8.72e+02  395.0000000  789.0000000  2.400000e+03
#> 10: 1240.00000000 2.52e+03 1.02e+03  220.0000000  572.0000000  2.740000e+03
#> 11: 6090.00000000 1.06e+03 9.41e+02 4660.0000000 5350.0000000  6.930000e+03
#> 12: 3120.00000000 6.60e+02 5.49e+02 2260.0000000 2670.0000000  3.630000e+03
#> 13: 1590.00000000 5.37e+02 4.40e+02  938.0000000 1240.0000000  2.000000e+03
#> 14:  790.00000000 4.99e+02 3.55e+02  352.0000000  531.0000000  1.170000e+03
#> 15:  388.00000000 4.53e+02 2.59e+02  117.0000000  218.0000000  7.020000e+02
#> 16:  194.00000000 4.56e+02 1.70e+02   33.0000000   85.0000000  4.380000e+02
#> 17:    0.25000000 4.17e-02 3.63e-02    0.1910000    0.2190000  2.830000e-01
#> 18:    0.37600000 5.78e-02 5.14e-02    0.2900000    0.3300000  4.210000e-01
#> 19:    0.52000000 6.78e-02 6.08e-02    0.4120000    0.4650000  5.730000e-01
#> 20:    0.66200000 6.75e-02 6.08e-02    0.5480000    0.6050000  7.120000e-01
#> 21:    0.77800000 5.81e-02 5.16e-02    0.6760000    0.7290000  8.200000e-01
#> 22:    0.86300000 4.48e-02 3.84e-02    0.7810000    0.8250000  8.930000e-01
#> 23:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 24:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 25:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 26:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 27:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 28:    1.58849291 3.08e-02 2.97e-02    1.5296997    1.5575948  1.618732e+00
#> 29:   -0.29542857 1.71e-01 1.60e-01   -0.5177857   -0.4015000 -1.901429e-01
#> 30:   -0.24200000 2.35e-01 2.20e-01   -0.5405714   -0.3802857 -9.192857e-02
#> 31:   -0.18935714 2.84e-01 2.66e-01   -0.5531429   -0.3637857 -4.392143e-03
#> 32:   -0.15085714 3.38e-01 3.07e-01   -0.5916429   -0.3590714  4.698571e-02
#> 33:   -0.04957857 8.02e-02 7.29e-02   -0.1626429   -0.0990000 -3.370714e-03
#> 34:   -0.06317143 1.02e-01 9.55e-02   -0.1893571   -0.1249286 -1.815000e-04
#> 35:   -0.07637143 1.70e-01 1.60e-01   -0.2938571   -0.1807143  3.166429e-02
#> 36:   -0.08485714 2.34e-01 2.19e-01   -0.3881429   -0.2231429  6.505714e-02
#> 37:   -0.08642857 2.84e-01 2.68e-01   -0.4549286   -0.2608571  9.507143e-02
#> 38:   -0.08485714 3.38e-01 3.11e-01   -0.5287857   -0.2954286  1.170714e-01
#> 39:   -0.51307143 8.31e-02 7.58e-02   -0.6277857   -0.5657143 -4.635714e-01
#> 40:   -0.52642857 1.07e-01 1.01e-01   -0.6592143   -0.5900714 -4.604286e-01
#> 41:   -0.53900000 1.73e-01 1.65e-01   -0.7621429   -0.6458571 -4.305714e-01
#> 42:   -0.54607143 2.37e-01 2.23e-01   -0.8564286   -0.6890714 -3.944286e-01
#> 43:   -0.55157143 2.86e-01 2.73e-01   -0.9192857   -0.7252143 -3.669286e-01
#> 44:   -0.55000000 3.40e-01 3.15e-01   -1.0057143   -0.7597857 -3.441429e-01
#> 45:    0.74421257 1.71e-01 1.60e-01    0.5958384    0.6693153  8.268410e-01
#> 46:    0.78505618 2.35e-01 2.20e-01    0.5824153    0.6836660  9.121703e-01
#> 47:    0.82749092 2.84e-01 2.66e-01    0.5751394    0.6950401  9.956175e-01
#> 48:    0.85997054 3.38e-01 3.07e-01    0.5534174    0.6983245  1.048107e+00
#> 49:    0.95163038 8.02e-02 7.29e-02    0.8498947    0.9057427  9.966350e-01
#> 50:    0.93878253 1.02e-01 9.55e-02    0.8274909    0.8825599  9.998185e-01
#> 51:    0.92647202 1.70e-01 1.60e-01    0.7453830    0.8346738  1.032171e+00
#> 52:    0.91864351 2.34e-01 2.19e-01    0.6783154    0.8000006  1.067220e+00
#> 53:    0.91720106 2.84e-01 2.68e-01    0.6344933    0.7703910  1.099737e+00
#> 54:    0.91864351 3.38e-01 3.11e-01    0.5893201    0.7442126  1.124200e+00
#> 55:    0.59865403 8.31e-02 7.58e-02    0.5337724    0.5679543  6.290331e-01
#> 56:    0.59071089 1.07e-01 1.01e-01    0.5172576    0.5542877  6.310132e-01
#> 57:    0.58333129 1.73e-01 1.65e-01    0.4666654    0.5242130  6.501375e-01
#> 58:    0.57922086 2.37e-01 2.23e-01    0.4246761    0.5020420  6.740651e-01
#> 59:    0.57604389 2.86e-01 2.73e-01    0.3988038    0.4842208  6.928591e-01
#> 60:    0.57694981 3.40e-01 3.15e-01    0.3657833    0.4677667  7.088277e-01
#>            median       sd      mad           q5          q20           q80
#>               q95
#>  1: 5280.00000000
#>  2: 5150.00000000
#>  3: 5790.00000000
#>  4: 7380.00000000
#>  5: 2970.00000000
#>  6: 2800.00000000
#>  7: 2970.00000000
#>  8: 3400.00000000
#>  9: 4580.00000000
#> 10: 6400.00000000
#> 11: 7930.00000000
#> 12: 4330.00000000
#> 13: 2640.00000000
#> 14: 1790.00000000
#> 15: 1310.00000000
#> 16: 1010.00000000
#> 17:    0.32500000
#> 18:    0.47500000
#> 19:    0.63100000
#> 20:    0.76300000
#> 21:    0.86000000
#> 22:    0.92100000
#> 23:    1.65214002
#> 24:    1.65214002
#> 25:    1.65214002
#> 26:    1.65214002
#> 27:    1.65214002
#> 28:    1.65214002
#> 29:   -0.07542857
#> 30:    0.04635714
#> 31:    0.16971429
#> 32:    0.26635714
#> 33:    0.04447143
#> 34:    0.06859286
#> 35:    0.14535714
#> 36:    0.20664286
#> 37:    0.27185714
#> 38:    0.33707143
#> 39:   -0.41564286
#> 40:   -0.38185714
#> 41:   -0.31821429
#> 42:   -0.25378571
#> 43:   -0.18857143
#> 44:   -0.13042857
#> 45:    0.92734597
#> 46:    1.04744843
#> 47:    1.18496624
#> 48:    1.30520112
#> 49:    1.04547511
#> 50:    1.07100007
#> 51:    1.15645252
#> 52:    1.22954337
#> 53:    1.31239950
#> 54:    1.40083912
#> 55:    0.65991591
#> 56:    0.68259256
#> 57:    0.72744689
#> 58:    0.77585804
#> 59:    0.82814135
#> 60:    0.87771918
#>               q95