Extract forecasts from a summarised posterior

Uses the observed variable returned by fv_tidy_posterior() to return posterior predictions for forecast dates only.

Usage

fv_extract_forecast(posterior)

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

Value

A data.frame of forecasts in the format returned by fv_tidy_posterior() but with fitting variables dropped.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")

fv_extract_forecast(p)
#> Available value types: cases, voc_frac, voc_advantage, growth, rt
#>        value_type     type       date horizon forecast_start          mean
#>  1:         cases Combined 2021-07-03       1          FALSE 3400.00000000
#>  2:         cases Combined 2021-07-10       2          FALSE 2510.00000000
#>  3:         cases Combined 2021-07-17       3          FALSE 2130.00000000
#>  4:         cases Combined 2021-07-24       4          FALSE 2160.00000000
#>  5:         cases      VOC 2021-06-19       1          FALSE 2110.00000000
#>  6:         cases      VOC 2021-06-26       2          FALSE 1950.00000000
#>  7:         cases      VOC 2021-07-03       3          FALSE 1780.00000000
#>  8:         cases      VOC 2021-07-10       4          FALSE 1670.00000000
#>  9:         cases      VOC 2021-07-17       5          FALSE 1660.00000000
#> 10:         cases      VOC 2021-07-24       6          FALSE 1860.00000000
#> 11:         cases  non-VOC 2021-06-19       1          FALSE 6200.00000000
#> 12:         cases  non-VOC 2021-06-26       2          FALSE 3190.00000000
#> 13:         cases  non-VOC 2021-07-03       3          FALSE 1620.00000000
#> 14:         cases  non-VOC 2021-07-10       4          FALSE  837.00000000
#> 15:         cases  non-VOC 2021-07-17       5          FALSE  467.00000000
#> 16:         cases  non-VOC 2021-07-24       6          FALSE  296.00000000
#> 17:      voc_frac      VOC 2021-06-19       1          FALSE    0.25400000
#> 18:      voc_frac      VOC 2021-06-26       2          FALSE    0.37900000
#> 19:      voc_frac      VOC 2021-07-03       3          FALSE    0.52300000
#> 20:      voc_frac      VOC 2021-07-10       4          FALSE    0.66200000
#> 21:      voc_frac      VOC 2021-07-17       5          FALSE    0.77700000
#> 22:      voc_frac      VOC 2021-07-24       6          FALSE    0.86100000
#> 23: voc_advantage      VOC 2021-06-19       1          FALSE    1.59099108
#> 24: voc_advantage      VOC 2021-06-26       2          FALSE    1.59099108
#> 25: voc_advantage      VOC 2021-07-03       3          FALSE    1.59099108
#> 26: voc_advantage      VOC 2021-07-10       4          FALSE    1.59099108
#> 27: voc_advantage      VOC 2021-07-17       5          FALSE    1.59099108
#> 28: voc_advantage      VOC 2021-07-24       6          FALSE    1.59099108
#> 29:        growth Combined 2021-06-26       1          FALSE   -0.30800000
#> 30:        growth Combined 2021-07-03       2          FALSE   -0.25378571
#> 31:        growth Combined 2021-07-10       3          FALSE   -0.21292857
#> 32:        growth Combined 2021-07-17       4          FALSE   -0.18228571
#> 33:        growth      VOC 2021-06-12       1          FALSE   -0.04816429
#> 34:        growth      VOC 2021-06-19       2          FALSE   -0.06442857
#> 35:        growth      VOC 2021-06-26       3          FALSE   -0.08800000
#> 36:        growth      VOC 2021-07-03       4          FALSE   -0.09742857
#> 37:        growth      VOC 2021-07-10       5          FALSE   -0.11078571
#> 38:        growth      VOC 2021-07-17       6          FALSE   -0.11864286
#> 39:        growth  non-VOC 2021-06-12       1          FALSE   -0.51228571
#> 40:        growth  non-VOC 2021-06-19       2          FALSE   -0.52878571
#> 41:        growth  non-VOC 2021-06-26       3          FALSE   -0.55235714
#> 42:        growth  non-VOC 2021-07-03       4          FALSE   -0.56178571
#> 43:        growth  non-VOC 2021-07-10       5          FALSE   -0.57514286
#> 44:        growth  non-VOC 2021-07-17       6          FALSE   -0.58300000
#> 45:            rt Combined 2021-06-26       1          FALSE    0.73491532
#> 46:            rt Combined 2021-07-03       2          FALSE    0.77585804
#> 47:            rt Combined 2021-07-10       3          FALSE    0.80821386
#> 48:            rt Combined 2021-07-17       4          FALSE    0.83336320
#> 49:            rt      VOC 2021-06-12       1          FALSE    0.95297721
#> 50:            rt      VOC 2021-06-19       2          FALSE    0.93760308
#> 51:            rt      VOC 2021-06-26       3          FALSE    0.91576088
#> 52:            rt      VOC 2021-07-03       4          FALSE    0.90716714
#> 53:            rt      VOC 2021-07-10       5          FALSE    0.89513054
#> 54:            rt      VOC 2021-07-17       6          FALSE    0.88812493
#> 55:            rt  non-VOC 2021-06-12       1          FALSE    0.59912458
#> 56:            rt  non-VOC 2021-06-19       2          FALSE    0.58932014
#> 57:            rt  non-VOC 2021-06-26       3          FALSE    0.57559146
#> 58:            rt  non-VOC 2021-07-03       4          FALSE    0.57018996
#> 59:            rt  non-VOC 2021-07-10       5          FALSE    0.56262449
#> 60:            rt  non-VOC 2021-07-17       6          FALSE    0.55822119
#>        value_type     type       date horizon forecast_start          mean
#>            median       sd      mad           q5           q20           q80
#>  1: 3300.00000000 9.00e+02 7.89e+02 2170.0000000 2690.00000000  4.020000e+03
#>  2: 2300.00000000 1.16e+03 9.05e+02 1120.0000000 1630.00000000  3.210000e+03
#>  3: 1710.00000000 2.14e+03 1.03e+03  551.0000000  999.00000000  2.880000e+03
#>  4: 1340.00000000 6.26e+03 1.11e+03  256.0000000  611.00000000  2.850000e+03
#>  5: 2080.00000000 4.90e+02 4.51e+02 1420.0000000 1710.00000000  2.460000e+03
#>  6: 1910.00000000 4.64e+02 4.22e+02 1300.0000000 1570.00000000  2.290000e+03
#>  7: 1700.00000000 5.75e+02 4.97e+02 1040.0000000 1320.00000000  2.170000e+03
#>  8: 1520.00000000 8.44e+02 6.30e+02  703.0000000 1050.00000000  2.160000e+03
#>  9: 1320.00000000 1.76e+03 8.24e+02  417.0000000  744.00000000  2.250000e+03
#> 10: 1150.00000000 5.59e+03 9.59e+02  222.0000000  522.00000000  2.450000e+03
#> 11: 6130.00000000 1.04e+03 9.01e+02 4680.0000000 5400.00000000  6.940000e+03
#> 12: 3130.00000000 6.64e+02 5.44e+02 2290.0000000 2700.00000000  3.620000e+03
#> 13: 1550.00000000 4.98e+02 4.23e+02  958.0000000 1240.00000000  1.960000e+03
#> 14:  753.00000000 4.15e+02 3.22e+02  351.0000000  517.00000000  1.090000e+03
#> 15:  368.00000000 4.32e+02 2.32e+02  113.0000000  212.00000000  6.410000e+02
#> 16:  177.00000000 7.09e+02 1.51e+02   32.0000000   79.00000000  3.810000e+02
#> 17:    0.25200000 3.95e-02 3.71e-02    0.1930000    0.22200000  2.850000e-01
#> 18:    0.37800000 5.44e-02 5.22e-02    0.2930000    0.33500000  4.230000e-01
#> 19:    0.52400000 6.34e-02 6.14e-02    0.4190000    0.47100000  5.740000e-01
#> 20:    0.66500000 6.25e-02 6.03e-02    0.5560000    0.61100000  7.140000e-01
#> 21:    0.78200000 5.31e-02 5.00e-02    0.6830000    0.73600000  8.210000e-01
#> 22:    0.86600000 4.02e-02 3.70e-02    0.7890000    0.83100000  8.950000e-01
#> 23:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 24:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 25:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 26:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 27:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 28:    1.59099108 2.83e-02 2.65e-02    1.5345149    1.56249783  1.620004e+00
#> 29:   -0.30878571 1.64e-01 1.49e-01   -0.5170000   -0.40542857 -2.105714e-01
#> 30:   -0.25378571 2.23e-01 1.99e-01   -0.5366429   -0.38657143 -1.210000e-01
#> 31:   -0.20350000 2.76e-01 2.54e-01   -0.5743571   -0.38185714 -4.612143e-02
#> 32:   -0.17521429 3.35e-01 2.96e-01   -0.6199286   -0.37478571  1.830714e-02
#> 33:   -0.04604286 7.73e-02 6.53e-02   -0.1492857   -0.09035714 -3.912857e-03
#> 34:   -0.06702143 9.50e-02 8.78e-02   -0.1815000   -0.12414286 -6.262143e-03
#> 35:   -0.08721429 1.62e-01 1.47e-01   -0.2962143   -0.18385714  9.585714e-03
#> 36:   -0.09742857 2.22e-01 1.98e-01   -0.3795000   -0.22942857  3.504286e-02
#> 37:   -0.10450000 2.76e-01 2.55e-01   -0.4761429   -0.27657143  5.649286e-02
#> 38:   -0.11078571 3.36e-01 2.97e-01   -0.5555000   -0.31035714  8.250000e-02
#> 39:   -0.51071429 7.87e-02 6.84e-02   -0.6144286   -0.55864286 -4.651429e-01
#> 40:   -0.52957143 9.89e-02 9.07e-02   -0.6521429   -0.59007143 -4.675000e-01
#> 41:   -0.55314286 1.65e-01 1.49e-01   -0.7676429   -0.64821429 -4.510000e-01
#> 42:   -0.56178571 2.26e-01 2.01e-01   -0.8485714   -0.69692857 -4.242857e-01
#> 43:   -0.56807143 2.79e-01 2.55e-01   -0.9428571   -0.74328571 -4.070000e-01
#> 44:   -0.57278571 3.39e-01 2.97e-01   -1.0214286   -0.77707143 -3.810714e-01
#> 45:    0.73433811 1.64e-01 1.49e-01    0.5963068    0.66669102  8.101212e-01
#> 46:    0.77585804 2.23e-01 1.99e-01    0.5847079    0.67938220  8.860340e-01
#> 47:    0.81587020 2.76e-01 2.54e-01    0.5630667    0.68259256  9.549260e-01
#> 48:    0.83927716 3.35e-01 2.96e-01    0.5379829    0.68743657  1.018476e+00
#> 49:    0.95500103 7.73e-02 6.53e-02    0.8613230    0.91360484  9.960948e-01
#> 50:    0.93517516 9.50e-02 8.78e-02    0.8340182    0.88325365  9.937574e-01
#> 51:    0.91648069 1.62e-01 1.47e-01    0.7436281    0.83205466  1.009632e+00
#> 52:    0.90716714 2.22e-01 1.98e-01    0.6842034    0.79498775  1.035664e+00
#> 53:    0.90077480 2.76e-01 2.55e-01    0.6211747    0.75837945  1.058119e+00
#> 54:    0.89513054 3.36e-01 2.97e-01    0.5737853    0.73318506  1.085999e+00
#> 55:    0.60006681 7.87e-02 6.84e-02    0.5409499    0.57198480  6.280454e-01
#> 56:    0.58885728 9.89e-02 9.07e-02    0.5209283    0.55428769  6.265667e-01
#> 57:    0.57513939 1.65e-01 1.49e-01    0.4641057    0.52297883  6.369908e-01
#> 58:    0.57018996 2.26e-01 2.01e-01    0.4280260    0.49811287  6.542369e-01
#> 59:    0.56661715 2.79e-01 2.55e-01    0.3895133    0.47554883  6.656442e-01
#> 60:    0.56395224 3.39e-01 2.97e-01    0.3600802    0.45975045  6.831291e-01
#>            median       sd      mad           q5           q20           q80
#>               q95
#>  1: 5010.00000000
#>  2: 4580.00000000
#>  3: 4920.00000000
#>  4: 5870.00000000
#>  5: 2950.00000000
#>  6: 2770.00000000
#>  7: 2810.00000000
#>  8: 3090.00000000
#>  9: 3840.00000000
#> 10: 5170.00000000
#> 11: 7860.00000000
#> 12: 4290.00000000
#> 13: 2490.00000000
#> 14: 1580.00000000
#> 15: 1080.00000000
#> 16:  827.00000000
#> 17:    0.31900000
#> 18:    0.46800000
#> 19:    0.62500000
#> 20:    0.75900000
#> 21:    0.85600000
#> 22:    0.91800000
#> 23:    1.64954584
#> 24:    1.64954584
#> 25:    1.64954584
#> 26:    1.64954584
#> 27:    1.64954584
#> 28:    1.64954584
#> 29:   -0.10292857
#> 30:    0.02530000
#> 31:    0.12414286
#> 32:    0.21371429
#> 33:    0.04140714
#> 34:    0.05665000
#> 35:    0.11628571
#> 36:    0.17914286
#> 37:    0.21921429
#> 38:    0.27657143
#> 39:   -0.42114286
#> 40:   -0.40071429
#> 41:   -0.34335714
#> 42:   -0.28600000
#> 43:   -0.24042857
#> 44:   -0.18228571
#> 45:    0.90219141
#> 46:    1.02562276
#> 47:    1.13217760
#> 48:    1.23826881
#> 49:    1.04227637
#> 50:    1.05828535
#> 51:    1.12331677
#> 52:    1.19619162
#> 53:    1.24509805
#> 54:    1.31860114
#> 55:    0.65629634
#> 56:    0.66984142
#> 57:    0.70938481
#> 58:    0.75126262
#> 59:    0.78629081
#> 60:    0.83336320
#>               q95