Extract forecasts from a summarised posterior

Uses the observed variable returned by fv_tidy_posterior() to return posterior predictions for forecast dates only.

Usage

fv_extract_forecast(posterior)

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

Value

A data.frame of forecasts in the format returned by fv_tidy_posterior() but with fitting variables dropped.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")

fv_extract_forecast(p)
#> Available value types: cases, voc_frac, voc_advantage, growth, rt
#>        value_type     type       date horizon forecast_start          mean
#>  1:         cases Combined 2021-07-03       1          FALSE 3470.00000000
#>  2:         cases Combined 2021-07-10       2          FALSE 2660.00000000
#>  3:         cases Combined 2021-07-17       3          FALSE 2370.00000000
#>  4:         cases Combined 2021-07-24       4          FALSE 2540.00000000
#>  5:         cases      VOC 2021-06-19       1          FALSE 2070.00000000
#>  6:         cases      VOC 2021-06-26       2          FALSE 1920.00000000
#>  7:         cases      VOC 2021-07-03       3          FALSE 1800.00000000
#>  8:         cases      VOC 2021-07-10       4          FALSE 1750.00000000
#>  9:         cases      VOC 2021-07-17       5          FALSE 1840.00000000
#> 10:         cases      VOC 2021-07-24       6          FALSE 2180.00000000
#> 11:         cases  non-VOC 2021-06-19       1          FALSE 6190.00000000
#> 12:         cases  non-VOC 2021-06-26       2          FALSE 3190.00000000
#> 13:         cases  non-VOC 2021-07-03       3          FALSE 1670.00000000
#> 14:         cases  non-VOC 2021-07-10       4          FALSE  910.00000000
#> 15:         cases  non-VOC 2021-07-17       5          FALSE  534.00000000
#> 16:         cases  non-VOC 2021-07-24       6          FALSE  361.00000000
#> 17:      voc_frac      VOC 2021-06-19       1          FALSE    0.25100000
#> 18:      voc_frac      VOC 2021-06-26       2          FALSE    0.37500000
#> 19:      voc_frac      VOC 2021-07-03       3          FALSE    0.51700000
#> 20:      voc_frac      VOC 2021-07-10       4          FALSE    0.65600000
#> 21:      voc_frac      VOC 2021-07-17       5          FALSE    0.77200000
#> 22:      voc_frac      VOC 2021-07-24       6          FALSE    0.85800000
#> 23: voc_advantage      VOC 2021-06-12       1          FALSE    1.58849291
#> 24: voc_advantage      VOC 2021-06-19       2          FALSE    1.58849291
#> 25: voc_advantage      VOC 2021-06-26       3          FALSE    1.58849291
#> 26: voc_advantage      VOC 2021-07-03       4          FALSE    1.58849291
#> 27: voc_advantage      VOC 2021-07-10       5          FALSE    1.58849291
#> 28: voc_advantage      VOC 2021-07-17       6          FALSE    1.58849291
#> 29:        growth Combined 2021-06-26       1          FALSE   -0.29307143
#> 30:        growth Combined 2021-07-03       2          FALSE   -0.23100000
#> 31:        growth Combined 2021-07-10       3          FALSE   -0.18700000
#> 32:        growth Combined 2021-07-17       4          FALSE   -0.15164286
#> 33:        growth      VOC 2021-06-12       1          FALSE   -0.05397857
#> 34:        growth      VOC 2021-06-19       2          FALSE   -0.05830000
#> 35:        growth      VOC 2021-06-26       3          FALSE   -0.07142143
#> 36:        growth      VOC 2021-07-03       4          FALSE   -0.07377857
#> 37:        growth      VOC 2021-07-10       5          FALSE   -0.08250000
#> 38:        growth      VOC 2021-07-17       6          FALSE   -0.08721429
#> 39:        growth  non-VOC 2021-06-12       1          FALSE   -0.51621429
#> 40:        growth  non-VOC 2021-06-19       2          FALSE   -0.52092857
#> 41:        growth  non-VOC 2021-06-26       3          FALSE   -0.53428571
#> 42:        growth  non-VOC 2021-07-03       4          FALSE   -0.53664286
#> 43:        growth  non-VOC 2021-07-10       5          FALSE   -0.54528571
#> 44:        growth  non-VOC 2021-07-17       6          FALSE   -0.54921429
#> 45:            rt Combined 2021-06-26       1          FALSE    0.74596886
#> 46:            rt Combined 2021-07-03       2          FALSE    0.79373947
#> 47:            rt Combined 2021-07-10       3          FALSE    0.82944374
#> 48:            rt Combined 2021-07-17       4          FALSE    0.85929512
#> 49:            rt      VOC 2021-06-12       1          FALSE    0.94745241
#> 50:            rt      VOC 2021-06-19       2          FALSE    0.94336689
#> 51:            rt      VOC 2021-06-26       3          FALSE    0.93106943
#> 52:            rt      VOC 2021-07-03       4          FALSE    0.92887735
#> 53:            rt      VOC 2021-07-10       5          FALSE    0.92081144
#> 54:            rt      VOC 2021-07-17       6          FALSE    0.91648069
#> 55:            rt  non-VOC 2021-06-12       1          FALSE    0.59677550
#> 56:            rt  non-VOC 2021-06-19       2          FALSE    0.59396875
#> 57:            rt  non-VOC 2021-06-26       3          FALSE    0.58608777
#> 58:            rt  non-VOC 2021-07-03       4          FALSE    0.58470791
#> 59:            rt  non-VOC 2021-07-10       5          FALSE    0.57967614
#> 60:            rt  non-VOC 2021-07-17       6          FALSE    0.57740331
#>        value_type     type       date horizon forecast_start          mean
#>            median       sd      mad           q5          q20           q80
#>  1: 3340.00000000 9.62e+02 8.30e+02 2200.0000000 2710.0000000  4.130000e+03
#>  2: 2390.00000000 1.44e+03 1.00e+03 1150.0000000 1650.0000000  3.420000e+03
#>  3: 1840.00000000 2.11e+03 1.18e+03  591.0000000 1040.0000000  3.240000e+03
#>  4: 1490.00000000 4.39e+03 1.26e+03  285.0000000  676.0000000  3.140000e+03
#>  5: 2030.00000000 4.42e+02 4.12e+02 1420.0000000 1710.0000000  2.420000e+03
#>  6: 1880.00000000 4.60e+02 3.88e+02 1270.0000000 1560.0000000  2.230000e+03
#>  7: 1710.00000000 5.85e+02 4.91e+02 1040.0000000 1320.0000000  2.190000e+03
#>  8: 1570.00000000 9.77e+02 6.81e+02  723.0000000 1080.0000000  2.260000e+03
#>  9: 1420.00000000 1.65e+03 9.27e+02  447.0000000  786.0000000  2.490000e+03
#> 10: 1280.00000000 3.73e+03 1.07e+03  251.0000000  571.0000000  2.700000e+03
#> 11: 6120.00000000 1.03e+03 8.92e+02 4690.0000000 5380.0000000  6.940000e+03
#> 12: 3120.00000000 6.48e+02 5.66e+02 2280.0000000 2680.0000000  3.640000e+03
#> 13: 1590.00000000 5.25e+02 4.49e+02  964.0000000 1260.0000000  2.050000e+03
#> 14:  808.00000000 5.38e+02 3.68e+02  354.0000000  542.0000000  1.180000e+03
#> 15:  402.00000000 5.15e+02 2.65e+02  118.0000000  226.0000000  7.250000e+02
#> 16:  204.00000000 7.10e+02 1.76e+02   37.0000000   91.0000000  4.440000e+02
#> 17:    0.24800000 3.88e-02 3.51e-02    0.1910000    0.2200000  2.800000e-01
#> 18:    0.37300000 5.40e-02 4.85e-02    0.2890000    0.3320000  4.170000e-01
#> 19:    0.51700000 6.37e-02 5.74e-02    0.4130000    0.4690000  5.670000e-01
#> 20:    0.65800000 6.35e-02 5.73e-02    0.5470000    0.6090000  7.080000e-01
#> 21:    0.77500000 5.47e-02 4.89e-02    0.6760000    0.7330000  8.170000e-01
#> 22:    0.86100000 4.19e-02 3.67e-02    0.7830000    0.8290000  8.920000e-01
#> 23:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 24:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 25:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 26:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 27:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 28:    1.58724530 2.87e-02 2.73e-02    1.5333097    1.5600444  1.616190e+00
#> 29:   -0.29464286 1.76e-01 1.59e-01   -0.5115000   -0.3991429 -1.854286e-01
#> 30:   -0.23021429 2.34e-01 2.12e-01   -0.5272143   -0.3755714 -9.428571e-02
#> 31:   -0.18464286 2.95e-01 2.62e-01   -0.5531429   -0.3645714 -1.288571e-02
#> 32:   -0.14614286 3.45e-01 3.02e-01   -0.5767143   -0.3582857  5.924286e-02
#> 33:   -0.05280000 8.11e-02 7.15e-02   -0.1626429   -0.1013571 -5.570714e-03
#> 34:   -0.06097143 9.81e-02 9.05e-02   -0.1760000   -0.1178571  2.695000e-04
#> 35:   -0.07322857 1.77e-01 1.59e-01   -0.2915000   -0.1791429  3.465000e-02
#> 36:   -0.07315000 2.35e-01 2.11e-01   -0.3677143   -0.2145000  6.655000e-02
#> 37:   -0.08171429 2.95e-01 2.63e-01   -0.4462857   -0.2585000  9.114286e-02
#> 38:   -0.08328571 3.45e-01 3.03e-01   -0.5209286   -0.2930714  1.217857e-01
#> 39:   -0.51542857 8.32e-02 7.35e-02   -0.6293571   -0.5641429 -4.635714e-01
#> 40:   -0.52092857 1.03e-01 9.48e-02   -0.6490000   -0.5853571 -4.580714e-01
#> 41:   -0.53428571 1.81e-01 1.62e-01   -0.7637143   -0.6458571 -4.266429e-01
#> 42:   -0.53821429 2.38e-01 2.16e-01   -0.8407143   -0.6772857 -3.920714e-01
#> 43:   -0.54292857 2.98e-01 2.67e-01   -0.9192857   -0.7228571 -3.700714e-01
#> 44:   -0.54292857 3.47e-01 3.08e-01   -0.9900000   -0.7542857 -3.465000e-01
#> 45:    0.74479754 1.76e-01 1.59e-01    0.5995955    0.6708949  8.307482e-01
#> 46:    0.79436336 2.34e-01 2.12e-01    0.5902469    0.6868967  9.100227e-01
#> 47:    0.83140116 2.95e-01 2.62e-01    0.5751394    0.6944942  9.871970e-01
#> 48:    0.86403426 3.45e-01 3.02e-01    0.5617411    0.6988734  1.061033e+00
#> 49:    0.94856971 8.11e-02 7.15e-02    0.8498947    0.9036103  9.944448e-01
#> 50:    0.94085012 9.81e-02 9.05e-02    0.8386180    0.8888230  1.000270e+00
#> 51:    0.92938837 1.77e-01 1.59e-01    0.7471420    0.8359865  1.035257e+00
#> 52:    0.92946140 2.35e-01 2.11e-01    0.6923150    0.8069448  1.068814e+00
#> 53:    0.92153522 2.95e-01 2.63e-01    0.6400009    0.7722090  1.095425e+00
#> 54:    0.92008823 3.45e-01 3.03e-01    0.5939687    0.7459689  1.129512e+00
#> 55:    0.59724458 8.32e-02 7.35e-02    0.5329343    0.5688475  6.290331e-01
#> 56:    0.59396875 1.03e-01 9.48e-02    0.5225681    0.5569069  6.325023e-01
#> 57:    0.58608777 1.81e-01 1.62e-01    0.4659326    0.5242130  6.526966e-01
#> 58:    0.58378980 2.38e-01 2.16e-01    0.4314023    0.5079940  6.756559e-01
#> 59:    0.58104413 2.98e-01 2.67e-01    0.3988038    0.4853635  6.906850e-01
#> 60:    0.58104413 3.47e-01 3.08e-01    0.3715767    0.4703465  7.071588e-01
#>            median       sd      mad           q5          q20           q80
#>               q95
#>  1: 5160.00000000
#>  2: 5000.00000000
#>  3: 5780.00000000
#>  4: 7810.00000000
#>  5: 2860.00000000
#>  6: 2670.00000000
#>  7: 2850.00000000
#>  8: 3290.00000000
#>  9: 4440.00000000
#> 10: 6610.00000000
#> 11: 7910.00000000
#> 12: 4340.00000000
#> 13: 2620.00000000
#> 14: 1780.00000000
#> 15: 1340.00000000
#> 16: 1090.00000000
#> 17:    0.31800000
#> 18:    0.46800000
#> 19:    0.62400000
#> 20:    0.75800000
#> 21:    0.85500000
#> 22:    0.91700000
#> 23:    1.64825027
#> 24:    1.64825027
#> 25:    1.64825027
#> 26:    1.64825027
#> 27:    1.64825027
#> 28:    1.64825027
#> 29:   -0.06435000
#> 30:    0.06309286
#> 31:    0.19407143
#> 32:    0.30014286
#> 33:    0.04541429
#> 34:    0.06906429
#> 35:    0.16028571
#> 36:    0.23335714
#> 37:    0.30171429
#> 38:    0.36535714
#> 39:   -0.41800000
#> 40:   -0.38657143
#> 41:   -0.29228571
#> 42:   -0.23650000
#> 43:   -0.15714286
#> 44:   -0.09821429
#> 45:    0.93767676
#> 46:    1.06512574
#> 47:    1.21418301
#> 48:    1.35005166
#> 49:    1.04646130
#> 50:    1.07150509
#> 51:    1.17384621
#> 52:    1.26283241
#> 53:    1.35217484
#> 54:    1.44102857
#> 55:    0.65836223
#> 56:    0.67938220
#> 57:    0.74655520
#> 58:    0.78938588
#> 59:    0.85458197
#> 60:    0.90645464
#>               q95