Summary method for fv_tidy_posterior

summary method for class "fv_tidy_posterior". Can be used to filter the posterior for variables of interest, to return forecats only, and to summarise using the data.table method

Usage

# S3 method for fv_posterior
summary(object, type = "model", forecast = FALSE, as_dt = FALSE, ...)

Arguments

object

An object of the class fv_posterior as returned by fv_tidy_posterior() .

type

A character string used to filter the summarised output and defaulting to "model". Current options are: "model" which returns a summary of key model parameters, "cases" which returns summarised cases, "voc_frac" which returns summarised estimates of the fraction of cases that have the variant of concern, "voc_advantage" that returns summarised estiamtes of the the transmission advantage of the variant of concern, "growth" which returns summarised variant specifc and overall growth rates, "rt" which returns summarised variant specifc and overall reproduction number estimates, "raw" which returns a raw posterior summary, and "all" which returns all tidied posterior estimates.

forecast

Logical defaults to FALSE. Should fv_extract_forecast() be used to return only forecasts rather than complete posterior.

as_dt

Logical defaults to FALSE. Once any filtering has been applied should summary() fall back to using the default data.table method.

...

Additional arguments passed to summary() when as_dt = TRUE.

Value

A summary data.table table unless type "all" is used in which case the output is still of type "fv_posterior"

See also

Examples

posterior <- fv_example(strains = 2, type = "posterior")

# case summary
summary(posterior, type = "cases")
#>               variable       date     type    obs observed forecast_start
#>  1:       sim_cases[1] 2021-03-20 Combined  87328     TRUE          FALSE
#>  2:       sim_cases[2] 2021-03-27 Combined 109442     TRUE          FALSE
#>  3:       sim_cases[3] 2021-04-03 Combined 117965     TRUE          FALSE
#>  4:       sim_cases[4] 2021-04-10 Combined 107223     TRUE          FALSE
#>  5:       sim_cases[5] 2021-04-17 Combined 142664     TRUE          FALSE
#>  6:       sim_cases[6] 2021-04-24 Combined 145568     TRUE          FALSE
#>  7:       sim_cases[7] 2021-05-01 Combined 131887     TRUE          FALSE
#>  8:       sim_cases[8] 2021-05-08 Combined 107141     TRUE          FALSE
#>  9:       sim_cases[9] 2021-05-15 Combined  77261     TRUE          FALSE
#> 10:      sim_cases[10] 2021-05-22 Combined  57310     TRUE          FALSE
#> 11:      sim_cases[11] 2021-05-29 Combined  33052     TRUE          FALSE
#> 12:      sim_cases[12] 2021-06-05 Combined  22631     TRUE          FALSE
#> 13:      sim_cases[13] 2021-06-12 Combined  15553     TRUE          FALSE
#> 14:      sim_cases[14] 2021-06-19 Combined   7659     TRUE          FALSE
#> 15:      sim_cases[15] 2021-06-26 Combined   5033     TRUE           TRUE
#> 16:      sim_cases[16] 2021-07-03 Combined     NA    FALSE          FALSE
#> 17:      sim_cases[17] 2021-07-10 Combined     NA    FALSE          FALSE
#> 18:      sim_cases[18] 2021-07-17 Combined     NA    FALSE          FALSE
#> 19:      sim_cases[19] 2021-07-24 Combined     NA    FALSE          FALSE
#> 20:   sim_voc_cases[1] 2021-04-17      VOC     NA     TRUE          FALSE
#> 21:   sim_voc_cases[2] 2021-04-24      VOC     NA     TRUE          FALSE
#> 22:   sim_voc_cases[3] 2021-05-01      VOC     NA     TRUE          FALSE
#> 23:   sim_voc_cases[4] 2021-05-08      VOC     NA     TRUE          FALSE
#> 24:   sim_voc_cases[5] 2021-05-15      VOC     NA     TRUE          FALSE
#> 25:   sim_voc_cases[6] 2021-05-22      VOC     NA     TRUE          FALSE
#> 26:   sim_voc_cases[7] 2021-05-29      VOC     NA     TRUE          FALSE
#> 27:   sim_voc_cases[8] 2021-06-05      VOC     NA     TRUE          FALSE
#> 28:   sim_voc_cases[9] 2021-06-12      VOC     NA     TRUE           TRUE
#> 29:  sim_voc_cases[10] 2021-06-19      VOC     NA    FALSE          FALSE
#> 30:  sim_voc_cases[11] 2021-06-26      VOC     NA    FALSE          FALSE
#> 31:  sim_voc_cases[12] 2021-07-03      VOC     NA    FALSE          FALSE
#> 32:  sim_voc_cases[13] 2021-07-10      VOC     NA    FALSE          FALSE
#> 33:  sim_voc_cases[14] 2021-07-17      VOC     NA    FALSE          FALSE
#> 34:  sim_voc_cases[15] 2021-07-24      VOC     NA    FALSE          FALSE
#> 35:  sim_nvoc_cases[1] 2021-03-20  non-VOC     NA     TRUE          FALSE
#> 36:  sim_nvoc_cases[2] 2021-03-27  non-VOC     NA     TRUE          FALSE
#> 37:  sim_nvoc_cases[3] 2021-04-03  non-VOC     NA     TRUE          FALSE
#> 38:  sim_nvoc_cases[4] 2021-04-10  non-VOC     NA     TRUE          FALSE
#> 39:  sim_nvoc_cases[5] 2021-04-17  non-VOC     NA     TRUE          FALSE
#> 40:  sim_nvoc_cases[6] 2021-04-24  non-VOC     NA     TRUE          FALSE
#> 41:  sim_nvoc_cases[7] 2021-05-01  non-VOC     NA     TRUE          FALSE
#> 42:  sim_nvoc_cases[8] 2021-05-08  non-VOC     NA     TRUE          FALSE
#> 43:  sim_nvoc_cases[9] 2021-05-15  non-VOC     NA     TRUE          FALSE
#> 44: sim_nvoc_cases[10] 2021-05-22  non-VOC     NA     TRUE          FALSE
#> 45: sim_nvoc_cases[11] 2021-05-29  non-VOC     NA     TRUE          FALSE
#> 46: sim_nvoc_cases[12] 2021-06-05  non-VOC     NA     TRUE          FALSE
#> 47: sim_nvoc_cases[13] 2021-06-12  non-VOC     NA     TRUE           TRUE
#> 48: sim_nvoc_cases[14] 2021-06-19  non-VOC     NA    FALSE          FALSE
#> 49: sim_nvoc_cases[15] 2021-06-26  non-VOC     NA    FALSE          FALSE
#> 50: sim_nvoc_cases[16] 2021-07-03  non-VOC     NA    FALSE          FALSE
#> 51: sim_nvoc_cases[17] 2021-07-10  non-VOC     NA    FALSE          FALSE
#> 52: sim_nvoc_cases[18] 2021-07-17  non-VOC     NA    FALSE          FALSE
#> 53: sim_nvoc_cases[19] 2021-07-24  non-VOC     NA    FALSE          FALSE
#>               variable       date     type    obs observed forecast_start
#>       mean median      sd     mad     q5    q20    q80    q95 rhat ess_bulk
#>  1:  90100  89500 13400.0 12100.0  69200  79700 100000 112000    1     2060
#>  2: 102000 102000 14700.0 13100.0  78800  91000 113000 126000    1     2130
#>  3: 114000 113000 17600.0 15200.0  86500 100000 126000 143000    1     1980
#>  4: 123000 122000 19100.0 16700.0  93100 109000 137000 154000    1     1820
#>  5: 133000 133000 19900.0 17400.0 104000 118000 148000 164000    1     1920
#>  6: 138000 137000 20400.0 18200.0 106000 122000 153000 172000    1     1910
#>  7: 128000 127000 19100.0 16700.0  98800 113000 142000 160000    1     1860
#>  8: 106000 105000 16300.0 13700.0  82400  94800 117000 132000    1     2090
#>  9:  79600  78800 11800.0 10600.0  62700  70300  88100  99600    1     1880
#> 10:  55700  55200  8510.0  6970.0  43400  49400  61500  69900    1     1890
#> 11:  35800  35500  5540.0  4500.0  27800  31900  39600  45200    1     1870
#> 12:  22900  22700  3230.0  2800.0  17900  20400  25200  28100    1     2200
#> 13:  14200  14000  1910.0  1690.0  11400  12600  15600  17300    1     2160
#> 14:   8310   8240  1130.0   936.0   6700   7440   9080  10200    1     2340
#> 15:   5140   5060   822.0   692.0   4050   4520   5680   6470    1     2140
#> 16:   3400   3300   900.0   789.0   2170   2690   4020   5010    1     2050
#> 17:   2510   2300  1160.0   905.0   1120   1630   3210   4580    1     1780
#> 18:   2130   1710  2140.0  1030.0    551    999   2880   4920    1     1810
#> 19:   2160   1340  6260.0  1110.0    256    611   2850   5870    1     1710
#> 20:    221    218    40.0    37.1    162    189    252    288    1     2010
#> 21:    410    408    71.1    68.2    301    351    469    528    1     1800
#> 22:    684    673   124.0   108.0    507    588    772    900    1     1970
#> 23:   1030   1010   194.0   175.0    742    869   1170   1360    1     2290
#> 24:   1380   1350   269.0   234.0    991   1170   1580   1830    1     2270
#> 25:   1720   1680   330.0   307.0   1230   1450   1960   2290    1     2280
#> 26:   1960   1920   413.0   354.0   1390   1650   2250   2670    1     2300
#> 27:   2170   2140   459.0   411.0   1500   1810   2500   2980    1     2250
#> 28:   2230   2200   490.0   442.0   1510   1830   2580   3060    1     2130
#> 29:   2110   2080   490.0   451.0   1420   1710   2460   2950    1     2270
#> 30:   1950   1910   464.0   422.0   1300   1570   2290   2770    1     2250
#> 31:   1780   1700   575.0   497.0   1040   1320   2170   2810    1     2210
#> 32:   1670   1520   844.0   630.0    703   1050   2160   3090    1     2010
#> 33:   1660   1320  1760.0   824.0    417    744   2250   3840    1     1850
#> 34:   1860   1150  5590.0   959.0    222    522   2450   5170    1     1760
#> 35:  90100  89500 13400.0 12100.0  69200  79700 100000 112000    1     2060
#> 36: 102000 102000 14700.0 13100.0  78800  91000 113000 126000    1     2130
#> 37: 114000 113000 17600.0 15200.0  86500 100000 126000 143000    1     1980
#> 38: 123000 122000 19100.0 16700.0  93100 109000 137000 154000    1     1820
#> 39: 133000 133000 19900.0 17400.0 103000 118000 147000 164000    1     1920
#> 40: 138000 137000 20400.0 18100.0 105000 122000 153000 172000    1     1910
#> 41: 127000 127000 19100.0 16600.0  98200 113000 141000 159000    1     1860
#> 42: 105000 105000 16200.0 13600.0  81500  93700 116000 131000    1     2090
#> 43:  78300  77400 11700.0 10500.0  61300  69000  86700  97800    1     1880
#> 44:  54000  53500  8450.0  6980.0  41700  47700  59700  67900    1     1890
#> 45:  33900  33500  5430.0  4340.0  26000  30100  37600  43100    1     1870
#> 46:  20700  20600  3130.0  2630.0  15900  18300  23000  25800    1     2180
#> 47:  11900  11800  1830.0  1630.0   9290  10500  13300  15100    1     2200
#> 48:   6200   6130  1040.0   901.0   4680   5400   6940   7860    1     2220
#> 49:   3190   3130   664.0   544.0   2290   2700   3620   4290    1     1950
#> 50:   1620   1550   498.0   423.0    958   1240   1960   2490    1     1950
#> 51:    837    753   415.0   322.0    351    517   1090   1580    1     1850
#> 52:    467    368   432.0   232.0    113    212    641   1080    1     1720
#> 53:    296    177   709.0   151.0     32     79    381    827    1     1510
#>       mean median      sd     mad     q5    q20    q80    q95 rhat ess_bulk
#>     ess_tail
#>  1:     1860
#>  2:     1850
#>  3:     1700
#>  4:     1760
#>  5:     1350
#>  6:     1860
#>  7:     1780
#>  8:     1820
#>  9:     1750
#> 10:     1650
#> 11:     1700
#> 12:     1850
#> 13:     1780
#> 14:     1850
#> 15:     1750
#> 16:     1580
#> 17:     1620
#> 18:     1460
#> 19:     1720
#> 20:     1850
#> 21:     1750
#> 22:     1770
#> 23:     1850
#> 24:     1820
#> 25:     1550
#> 26:     1650
#> 27:     1720
#> 28:     1660
#> 29:     1660
#> 30:     1630
#> 31:     1450
#> 32:     1640
#> 33:     1610
#> 34:     1720
#> 35:     1860
#> 36:     1850
#> 37:     1700
#> 38:     1760
#> 39:     1350
#> 40:     1790
#> 41:     1780
#> 42:     1820
#> 43:     1780
#> 44:     1670
#> 45:     1670
#> 46:     1880
#> 47:     1850
#> 48:     1700
#> 49:     1420
#> 50:     1420
#> 51:     1720
#> 52:     1500
#> 53:     1490
#>     ess_tail

# summary of the case summary
summary(posterior, type = "cases", as_dt = TRUE)
#>    variable              date                type                obs        
#>  Length:53          Min.   :2021-03-20   Length:53          Min.   :  5033  
#>  Class :character   1st Qu.:2021-04-24   Class :character   1st Qu.: 27842  
#>  Mode  :character   Median :2021-05-29   Mode  :character   Median : 87328  
#>                     Mean   :2021-05-25                      Mean   : 77848  
#>                     3rd Qu.:2021-06-26                      3rd Qu.:113704  
#>                     Max.   :2021-07-24                      Max.   :145568  
#>                                                             NA's   :38      
#>   observed       forecast_start       mean            median      
#>  Mode :logical   Mode :logical   Min.   :   221   Min.   :   177  
#>  FALSE:16        FALSE:50        1st Qu.:  1860   1st Qu.:  1680  
#>  TRUE :37        TRUE :3         Median :  8310   Median :  8240  
#>                                  Mean   : 44006   Mean   : 43686  
#>                                  3rd Qu.:102000   3rd Qu.:102000  
#>                                  Max.   :138000   Max.   :137000  
#>                                                                   
#>        sd             mad                q5              q20        
#>  Min.   :   40   Min.   :   37.1   Min.   :    32   Min.   :    79  
#>  1st Qu.:  498   1st Qu.:  442.0   1st Qu.:   991   1st Qu.:  1240  
#>  Median : 2140   Median : 1110.0   Median :  6700   Median :  7440  
#>  Mean   : 6995   Mean   : 5896.1   Mean   : 33740   Mean   : 38848  
#>  3rd Qu.:14700   3rd Qu.:13100.0   3rd Qu.: 78800   3rd Qu.: 91000  
#>  Max.   :20400   Max.   :18200.0   Max.   :106000   Max.   :122000  
#>                                                                     
#>       q80              q95              rhat      ess_bulk       ess_tail   
#>  Min.   :   252   Min.   :   288   Min.   :1   Min.   :1510   Min.   :1350  
#>  1st Qu.:  2250   1st Qu.:  2950   1st Qu.:1   1st Qu.:1870   1st Qu.:1640  
#>  Median :  9080   Median : 10200   Median :1   Median :1980   Median :1720  
#>  Mean   : 48848   Mean   : 55198   Mean   :1   Mean   :2005   Mean   :1699  
#>  3rd Qu.:113000   3rd Qu.:126000   3rd Qu.:1   3rd Qu.:2160   3rd Qu.:1820  
#>  Max.   :153000   Max.   :172000   Max.   :1   Max.   :2340   Max.   :1880  
#>                                                                             

# case forecast only
summary(posterior, type = "cases", forecast = TRUE)
#>         type       date horizon forecast_start mean median   sd  mad   q5  q20
#>  1: Combined 2021-07-03       1          FALSE 3400   3300  900  789 2170 2690
#>  2: Combined 2021-07-10       2          FALSE 2510   2300 1160  905 1120 1630
#>  3: Combined 2021-07-17       3          FALSE 2130   1710 2140 1030  551  999
#>  4: Combined 2021-07-24       4          FALSE 2160   1340 6260 1110  256  611
#>  5:      VOC 2021-06-19       1          FALSE 2110   2080  490  451 1420 1710
#>  6:      VOC 2021-06-26       2          FALSE 1950   1910  464  422 1300 1570
#>  7:      VOC 2021-07-03       3          FALSE 1780   1700  575  497 1040 1320
#>  8:      VOC 2021-07-10       4          FALSE 1670   1520  844  630  703 1050
#>  9:      VOC 2021-07-17       5          FALSE 1660   1320 1760  824  417  744
#> 10:      VOC 2021-07-24       6          FALSE 1860   1150 5590  959  222  522
#> 11:  non-VOC 2021-06-19       1          FALSE 6200   6130 1040  901 4680 5400
#> 12:  non-VOC 2021-06-26       2          FALSE 3190   3130  664  544 2290 2700
#> 13:  non-VOC 2021-07-03       3          FALSE 1620   1550  498  423  958 1240
#> 14:  non-VOC 2021-07-10       4          FALSE  837    753  415  322  351  517
#> 15:  non-VOC 2021-07-17       5          FALSE  467    368  432  232  113  212
#> 16:  non-VOC 2021-07-24       6          FALSE  296    177  709  151   32   79
#>      q80  q95
#>  1: 4020 5010
#>  2: 3210 4580
#>  3: 2880 4920
#>  4: 2850 5870
#>  5: 2460 2950
#>  6: 2290 2770
#>  7: 2170 2810
#>  8: 2160 3090
#>  9: 2250 3840
#> 10: 2450 5170
#> 11: 6940 7860
#> 12: 3620 4290
#> 13: 1960 2490
#> 14: 1090 1580
#> 15:  641 1080
#> 16:  381  827

# voc fraction summary
summary(posterior, type = "voc_frac")
#>         variable       date type         obs observed forecast_start    mean
#>  1:  frac_voc[1] 2021-04-17  VOC 0.001465917     TRUE          FALSE 0.00166
#>  2:  frac_voc[2] 2021-04-24  VOC 0.007085917     TRUE          FALSE 0.00298
#>  3:  frac_voc[3] 2021-05-01  VOC 0.015085025     TRUE          FALSE 0.00537
#>  4:  frac_voc[4] 2021-05-08  VOC 0.019443217     TRUE          FALSE 0.00965
#>  5:  frac_voc[5] 2021-05-15  VOC 0.026459592     TRUE          FALSE 0.01730
#>  6:  frac_voc[6] 2021-05-22  VOC 0.032900194     TRUE          FALSE 0.03090
#>  7:  frac_voc[7] 2021-05-29  VOC 0.038084020     TRUE          FALSE 0.05450
#>  8:  frac_voc[8] 2021-06-05  VOC 0.085600907     TRUE          FALSE 0.09440
#>  9:  frac_voc[9] 2021-06-12  VOC 0.166666667     TRUE           TRUE 0.15900
#> 10: frac_voc[10] 2021-06-19  VOC          NA    FALSE          FALSE 0.25400
#> 11: frac_voc[11] 2021-06-26  VOC          NA    FALSE          FALSE 0.37900
#> 12: frac_voc[12] 2021-07-03  VOC          NA    FALSE          FALSE 0.52300
#> 13: frac_voc[13] 2021-07-10  VOC          NA    FALSE          FALSE 0.66200
#> 14: frac_voc[14] 2021-07-17  VOC          NA    FALSE          FALSE 0.77700
#> 15: frac_voc[15] 2021-07-24  VOC          NA    FALSE          FALSE 0.86100
#>      median       sd      mad      q5     q20     q80     q95 rhat ess_bulk
#>  1: 0.00164 0.000201 0.000189 0.00136 0.00149 0.00181 0.00200    1     1790
#>  2: 0.00296 0.000318 0.000309 0.00250 0.00272 0.00324 0.00352    1     1940
#>  3: 0.00534 0.000529 0.000507 0.00456 0.00492 0.00580 0.00629    1     2170
#>  4: 0.00962 0.000947 0.000916 0.00817 0.00885 0.01040 0.01130    1     2420
#>  5: 0.01720 0.001820 0.001750 0.01450 0.01580 0.01880 0.02040    1     2610
#>  6: 0.03080 0.003640 0.003500 0.02520 0.02790 0.03370 0.03690    1     2600
#>  7: 0.05410 0.007260 0.006880 0.04350 0.04870 0.05990 0.06670    1     2530
#>  8: 0.09360 0.013900 0.012900 0.07340 0.08350 0.10500 0.11800    1     2440
#>  9: 0.15700 0.024800 0.023000 0.12100 0.13900 0.17800 0.19900    1     2350
#> 10: 0.25200 0.039500 0.037100 0.19300 0.22200 0.28500 0.31900    1     2270
#> 11: 0.37800 0.054400 0.052200 0.29300 0.33500 0.42300 0.46800    1     2200
#> 12: 0.52400 0.063400 0.061400 0.41900 0.47100 0.57400 0.62500    1     2150
#> 13: 0.66500 0.062500 0.060300 0.55600 0.61100 0.71400 0.75900    1     2110
#> 14: 0.78200 0.053100 0.050000 0.68300 0.73600 0.82100 0.85600    1     2080
#> 15: 0.86600 0.040200 0.037000 0.78900 0.83100 0.89500 0.91800    1     2060
#>     ess_tail
#>  1:     1480
#>  2:     1580
#>  3:     1700
#>  4:     1590
#>  5:     1650
#>  6:     1590
#>  7:     1670
#>  8:     1440
#>  9:     1340
#> 10:     1270
#> 11:     1260
#> 12:     1260
#> 13:     1220
#> 14:     1220
#> 15:     1310

# voc advantage summary
summary(posterior, type = "voc_advantage")
#>              variable       date type observed forecast_start     mean   median
#>  1:  voc_advantage[1] 2021-04-17  VOC     TRUE          FALSE 1.590991 1.590991
#>  2:  voc_advantage[2] 2021-04-24  VOC     TRUE          FALSE 1.590991 1.590991
#>  3:  voc_advantage[3] 2021-05-01  VOC     TRUE          FALSE 1.590991 1.590991
#>  4:  voc_advantage[4] 2021-05-08  VOC     TRUE          FALSE 1.590991 1.590991
#>  5:  voc_advantage[5] 2021-05-15  VOC     TRUE          FALSE 1.590991 1.590991
#>  6:  voc_advantage[6] 2021-05-22  VOC     TRUE          FALSE 1.590991 1.590991
#>  7:  voc_advantage[7] 2021-05-29  VOC     TRUE          FALSE 1.590991 1.590991
#>  8:  voc_advantage[8] 2021-06-05  VOC     TRUE          FALSE 1.590991 1.590991
#>  9:  voc_advantage[9] 2021-06-12  VOC     TRUE           TRUE 1.590991 1.590991
#> 10: voc_advantage[10] 2021-06-19  VOC    FALSE          FALSE 1.590991 1.590991
#> 11: voc_advantage[11] 2021-06-26  VOC    FALSE          FALSE 1.590991 1.590991
#> 12: voc_advantage[12] 2021-07-03  VOC    FALSE          FALSE 1.590991 1.590991
#> 13: voc_advantage[13] 2021-07-10  VOC    FALSE          FALSE 1.590991 1.590991
#> 14: voc_advantage[14] 2021-07-17  VOC    FALSE          FALSE 1.590991 1.590991
#> 15: avg_voc_advantage 2021-07-24  VOC    FALSE          FALSE 1.590991 1.590991
#>         sd    mad       q5      q20      q80      q95 rhat ess_bulk ess_tail
#>  1: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  2: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  3: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  4: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  5: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  6: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  7: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  8: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#>  9: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 10: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 11: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 12: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 13: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 14: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410
#> 15: 0.0283 0.0265 1.534515 1.562498 1.620004 1.649546    1     1880     1410

# growth summary
summary(posterior, type = "growth")
#>      variable       date     type observed forecast_start        mean
#>  1:  com_r[1] 2021-03-20 Combined     TRUE          FALSE  0.10292857
#>  2:  com_r[2] 2021-03-27 Combined     TRUE          FALSE  0.08250000
#>  3:  com_r[3] 2021-04-03 Combined     TRUE          FALSE  0.05830000
#>  4:  com_r[4] 2021-04-10 Combined     TRUE          FALSE  0.06600000
#>  5:  com_r[5] 2021-04-17 Combined     TRUE          FALSE  0.02325714
#>  6:  com_r[6] 2021-04-24 Combined     TRUE          FALSE -0.05672857
#>  7:  com_r[7] 2021-05-01 Combined     TRUE          FALSE -0.14457143
#>  8:  com_r[8] 2021-05-08 Combined     TRUE          FALSE -0.22392857
#>  9:  com_r[9] 2021-05-15 Combined     TRUE          FALSE -0.28050000
#> 10: com_r[10] 2021-05-22 Combined     TRUE          FALSE -0.33314286
#> 11: com_r[11] 2021-05-29 Combined     TRUE          FALSE -0.34571429
#> 12: com_r[12] 2021-06-05 Combined     TRUE          FALSE -0.36535714
#> 13: com_r[13] 2021-06-12 Combined     TRUE          FALSE -0.39364286
#> 14: com_r[14] 2021-06-19 Combined     TRUE           TRUE -0.35121429
#> 15: com_r[15] 2021-06-26 Combined    FALSE          FALSE -0.30800000
#> 16: com_r[16] 2021-07-03 Combined    FALSE          FALSE -0.25378571
#> 17: com_r[17] 2021-07-10 Combined    FALSE          FALSE -0.21292857
#> 18: com_r[18] 2021-07-17 Combined    FALSE          FALSE -0.18228571
#> 19:  voc_r[1] 2021-04-17      VOC     TRUE          FALSE  0.48635714
#> 20:  voc_r[2] 2021-04-24      VOC     TRUE          FALSE  0.40542857
#> 21:  voc_r[3] 2021-05-01      VOC     TRUE          FALSE  0.31507143
#> 22:  voc_r[4] 2021-05-08      VOC     TRUE          FALSE  0.23257143
#> 23:  voc_r[5] 2021-05-15      VOC     TRUE          FALSE  0.16971429
#> 24:  voc_r[6] 2021-05-22      VOC     TRUE          FALSE  0.10607143
#> 25:  voc_r[7] 2021-05-29      VOC     TRUE          FALSE  0.07472143
#> 26:  voc_r[8] 2021-06-05      VOC     TRUE           TRUE  0.02459286
#> 27:  voc_r[9] 2021-06-12      VOC    FALSE          FALSE -0.04816429
#> 28: voc_r[10] 2021-06-19      VOC    FALSE          FALSE -0.06442857
#> 29: voc_r[11] 2021-06-26      VOC    FALSE          FALSE -0.08800000
#> 30: voc_r[12] 2021-07-03      VOC    FALSE          FALSE -0.09742857
#> 31: voc_r[13] 2021-07-10      VOC    FALSE          FALSE -0.11078571
#> 32: voc_r[14] 2021-07-17      VOC    FALSE          FALSE -0.11864286
#> 33:      r[1] 2021-03-20  non-VOC     TRUE          FALSE  0.10292857
#> 34:      r[2] 2021-03-27  non-VOC     TRUE          FALSE  0.08250000
#> 35:      r[3] 2021-04-03  non-VOC     TRUE          FALSE  0.05830000
#> 36:      r[4] 2021-04-10  non-VOC     TRUE          FALSE  0.06600000
#> 37:      r[5] 2021-04-17  non-VOC     TRUE          FALSE  0.02184286
#> 38:      r[6] 2021-04-24  non-VOC     TRUE          FALSE -0.05916429
#> 39:      r[7] 2021-05-01  non-VOC     TRUE          FALSE -0.14928571
#> 40:      r[8] 2021-05-08  non-VOC     TRUE          FALSE -0.23178571
#> 41:      r[9] 2021-05-15  non-VOC     TRUE          FALSE -0.29464286
#> 42:     r[10] 2021-05-22  non-VOC     TRUE          FALSE -0.35828571
#> 43:     r[11] 2021-05-29  non-VOC     TRUE          FALSE -0.38971429
#> 44:     r[12] 2021-06-05  non-VOC     TRUE           TRUE -0.44000000
#> 45:     r[13] 2021-06-12  non-VOC    FALSE          FALSE -0.51228571
#> 46:     r[14] 2021-06-19  non-VOC    FALSE          FALSE -0.52878571
#> 47:     r[15] 2021-06-26  non-VOC    FALSE          FALSE -0.55235714
#> 48:     r[16] 2021-07-03  non-VOC    FALSE          FALSE -0.56178571
#> 49:     r[17] 2021-07-10  non-VOC    FALSE          FALSE -0.57514286
#> 50:     r[18] 2021-07-17  non-VOC    FALSE          FALSE -0.58300000
#>      variable       date     type observed forecast_start        mean
#>          median     sd    mad          q5         q20          q80         q95
#>  1:  0.10214286 0.0740 0.0652  0.01186429  0.05845714  0.146928571  0.19721429
#>  2:  0.08564286 0.0624 0.0524  0.00302500  0.04753571  0.120214286  0.15635714
#>  3:  0.06395714 0.0709 0.0535 -0.03912857  0.02255000  0.096642857  0.13514286
#>  4:  0.06254286 0.0665 0.0550 -0.01100000  0.02852143  0.102928571  0.15321429
#>  5:  0.02019286 0.0668 0.0541 -0.05476429 -0.01422143  0.059557143  0.10921429
#>  6: -0.05877143 0.0664 0.0501 -0.13278571 -0.09192857 -0.024592857  0.02404286
#>  7: -0.14300000 0.0665 0.0526 -0.22550000 -0.17992857 -0.108428571 -0.06411429
#>  8: -0.22157143 0.0639 0.0530 -0.30800000 -0.26007143 -0.188571429 -0.14614286
#>  9: -0.27971429 0.0668 0.0544 -0.36535714 -0.31664286 -0.243571429 -0.19721429
#> 10: -0.33078571 0.0666 0.0507 -0.41957143 -0.36535714 -0.297785714 -0.25221429
#> 11: -0.34885714 0.0645 0.0546 -0.42192857 -0.38185714 -0.308785714 -0.25928571
#> 12: -0.36850000 0.0654 0.0521 -0.44471429 -0.40071429 -0.330785714 -0.27892857
#> 13: -0.39207143 0.0766 0.0645 -0.49342857 -0.43842857 -0.349642857 -0.30328571
#> 14: -0.35200000 0.0957 0.0913 -0.46671429 -0.41092857 -0.291500000 -0.22785714
#> 15: -0.30878571 0.1640 0.1490 -0.51700000 -0.40542857 -0.210571429 -0.10292857
#> 16: -0.25378571 0.2230 0.1990 -0.53664286 -0.38657143 -0.121000000  0.02530000
#> 17: -0.20350000 0.2760 0.2540 -0.57435714 -0.38185714 -0.046121429  0.12414286
#> 18: -0.17521429 0.3350 0.2960 -0.61992857 -0.37478571  0.018307143  0.21371429
#> 19:  0.48400000 0.0681 0.0603  0.40385714  0.44707143  0.525642857  0.57750000
#> 20:  0.40228571 0.0685 0.0581  0.32528571  0.36692857  0.442357143  0.49264286
#> 21:  0.31742857 0.0711 0.0598  0.22785714  0.27500000  0.355142857  0.40071429
#> 22:  0.23414286 0.0685 0.0591  0.14300000  0.19407143  0.271857143  0.31350000
#> 23:  0.17050000 0.0716 0.0607  0.07841429  0.12964286  0.209785714  0.25614286
#> 24:  0.10921429 0.0710 0.0605  0.01343571  0.06623571  0.145357143  0.18935714
#> 25:  0.07338571 0.0673 0.0601 -0.00566500  0.03488571  0.114714286  0.16342857
#> 26:  0.02223571 0.0670 0.0543 -0.05649286 -0.01249286  0.061364286  0.11078571
#> 27: -0.04604286 0.0773 0.0653 -0.14928571 -0.09035714 -0.003912857  0.04140714
#> 28: -0.06702143 0.0950 0.0878 -0.18150000 -0.12414286 -0.006262143  0.05665000
#> 29: -0.08721429 0.1620 0.1470 -0.29621429 -0.18385714  0.009585714  0.11628571
#> 30: -0.09742857 0.2220 0.1980 -0.37950000 -0.22942857  0.035042857  0.17914286
#> 31: -0.10450000 0.2760 0.2550 -0.47614286 -0.27657143  0.056492857  0.21921429
#> 32: -0.11078571 0.3360 0.2970 -0.55550000 -0.31035714  0.082500000  0.27657143
#> 33:  0.10214286 0.0740 0.0652  0.01186429  0.05845714  0.146928571  0.19721429
#> 34:  0.08564286 0.0624 0.0524  0.00302500  0.04753571  0.120214286  0.15635714
#> 35:  0.06395714 0.0709 0.0535 -0.03912857  0.02255000  0.096642857  0.13514286
#> 36:  0.06254286 0.0665 0.0550 -0.01100000  0.02852143  0.102928571  0.15321429
#> 37:  0.01885714 0.0667 0.0541 -0.05610000 -0.01555714  0.058300000  0.10842857
#> 38: -0.06136429 0.0664 0.0498 -0.13514286 -0.09428571 -0.027264286  0.02176429
#> 39: -0.14771429 0.0664 0.0523 -0.23021429 -0.18464286 -0.113142857 -0.06827857
#> 40: -0.23021429 0.0639 0.0530 -0.31585714 -0.26792857 -0.196428571 -0.15478571
#> 41: -0.29464286 0.0669 0.0540 -0.37950000 -0.33078571 -0.257714286 -0.21057143
#> 42: -0.35671429 0.0668 0.0503 -0.44550000 -0.39207143 -0.322928571 -0.27892857
#> 43: -0.39207143 0.0652 0.0561 -0.46750000 -0.42742857 -0.352000000 -0.30250000
#> 44: -0.44157143 0.0669 0.0534 -0.51935714 -0.47692857 -0.403857143 -0.35042857
#> 45: -0.51071429 0.0787 0.0684 -0.61442857 -0.55864286 -0.465142857 -0.42114286
#> 46: -0.52957143 0.0989 0.0907 -0.65214286 -0.59007143 -0.467500000 -0.40071429
#> 47: -0.55314286 0.1650 0.1490 -0.76764286 -0.64821429 -0.451000000 -0.34335714
#> 48: -0.56178571 0.2260 0.2010 -0.84857143 -0.69692857 -0.424285714 -0.28600000
#> 49: -0.56807143 0.2790 0.2550 -0.94285714 -0.74328571 -0.407000000 -0.24042857
#> 50: -0.57278571 0.3390 0.2970 -1.02142857 -0.77707143 -0.381071429 -0.18228571
#>          median     sd    mad          q5         q20          q80         q95
#>     rhat ess_bulk ess_tail
#>  1:    1     1240      859
#>  2:    1     1970     1520
#>  3:    1     1100     1140
#>  4:    1     1390     1410
#>  5:    1     1590     1460
#>  6:    1     1830     1580
#>  7:    1     2050     1730
#>  8:    1     2100     1680
#>  9:    1     1780     1600
#> 10:    1     2050     1700
#> 11:    1     1770     1650
#> 12:    1     1890     1360
#> 13:    1     1650     1640
#> 14:    1     2690     1670
#> 15:    1     1880     1530
#> 16:    1     1630     1540
#> 17:    1     1810     1620
#> 18:    1     1650     1480
#> 19:    1     1680     1530
#> 20:    1     1950     1660
#> 21:    1     2060     1380
#> 22:    1     1900     1640
#> 23:    1     1600     1550
#> 24:    1     1950     1560
#> 25:    1     1870     1460
#> 26:    1     1880     1280
#> 27:    1     1620     1790
#> 28:    1     2690     1680
#> 29:    1     1730     1590
#> 30:    1     1520     1600
#> 31:    1     1790     1630
#> 32:    1     1650     1520
#> 33:    1     1240      859
#> 34:    1     1970     1520
#> 35:    1     1100     1140
#> 36:    1     1390     1410
#> 37:    1     1590     1460
#> 38:    1     1830     1580
#> 39:    1     2050     1730
#> 40:    1     2110     1680
#> 41:    1     1790     1550
#> 42:    1     2040     1730
#> 43:    1     1680     1800
#> 44:    1     1930     1260
#> 45:    1     1610     1560
#> 46:    1     2520     1650
#> 47:    1     1580     1370
#> 48:    1     1520     1300
#> 49:    1     1780     1630
#> 50:    1     1650     1570
#>     rhat ess_bulk ess_tail

# Rt summary
summary(posterior, type = "rt")
#>      variable       date     type observed forecast_start      mean    median
#>  1:  com_r[1] 2021-03-20 Combined     TRUE          FALSE 1.1084122 1.1075417
#>  2:  com_r[2] 2021-03-27 Combined     TRUE          FALSE 1.0859987 1.0894172
#>  3:  com_r[3] 2021-04-03 Combined     TRUE          FALSE 1.0600330 1.0660467
#>  4:  com_r[4] 2021-04-10 Combined     TRUE          FALSE 1.0682267 1.0645401
#>  5:  com_r[5] 2021-04-17 Combined     TRUE          FALSE 1.0235297 1.0203981
#>  6:  com_r[6] 2021-04-24 Combined     TRUE          FALSE 0.9448505 0.9429223
#>  7:  com_r[7] 2021-05-01 Combined     TRUE          FALSE 0.8653931 0.8667541
#>  8:  com_r[8] 2021-05-08 Combined     TRUE          FALSE 0.7993722 0.8012587
#>  9:  com_r[9] 2021-05-15 Combined     TRUE          FALSE 0.7554059 0.7559997
#> 10: com_r[10] 2021-05-22 Combined     TRUE          FALSE 0.7166678 0.7183591
#> 11: com_r[11] 2021-05-29 Combined     TRUE          FALSE 0.7077147 0.7054939
#> 12: com_r[12] 2021-06-05 Combined     TRUE          FALSE 0.6939488 0.6917712
#> 13: com_r[13] 2021-06-12 Combined     TRUE          FALSE 0.6745949 0.6756559
#> 14: com_r[14] 2021-06-19 Combined     TRUE           TRUE 0.7038329 0.7032801
#> 15: com_r[15] 2021-06-26 Combined    FALSE          FALSE 0.7349153 0.7343381
#> 16: com_r[16] 2021-07-03 Combined    FALSE          FALSE 0.7758580 0.7758580
#> 17: com_r[17] 2021-07-10 Combined    FALSE          FALSE 0.8082139 0.8158702
#> 18: com_r[18] 2021-07-17 Combined    FALSE          FALSE 0.8333632 0.8392772
#> 19:  voc_r[1] 2021-04-17      VOC     TRUE          FALSE 1.6263807 1.6225516
#> 20:  voc_r[2] 2021-04-24      VOC     TRUE          FALSE 1.4999452 1.4952385
#> 21:  voc_r[3] 2021-05-01      VOC     TRUE          FALSE 1.3703572 1.3735911
#> 22:  voc_r[4] 2021-05-08      VOC     TRUE          FALSE 1.2618406 1.2638250
#> 23:  voc_r[5] 2021-05-15      VOC     TRUE          FALSE 1.1849662 1.1858977
#> 24:  voc_r[6] 2021-05-22      VOC     TRUE          FALSE 1.1119013 1.1154013
#> 25:  voc_r[7] 2021-05-29      VOC     TRUE          FALSE 1.0775839 1.0761455
#> 26:  voc_r[8] 2021-06-05      VOC     TRUE           TRUE 1.0248978 1.0224848
#> 27:  voc_r[9] 2021-06-12      VOC    FALSE          FALSE 0.9529772 0.9550010
#> 28: voc_r[10] 2021-06-19      VOC    FALSE          FALSE 0.9376031 0.9351752
#> 29: voc_r[11] 2021-06-26      VOC    FALSE          FALSE 0.9157609 0.9164807
#> 30: voc_r[12] 2021-07-03      VOC    FALSE          FALSE 0.9071671 0.9071671
#> 31: voc_r[13] 2021-07-10      VOC    FALSE          FALSE 0.8951305 0.9007748
#> 32: voc_r[14] 2021-07-17      VOC    FALSE          FALSE 0.8881249 0.8951305
#> 33:      r[1] 2021-03-20  non-VOC     TRUE          FALSE 1.1084122 1.1075417
#> 34:      r[2] 2021-03-27  non-VOC     TRUE          FALSE 1.0859987 1.0894172
#> 35:      r[3] 2021-04-03  non-VOC     TRUE          FALSE 1.0600330 1.0660467
#> 36:      r[4] 2021-04-10  non-VOC     TRUE          FALSE 1.0682267 1.0645401
#> 37:      r[5] 2021-04-17  non-VOC     TRUE          FALSE 1.0220832 1.0190361
#> 38:      r[6] 2021-04-24  non-VOC     TRUE          FALSE 0.9425519 0.9404806
#> 39:      r[7] 2021-05-01  non-VOC     TRUE          FALSE 0.8613230 0.8626776
#> 40:      r[8] 2021-05-08  non-VOC     TRUE          FALSE 0.7931161 0.7943634
#> 41:      r[9] 2021-05-15  non-VOC     TRUE          FALSE 0.7447975 0.7447975
#> 42:     r[10] 2021-05-22  non-VOC     TRUE          FALSE 0.6988734 0.6999725
#> 43:     r[11] 2021-05-29  non-VOC     TRUE          FALSE 0.6772503 0.6756559
#> 44:     r[12] 2021-06-05  non-VOC     TRUE           TRUE 0.6440364 0.6430252
#> 45:     r[13] 2021-06-12  non-VOC    FALSE          FALSE 0.5991246 0.6000668
#> 46:     r[14] 2021-06-19  non-VOC    FALSE          FALSE 0.5893201 0.5888573
#> 47:     r[15] 2021-06-26  non-VOC    FALSE          FALSE 0.5755915 0.5751394
#> 48:     r[16] 2021-07-03  non-VOC    FALSE          FALSE 0.5701900 0.5701900
#> 49:     r[17] 2021-07-10  non-VOC    FALSE          FALSE 0.5626245 0.5666171
#> 50:     r[18] 2021-07-17  non-VOC    FALSE          FALSE 0.5582212 0.5639522
#>      variable       date     type observed forecast_start      mean    median
#>         sd    mad        q5       q20       q80       q95 rhat ess_bulk
#>  1: 0.0740 0.0652 1.0119349 1.0601995 1.1582712 1.2180050    1     1240
#>  2: 0.0624 0.0524 1.0030296 1.0486837 1.1277385 1.1692437    1     1970
#>  3: 0.0709 0.0535 0.9616271 1.0228062 1.1014669 1.1447003    1     1100
#>  4: 0.0665 0.0550 0.9890603 1.0289321 1.1084122 1.1655747    1     1390
#>  5: 0.0668 0.0541 0.9467083 0.9858792 1.0613664 1.1154013    1     1590
#>  6: 0.0664 0.0501 0.8756527 0.9121703 0.9757071 1.0243342    1     1830
#>  7: 0.0665 0.0526 0.7981171 0.8353299 0.8972430 0.9378978    1     2050
#>  8: 0.0639 0.0530 0.7349153 0.7709965 0.8281413 0.8640343    1     2100
#>  9: 0.0668 0.0544 0.6939488 0.7285909 0.7838235 0.8210147    1     1780
#> 10: 0.0666 0.0507 0.6573285 0.6939488 0.7424604 0.7770782    1     2050
#> 11: 0.0645 0.0546 0.6557809 0.6825926 0.7343381 0.7716025    1     1770
#> 12: 0.0654 0.0521 0.6410074 0.6698414 0.7183591 0.7565939    1     1890
#> 13: 0.0766 0.0645 0.6105296 0.6450493 0.7049398 0.7383881    1     1650
#> 14: 0.0957 0.0913 0.6270592 0.6630343 0.7471420 0.7962380    1     2690
#> 15: 0.1640 0.1490 0.5963068 0.6666910 0.8101212 0.9021914    1     1880
#> 16: 0.2230 0.1990 0.5847079 0.6793822 0.8860340 1.0256228    1     1630
#> 17: 0.2760 0.2540 0.5630667 0.6825926 0.9549260 1.1321776    1     1810
#> 18: 0.3350 0.2960 0.5379829 0.6874366 1.0184757 1.2382688    1     1650
#> 19: 0.0681 0.0603 1.4975900 1.5637260 1.6915459 1.7815789    1     1680
#> 20: 0.0685 0.0581 1.3844261 1.4432948 1.5563715 1.6366359    1     1950
#> 21: 0.0711 0.0598 1.2559059 1.3165307 1.4263844 1.4928907    1     2060
#> 22: 0.0685 0.0591 1.1537298 1.2141830 1.3123995 1.3682055    1     1900
#> 23: 0.0716 0.0607 1.0815706 1.1384217 1.2334137 1.2919373    1     1600
#> 24: 0.0710 0.0605 1.0135264 1.0684785 1.1564525 1.2084725    1     1950
#> 25: 0.0673 0.0601 0.9943510 1.0355014 1.1215529 1.1775412    1     1870
#> 26: 0.0670 0.0543 0.9450732 0.9875849 1.0632862 1.1171555    1     1880
#> 27: 0.0773 0.0653 0.8613230 0.9136048 0.9960948 1.0422764    1     1620
#> 28: 0.0950 0.0878 0.8340182 0.8832537 0.9937574 1.0582853    1     2690
#> 29: 0.1620 0.1470 0.7436281 0.8320547 1.0096318 1.1233168    1     1730
#> 30: 0.2220 0.1980 0.6842034 0.7949878 1.0356641 1.1961916    1     1520
#> 31: 0.2760 0.2550 0.6211747 0.7583794 1.0581191 1.2450981    1     1790
#> 32: 0.3360 0.2970 0.5737853 0.7331851 1.0859987 1.3186011    1     1650
#> 33: 0.0740 0.0652 1.0119349 1.0601995 1.1582712 1.2180050    1     1240
#> 34: 0.0624 0.0524 1.0030296 1.0486837 1.1277385 1.1692437    1     1970
#> 35: 0.0709 0.0535 0.9616271 1.0228062 1.1014669 1.1447003    1     1100
#> 36: 0.0665 0.0550 0.9890603 1.0289321 1.1084122 1.1655747    1     1390
#> 37: 0.0667 0.0541 0.9454446 0.9845632 1.0600330 1.1145253    1     1590
#> 38: 0.0664 0.0498 0.8735911 0.9100227 0.9731040 1.0220029    1     1830
#> 39: 0.0664 0.0523 0.7943634 0.8314012 0.8930231 0.9340003    1     2050
#> 40: 0.0639 0.0530 0.7291636 0.7649624 0.8216600 0.8565987    1     2110
#> 41: 0.0669 0.0540 0.6842034 0.7183591 0.7728160 0.8101212    1     1790
#> 42: 0.0668 0.0503 0.6405039 0.6756559 0.7240256 0.7565939    1     2040
#> 43: 0.0652 0.0561 0.6265667 0.6521840 0.7032801 0.7389685    1     1680
#> 44: 0.0669 0.0534 0.5949029 0.6206869 0.6677395 0.7043861    1     1930
#> 45: 0.0787 0.0684 0.5409499 0.5719848 0.6280454 0.6562963    1     1610
#> 46: 0.0989 0.0907 0.5209283 0.5542877 0.6265667 0.6698414    1     2520
#> 47: 0.1650 0.1490 0.4641057 0.5229788 0.6369908 0.7093848    1     1580
#> 48: 0.2260 0.2010 0.4280260 0.4981129 0.6542369 0.7512626    1     1520
#> 49: 0.2790 0.2550 0.3895133 0.4755488 0.6656442 0.7862908    1     1780
#> 50: 0.3390 0.2970 0.3600802 0.4597505 0.6831291 0.8333632    1     1650
#>         sd    mad        q5       q20       q80       q95 rhat ess_bulk
#>     ess_tail
#>  1:      859
#>  2:     1520
#>  3:     1140
#>  4:     1410
#>  5:     1460
#>  6:     1580
#>  7:     1730
#>  8:     1680
#>  9:     1600
#> 10:     1700
#> 11:     1650
#> 12:     1360
#> 13:     1640
#> 14:     1670
#> 15:     1530
#> 16:     1540
#> 17:     1620
#> 18:     1480
#> 19:     1530
#> 20:     1660
#> 21:     1380
#> 22:     1640
#> 23:     1550
#> 24:     1560
#> 25:     1460
#> 26:     1280
#> 27:     1790
#> 28:     1680
#> 29:     1590
#> 30:     1600
#> 31:     1630
#> 32:     1520
#> 33:      859
#> 34:     1520
#> 35:     1140
#> 36:     1410
#> 37:     1460
#> 38:     1580
#> 39:     1730
#> 40:     1680
#> 41:     1550
#> 42:     1730
#> 43:     1800
#> 44:     1260
#> 45:     1560
#> 46:     1650
#> 47:     1370
#> 48:     1300
#> 49:     1630
#> 50:     1570
#>     ess_tail

# model parameter summary
summary(posterior, type = "model")
#>             variable                  clean_name exponentiated         mean
#> 1: avg_voc_advantage          Average VOC effect          TRUE     1.805793
#> 2:              beta                        Beta         FALSE     0.359000
#> 3:     init_cases[1]               Initial cases          TRUE 89321.723361
#> 4:     init_cases[2]           Initial VOC cases          TRUE   219.203386
#> 5:            phi[1] Notification overdispersion         FALSE    92.000000
#> 6:            phi[2]   Sequencing overdispersion         FALSE   189.000000
#> 7:            r_init              Initial growth         FALSE     0.131000
#> 8:           r_noise                 Growth (sd)         FALSE     0.105000
#> 9:           voc_mod          Initial VOC effect          TRUE     1.805793
#>          median       sd     mad           q5          q20          q80
#> 1:     1.805793   0.0283  0.0265     1.724608     1.764734     1.847808
#> 2:     0.457000   0.4700  0.4870    -0.523000    -0.086800     0.798000
#> 3: 89321.723361   0.0713  0.0721 80821.637540 80821.637540 98715.771011
#> 4:   219.203386   0.0975  0.0971   186.792804   202.350228   237.460193
#> 5:    77.000000  62.0000 47.8000    23.600000    43.500000   130.000000
#> 6:   164.000000 119.0000 99.5000    50.400000    92.800000   266.000000
#> 7:     0.130000   0.0740  0.0652     0.015100     0.074400     0.187000
#> 8:     0.092800   0.0597  0.0576     0.031900     0.052500     0.154000
#> 9:     1.805793   0.0283  0.0265     1.724608     1.764734     1.847808
#>           q95 rhat ess_bulk ess_tail
#> 1:     1.8908    1     1880     1410
#> 2:     0.9490    1      616     1090
#> 3: 98715.7710    1     1720     1260
#> 4:   257.2376    1     2220     1480
#> 5:   212.0000    1      809     1100
#> 6:   424.0000    1     2140     1190
#> 7:     0.2510    1     1240      859
#> 8:     0.2200    1      578     1160
#> 9:     1.8908    1     1880     1410

# raw posterior values
summary(posterior, type = "raw")
#>         variable      mean    median     sd    mad        q5       q20
#>   1:        lp__ -3340.000 -3.34e+03 4.5800 4.3800 -3.35e+03 -3.35e+03
#>   2:      r_init     0.131  1.30e-01 0.0740 0.0652  1.51e-02  7.44e-02
#>   3:     r_noise     0.105  9.28e-02 0.0597 0.0576  3.19e-02  5.25e-02
#>   4:        beta     0.359  4.57e-01 0.4700 0.4870 -5.23e-01 -8.68e-02
#>   5:     voc_mod     0.591  5.91e-01 0.0283 0.0265  5.45e-01  5.68e-01
#>  ---                                                                  
#> 242: log_lik[11]   -14.800 -1.47e+01 0.7130 0.6130 -1.60e+01 -1.53e+01
#> 243: log_lik[12]   -14.100 -1.40e+01 0.6710 0.6100 -1.53e+01 -1.46e+01
#> 244: log_lik[13]   -13.700 -1.36e+01 0.8350 0.7320 -1.52e+01 -1.43e+01
#> 245: log_lik[14]    -8.160 -8.06e+00 0.5840 0.4510 -9.24e+00 -8.53e+00
#> 246: log_lik[15]    -7.690 -7.56e+00 0.6060 0.4880 -8.86e+00 -8.10e+00
#>            q80       q95 rhat ess_bulk ess_tail
#>   1: -3340.000 -3330.000 1.01      546      879
#>   2:     0.187     0.251 1.00     1240      859
#>   3:     0.154     0.220 1.00      578     1160
#>   4:     0.798     0.949 1.00      616     1090
#>   5:     0.614     0.637 1.00     1880     1410
#>  ---                                           
#> 242:   -14.200   -13.800 1.00     1100     1210
#> 243:   -13.500   -13.100 1.00      909     1030
#> 244:   -13.000   -12.500 1.00     1280     1510
#> 245:    -7.730    -7.430 1.00     1290     1270
#> 246:    -7.210    -6.950 1.00      903     1490