Label the Variant of Concern

Assign a custom label to the variant of concern in the output from fv_tidy_posterior().

Usage

update_voc_label(posterior, label, target_label = "VOC")

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

label

Character string indicating the new label to use for the variant of concern.

target_label

A character string defaulting to "VOC". Indicates the current label for the variant of concern.

Value

A list of data frames as returned by `fv_tidy_posterior() but with updated labels.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")
p <- update_voc_label(p, "Delta")
summary(p, type = "cases")
#>               variable       date      type    obs observed forecast_start
#>  1:       sim_cases[1] 2021-03-20  Combined  87328     TRUE          FALSE
#>  2:       sim_cases[2] 2021-03-27  Combined 109442     TRUE          FALSE
#>  3:       sim_cases[3] 2021-04-03  Combined 117965     TRUE          FALSE
#>  4:       sim_cases[4] 2021-04-10  Combined 107223     TRUE          FALSE
#>  5:       sim_cases[5] 2021-04-17  Combined 142664     TRUE          FALSE
#>  6:       sim_cases[6] 2021-04-24  Combined 145568     TRUE          FALSE
#>  7:       sim_cases[7] 2021-05-01  Combined 131887     TRUE          FALSE
#>  8:       sim_cases[8] 2021-05-08  Combined 107141     TRUE          FALSE
#>  9:       sim_cases[9] 2021-05-15  Combined  77261     TRUE          FALSE
#> 10:      sim_cases[10] 2021-05-22  Combined  57310     TRUE          FALSE
#> 11:      sim_cases[11] 2021-05-29  Combined  33052     TRUE          FALSE
#> 12:      sim_cases[12] 2021-06-05  Combined  22631     TRUE          FALSE
#> 13:      sim_cases[13] 2021-06-12  Combined  15553     TRUE          FALSE
#> 14:      sim_cases[14] 2021-06-19  Combined   7659     TRUE          FALSE
#> 15:      sim_cases[15] 2021-06-26  Combined   5033     TRUE           TRUE
#> 16:      sim_cases[16] 2021-07-03  Combined     NA    FALSE          FALSE
#> 17:      sim_cases[17] 2021-07-10  Combined     NA    FALSE          FALSE
#> 18:      sim_cases[18] 2021-07-17  Combined     NA    FALSE          FALSE
#> 19:      sim_cases[19] 2021-07-24  Combined     NA    FALSE          FALSE
#> 20:   sim_voc_cases[1] 2021-04-17     Delta     NA     TRUE          FALSE
#> 21:   sim_voc_cases[2] 2021-04-24     Delta     NA     TRUE          FALSE
#> 22:   sim_voc_cases[3] 2021-05-01     Delta     NA     TRUE          FALSE
#> 23:   sim_voc_cases[4] 2021-05-08     Delta     NA     TRUE          FALSE
#> 24:   sim_voc_cases[5] 2021-05-15     Delta     NA     TRUE          FALSE
#> 25:   sim_voc_cases[6] 2021-05-22     Delta     NA     TRUE          FALSE
#> 26:   sim_voc_cases[7] 2021-05-29     Delta     NA     TRUE          FALSE
#> 27:   sim_voc_cases[8] 2021-06-05     Delta     NA     TRUE          FALSE
#> 28:   sim_voc_cases[9] 2021-06-12     Delta     NA     TRUE           TRUE
#> 29:  sim_voc_cases[10] 2021-06-19     Delta     NA    FALSE          FALSE
#> 30:  sim_voc_cases[11] 2021-06-26     Delta     NA    FALSE          FALSE
#> 31:  sim_voc_cases[12] 2021-07-03     Delta     NA    FALSE          FALSE
#> 32:  sim_voc_cases[13] 2021-07-10     Delta     NA    FALSE          FALSE
#> 33:  sim_voc_cases[14] 2021-07-17     Delta     NA    FALSE          FALSE
#> 34:  sim_voc_cases[15] 2021-07-24     Delta     NA    FALSE          FALSE
#> 35:  sim_nvoc_cases[1] 2021-03-20 non-Delta     NA     TRUE          FALSE
#> 36:  sim_nvoc_cases[2] 2021-03-27 non-Delta     NA     TRUE          FALSE
#> 37:  sim_nvoc_cases[3] 2021-04-03 non-Delta     NA     TRUE          FALSE
#> 38:  sim_nvoc_cases[4] 2021-04-10 non-Delta     NA     TRUE          FALSE
#> 39:  sim_nvoc_cases[5] 2021-04-17 non-Delta     NA     TRUE          FALSE
#> 40:  sim_nvoc_cases[6] 2021-04-24 non-Delta     NA     TRUE          FALSE
#> 41:  sim_nvoc_cases[7] 2021-05-01 non-Delta     NA     TRUE          FALSE
#> 42:  sim_nvoc_cases[8] 2021-05-08 non-Delta     NA     TRUE          FALSE
#> 43:  sim_nvoc_cases[9] 2021-05-15 non-Delta     NA     TRUE          FALSE
#> 44: sim_nvoc_cases[10] 2021-05-22 non-Delta     NA     TRUE          FALSE
#> 45: sim_nvoc_cases[11] 2021-05-29 non-Delta     NA     TRUE          FALSE
#> 46: sim_nvoc_cases[12] 2021-06-05 non-Delta     NA     TRUE          FALSE
#> 47: sim_nvoc_cases[13] 2021-06-12 non-Delta     NA     TRUE           TRUE
#> 48: sim_nvoc_cases[14] 2021-06-19 non-Delta     NA    FALSE          FALSE
#> 49: sim_nvoc_cases[15] 2021-06-26 non-Delta     NA    FALSE          FALSE
#> 50: sim_nvoc_cases[16] 2021-07-03 non-Delta     NA    FALSE          FALSE
#> 51: sim_nvoc_cases[17] 2021-07-10 non-Delta     NA    FALSE          FALSE
#> 52: sim_nvoc_cases[18] 2021-07-17 non-Delta     NA    FALSE          FALSE
#> 53: sim_nvoc_cases[19] 2021-07-24 non-Delta     NA    FALSE          FALSE
#>               variable       date      type    obs observed forecast_start
#>       mean median      sd     mad     q5    q20    q80    q95  rhat ess_bulk
#>  1:  90000  89200 13600.0 11800.0  70300  79500  99900 113000 1.000     1700
#>  2: 103000 103000 15300.0 13200.0  78900  91000 114000 127000 1.000     1820
#>  3: 113000 113000 16300.0 14500.0  88600 101000 126000 141000 1.000     2090
#>  4: 121000 120000 18000.0 15300.0  93300 108000 133000 152000 1.000     1320
#>  5: 133000 133000 20600.0 17500.0 102000 118000 147000 167000 1.000     2040
#>  6: 140000 139000 20900.0 18500.0 107000 123000 155000 174000 1.000     1980
#>  7: 130000 129000 19500.0 16200.0  99600 115000 143000 162000 1.000     2050
#>  8: 107000 106000 16900.0 14000.0  81900  94400 119000 135000 1.000     2090
#>  9:  79600  78700 12100.0 10500.0  61500  70100  88000  99300 1.000     1920
#> 10:  55300  55000  8030.0  6940.0  42900  49200  61000  69000 1.000     2120
#> 11:  35700  35100  5400.0  4360.0  28300  31800  39200  44900 1.000     1980
#> 12:  22900  22700  3240.0  2840.0  17900  20400  25300  28500 1.000     1810
#> 13:  14100  14100  1900.0  1660.0  11100  12700  15500  17200 1.000     1710
#> 14:   8260   8180  1120.0   937.0   6600   7410   9040  10100 1.000     2240
#> 15:   5110   5010   826.0   678.0   3980   4500   5660   6490 1.000     2070
#> 16:   3470   3340   962.0   830.0   2200   2710   4130   5160 1.000     2070
#> 17:   2660   2390  1440.0  1000.0   1150   1650   3420   5000 1.000     1950
#> 18:   2370   1840  2110.0  1180.0    591   1040   3240   5780 1.000     1990
#> 19:   2540   1490  4390.0  1260.0    285    676   3140   7810 1.000     2080
#> 20:    220    219    37.9    35.6    161    190    251    284 1.000     2160
#> 21:    413    410    72.5    69.7    302    352    469    532 1.000     1900
#> 22:    690    680   121.0   113.0    506    593    782    891 1.000     2050
#> 23:   1020   1010   186.0   164.0    764    878   1150   1350 0.999     2020
#> 24:   1370   1360   254.0   234.0    990   1160   1560   1800 0.999     2430
#> 25:   1700   1670   341.0   309.0   1210   1410   1940   2250 1.000     1790
#> 26:   1930   1900   404.0   354.0   1340   1610   2220   2620 1.000     1520
#> 27:   2130   2090   459.0   411.0   1470   1760   2450   2930 1.000     1840
#> 28:   2210   2170   476.0   441.0   1480   1820   2570   3060 1.000     1600
#> 29:   2070   2030   442.0   412.0   1420   1710   2420   2860 1.000     2270
#> 30:   1920   1880   460.0   388.0   1270   1560   2230   2670 1.000     2030
#> 31:   1800   1710   585.0   491.0   1040   1320   2190   2850 1.000     2190
#> 32:   1750   1570   977.0   681.0    723   1080   2260   3290 1.000     1970
#> 33:   1840   1420  1650.0   927.0    447    786   2490   4440 1.000     2100
#> 34:   2180   1280  3730.0  1070.0    251    571   2700   6610 1.000     2070
#> 35:  90000  89200 13600.0 11800.0  70300  79500  99900 113000 1.000     1700
#> 36: 103000 103000 15300.0 13200.0  78900  91000 114000 127000 1.000     1820
#> 37: 113000 113000 16300.0 14500.0  88600 101000 126000 141000 1.000     2090
#> 38: 121000 120000 18000.0 15300.0  93300 108000 133000 152000 1.000     1320
#> 39: 133000 132000 20600.0 17500.0 102000 118000 147000 167000 1.000     2040
#> 40: 139000 139000 20900.0 18400.0 107000 123000 154000 174000 1.000     1970
#> 41: 129000 128000 19500.0 16200.0  99000 115000 142000 161000 1.000     2050
#> 42: 106000 105000 16800.0 14100.0  81000  93400 118000 134000 1.000     2080
#> 43:  78200  77400 12000.0 10400.0  60300  68700  86600  97900 1.000     1920
#> 44:  53600  53400  7960.0  6830.0  41300  47600  59200  67300 1.000     2120
#> 45:  33800  33300  5310.0  4330.0  26500  29900  37300  43100 1.000     1990
#> 46:  20700  20600  3160.0  2810.0  16000  18200  23100  26300 1.000     1870
#> 47:  11900  11900  1830.0  1620.0   9070  10600  13300  14900 1.000     1760
#> 48:   6190   6120  1030.0   892.0   4690   5380   6940   7910 1.000     2310
#> 49:   3190   3120   648.0   566.0   2280   2680   3640   4340 1.000     2160
#> 50:   1670   1590   525.0   449.0    964   1260   2050   2620 1.000     1800
#> 51:    910    808   538.0   368.0    354    542   1180   1780 1.000     1790
#> 52:    534    402   515.0   265.0    118    226    725   1340 1.000     1810
#> 53:    361    204   710.0   176.0     37     91    444   1090 1.000     1910
#>       mean median      sd     mad     q5    q20    q80    q95  rhat ess_bulk
#>     ess_tail
#>  1:     1610
#>  2:     1620
#>  3:     1900
#>  4:     1550
#>  5:     1710
#>  6:     1730
#>  7:     1750
#>  8:     1890
#>  9:     1950
#> 10:     1800
#> 11:     1700
#> 12:     1530
#> 13:     1730
#> 14:     1780
#> 15:     1820
#> 16:     1740
#> 17:     1590
#> 18:     1530
#> 19:     1340
#> 20:     1750
#> 21:     1850
#> 22:     1810
#> 23:     1620
#> 24:     1940
#> 25:     1930
#> 26:     1480
#> 27:     1670
#> 28:     1480
#> 29:     1750
#> 30:     1920
#> 31:     1710
#> 32:     1700
#> 33:     1580
#> 34:     1540
#> 35:     1610
#> 36:     1620
#> 37:     1900
#> 38:     1550
#> 39:     1710
#> 40:     1730
#> 41:     1750
#> 42:     1890
#> 43:     1990
#> 44:     1810
#> 45:     1690
#> 46:     1530
#> 47:     1780
#> 48:     1930
#> 49:     1380
#> 50:     1810
#> 51:     1390
#> 52:     1430
#> 53:     1400
#>     ess_tail
summary(p, type = "model")
#>             variable                  clean_name exponentiated         mean
#> 1: avg_voc_advantage        Average Delta effect         FALSE     0.589000
#> 2:              beta                        Beta         FALSE     0.203000
#> 3:     init_cases[1]               Initial cases          TRUE 89321.723361
#> 4:     init_cases[2]         Initial Delta cases          TRUE   219.203386
#> 5:            phi[1] Notification overdispersion         FALSE   103.000000
#> 6:            phi[2]   Sequencing overdispersion         FALSE   188.000000
#> 7:            r_init              Initial growth         FALSE     0.135000
#> 8:           r_scale                 Growth (sd)         FALSE     0.122000
#> 9:           voc_mod        Initial Delta effect          TRUE     1.802185
#>          median       sd     mad           q5          q20          q80
#> 1:     0.588000   0.0287  0.0273     0.544000     0.566000     0.611000
#> 2:     0.209000   0.4080  0.4480    -0.499000    -0.160000     0.582000
#> 3: 89321.723361   0.0670  0.0652 80821.637540 80821.637540 98715.771011
#> 4:   219.203386   0.0982  0.0948   186.792804   202.350228   237.460193
#> 5:    79.100000 121.0000 52.4000    23.500000    44.200000   137.000000
#> 6:   165.000000 111.0000 94.0000    55.400000    95.000000   266.000000
#> 7:     0.137000   0.0759  0.0750     0.010000     0.072800     0.197000
#> 8:     0.117000   0.0577  0.0573     0.040300     0.070800     0.170000
#> 9:     1.800384   0.0287  0.0273     1.722885     1.761208     1.842273
#>            q95 rhat ess_bulk ess_tail
#> 1:     0.63600    1     1670     1510
#> 2:     0.85300    1      892     1010
#> 3: 98715.77101    1     1570     1320
#> 4:   259.82284    1     2530     1550
#> 5:   232.00000    1      675      729
#> 6:   409.00000    1     2540     1510
#> 7:     0.25600    1      952     1190
#> 8:     0.22100    1      749      773
#> 9:     1.88891    1     1670     1510