Label the Variant of Concern

Assign a custom label to the variant of concern in the output from fv_tidy_posterior().

Usage

update_voc_label(posterior, label, target_label = "VOC")

Arguments

posterior

A dataframe of posterior output as produced by fv_tidy_posterior(). For forecast dates to be extracted data with value_type == "cases" must be present.

label

Character string indicating the new label to use for the variant of concern.

target_label

A character string defaulting to "VOC". Indicates the current label for the variant of concern.

Value

A list of data frames as returned by `fv_tidy_posterior() but with updated labels.

See also

Examples

p <- fv_example(strains = 2, type = "posterior")
p <- update_voc_label(p, "Delta")
summary(p, type = "cases")
#>               variable       date      type    obs observed forecast_start
#>  1:       sim_cases[1] 2021-03-20  Combined  87328     TRUE          FALSE
#>  2:       sim_cases[2] 2021-03-27  Combined 109442     TRUE          FALSE
#>  3:       sim_cases[3] 2021-04-03  Combined 117965     TRUE          FALSE
#>  4:       sim_cases[4] 2021-04-10  Combined 107223     TRUE          FALSE
#>  5:       sim_cases[5] 2021-04-17  Combined 142664     TRUE          FALSE
#>  6:       sim_cases[6] 2021-04-24  Combined 145568     TRUE          FALSE
#>  7:       sim_cases[7] 2021-05-01  Combined 131887     TRUE          FALSE
#>  8:       sim_cases[8] 2021-05-08  Combined 107141     TRUE          FALSE
#>  9:       sim_cases[9] 2021-05-15  Combined  77261     TRUE          FALSE
#> 10:      sim_cases[10] 2021-05-22  Combined  57310     TRUE          FALSE
#> 11:      sim_cases[11] 2021-05-29  Combined  33052     TRUE          FALSE
#> 12:      sim_cases[12] 2021-06-05  Combined  22631     TRUE          FALSE
#> 13:      sim_cases[13] 2021-06-12  Combined  15553     TRUE          FALSE
#> 14:      sim_cases[14] 2021-06-19  Combined   7659     TRUE          FALSE
#> 15:      sim_cases[15] 2021-06-26  Combined   5033     TRUE           TRUE
#> 16:      sim_cases[16] 2021-07-03  Combined     NA    FALSE          FALSE
#> 17:      sim_cases[17] 2021-07-10  Combined     NA    FALSE          FALSE
#> 18:      sim_cases[18] 2021-07-17  Combined     NA    FALSE          FALSE
#> 19:      sim_cases[19] 2021-07-24  Combined     NA    FALSE          FALSE
#> 20:   sim_voc_cases[1] 2021-04-17     Delta     NA     TRUE          FALSE
#> 21:   sim_voc_cases[2] 2021-04-24     Delta     NA     TRUE          FALSE
#> 22:   sim_voc_cases[3] 2021-05-01     Delta     NA     TRUE          FALSE
#> 23:   sim_voc_cases[4] 2021-05-08     Delta     NA     TRUE          FALSE
#> 24:   sim_voc_cases[5] 2021-05-15     Delta     NA     TRUE          FALSE
#> 25:   sim_voc_cases[6] 2021-05-22     Delta     NA     TRUE          FALSE
#> 26:   sim_voc_cases[7] 2021-05-29     Delta     NA     TRUE          FALSE
#> 27:   sim_voc_cases[8] 2021-06-05     Delta     NA     TRUE          FALSE
#> 28:   sim_voc_cases[9] 2021-06-12     Delta     NA     TRUE           TRUE
#> 29:  sim_voc_cases[10] 2021-06-19     Delta     NA    FALSE          FALSE
#> 30:  sim_voc_cases[11] 2021-06-26     Delta     NA    FALSE          FALSE
#> 31:  sim_voc_cases[12] 2021-07-03     Delta     NA    FALSE          FALSE
#> 32:  sim_voc_cases[13] 2021-07-10     Delta     NA    FALSE          FALSE
#> 33:  sim_voc_cases[14] 2021-07-17     Delta     NA    FALSE          FALSE
#> 34:  sim_voc_cases[15] 2021-07-24     Delta     NA    FALSE          FALSE
#> 35:  sim_nvoc_cases[1] 2021-03-20 non-Delta     NA     TRUE          FALSE
#> 36:  sim_nvoc_cases[2] 2021-03-27 non-Delta     NA     TRUE          FALSE
#> 37:  sim_nvoc_cases[3] 2021-04-03 non-Delta     NA     TRUE          FALSE
#> 38:  sim_nvoc_cases[4] 2021-04-10 non-Delta     NA     TRUE          FALSE
#> 39:  sim_nvoc_cases[5] 2021-04-17 non-Delta     NA     TRUE          FALSE
#> 40:  sim_nvoc_cases[6] 2021-04-24 non-Delta     NA     TRUE          FALSE
#> 41:  sim_nvoc_cases[7] 2021-05-01 non-Delta     NA     TRUE          FALSE
#> 42:  sim_nvoc_cases[8] 2021-05-08 non-Delta     NA     TRUE          FALSE
#> 43:  sim_nvoc_cases[9] 2021-05-15 non-Delta     NA     TRUE          FALSE
#> 44: sim_nvoc_cases[10] 2021-05-22 non-Delta     NA     TRUE          FALSE
#> 45: sim_nvoc_cases[11] 2021-05-29 non-Delta     NA     TRUE          FALSE
#> 46: sim_nvoc_cases[12] 2021-06-05 non-Delta     NA     TRUE          FALSE
#> 47: sim_nvoc_cases[13] 2021-06-12 non-Delta     NA     TRUE           TRUE
#> 48: sim_nvoc_cases[14] 2021-06-19 non-Delta     NA    FALSE          FALSE
#> 49: sim_nvoc_cases[15] 2021-06-26 non-Delta     NA    FALSE          FALSE
#> 50: sim_nvoc_cases[16] 2021-07-03 non-Delta     NA    FALSE          FALSE
#> 51: sim_nvoc_cases[17] 2021-07-10 non-Delta     NA    FALSE          FALSE
#> 52: sim_nvoc_cases[18] 2021-07-17 non-Delta     NA    FALSE          FALSE
#> 53: sim_nvoc_cases[19] 2021-07-24 non-Delta     NA    FALSE          FALSE
#>               variable       date      type    obs observed forecast_start
#>       mean median      sd     mad     q5    q20    q80    q95 rhat ess_bulk
#>  1:  90100  89500 13400.0 12100.0  69200  79700 100000 112000    1     2060
#>  2: 102000 102000 14700.0 13100.0  78800  91000 113000 126000    1     2130
#>  3: 114000 113000 17600.0 15200.0  86500 100000 126000 143000    1     1980
#>  4: 123000 122000 19100.0 16700.0  93100 109000 137000 154000    1     1820
#>  5: 133000 133000 19900.0 17400.0 104000 118000 148000 164000    1     1920
#>  6: 138000 137000 20400.0 18200.0 106000 122000 153000 172000    1     1910
#>  7: 128000 127000 19100.0 16700.0  98800 113000 142000 160000    1     1860
#>  8: 106000 105000 16300.0 13700.0  82400  94800 117000 132000    1     2090
#>  9:  79600  78800 11800.0 10600.0  62700  70300  88100  99600    1     1880
#> 10:  55700  55200  8510.0  6970.0  43400  49400  61500  69900    1     1890
#> 11:  35800  35500  5540.0  4500.0  27800  31900  39600  45200    1     1870
#> 12:  22900  22700  3230.0  2800.0  17900  20400  25200  28100    1     2200
#> 13:  14200  14000  1910.0  1690.0  11400  12600  15600  17300    1     2160
#> 14:   8310   8240  1130.0   936.0   6700   7440   9080  10200    1     2340
#> 15:   5140   5060   822.0   692.0   4050   4520   5680   6470    1     2140
#> 16:   3400   3300   900.0   789.0   2170   2690   4020   5010    1     2050
#> 17:   2510   2300  1160.0   905.0   1120   1630   3210   4580    1     1780
#> 18:   2130   1710  2140.0  1030.0    551    999   2880   4920    1     1810
#> 19:   2160   1340  6260.0  1110.0    256    611   2850   5870    1     1710
#> 20:    221    218    40.0    37.1    162    189    252    288    1     2010
#> 21:    410    408    71.1    68.2    301    351    469    528    1     1800
#> 22:    684    673   124.0   108.0    507    588    772    900    1     1970
#> 23:   1030   1010   194.0   175.0    742    869   1170   1360    1     2290
#> 24:   1380   1350   269.0   234.0    991   1170   1580   1830    1     2270
#> 25:   1720   1680   330.0   307.0   1230   1450   1960   2290    1     2280
#> 26:   1960   1920   413.0   354.0   1390   1650   2250   2670    1     2300
#> 27:   2170   2140   459.0   411.0   1500   1810   2500   2980    1     2250
#> 28:   2230   2200   490.0   442.0   1510   1830   2580   3060    1     2130
#> 29:   2110   2080   490.0   451.0   1420   1710   2460   2950    1     2270
#> 30:   1950   1910   464.0   422.0   1300   1570   2290   2770    1     2250
#> 31:   1780   1700   575.0   497.0   1040   1320   2170   2810    1     2210
#> 32:   1670   1520   844.0   630.0    703   1050   2160   3090    1     2010
#> 33:   1660   1320  1760.0   824.0    417    744   2250   3840    1     1850
#> 34:   1860   1150  5590.0   959.0    222    522   2450   5170    1     1760
#> 35:  90100  89500 13400.0 12100.0  69200  79700 100000 112000    1     2060
#> 36: 102000 102000 14700.0 13100.0  78800  91000 113000 126000    1     2130
#> 37: 114000 113000 17600.0 15200.0  86500 100000 126000 143000    1     1980
#> 38: 123000 122000 19100.0 16700.0  93100 109000 137000 154000    1     1820
#> 39: 133000 133000 19900.0 17400.0 103000 118000 147000 164000    1     1920
#> 40: 138000 137000 20400.0 18100.0 105000 122000 153000 172000    1     1910
#> 41: 127000 127000 19100.0 16600.0  98200 113000 141000 159000    1     1860
#> 42: 105000 105000 16200.0 13600.0  81500  93700 116000 131000    1     2090
#> 43:  78300  77400 11700.0 10500.0  61300  69000  86700  97800    1     1880
#> 44:  54000  53500  8450.0  6980.0  41700  47700  59700  67900    1     1890
#> 45:  33900  33500  5430.0  4340.0  26000  30100  37600  43100    1     1870
#> 46:  20700  20600  3130.0  2630.0  15900  18300  23000  25800    1     2180
#> 47:  11900  11800  1830.0  1630.0   9290  10500  13300  15100    1     2200
#> 48:   6200   6130  1040.0   901.0   4680   5400   6940   7860    1     2220
#> 49:   3190   3130   664.0   544.0   2290   2700   3620   4290    1     1950
#> 50:   1620   1550   498.0   423.0    958   1240   1960   2490    1     1950
#> 51:    837    753   415.0   322.0    351    517   1090   1580    1     1850
#> 52:    467    368   432.0   232.0    113    212    641   1080    1     1720
#> 53:    296    177   709.0   151.0     32     79    381    827    1     1510
#>       mean median      sd     mad     q5    q20    q80    q95 rhat ess_bulk
#>     ess_tail
#>  1:     1860
#>  2:     1850
#>  3:     1700
#>  4:     1760
#>  5:     1350
#>  6:     1860
#>  7:     1780
#>  8:     1820
#>  9:     1750
#> 10:     1650
#> 11:     1700
#> 12:     1850
#> 13:     1780
#> 14:     1850
#> 15:     1750
#> 16:     1580
#> 17:     1620
#> 18:     1460
#> 19:     1720
#> 20:     1850
#> 21:     1750
#> 22:     1770
#> 23:     1850
#> 24:     1820
#> 25:     1550
#> 26:     1650
#> 27:     1720
#> 28:     1660
#> 29:     1660
#> 30:     1630
#> 31:     1450
#> 32:     1640
#> 33:     1610
#> 34:     1720
#> 35:     1860
#> 36:     1850
#> 37:     1700
#> 38:     1760
#> 39:     1350
#> 40:     1790
#> 41:     1780
#> 42:     1820
#> 43:     1780
#> 44:     1670
#> 45:     1670
#> 46:     1880
#> 47:     1850
#> 48:     1700
#> 49:     1420
#> 50:     1420
#> 51:     1720
#> 52:     1500
#> 53:     1490
#>     ess_tail
summary(p, type = "model")
#>             variable                  clean_name exponentiated         mean
#> 1: avg_voc_advantage        Average Delta effect          TRUE     1.805793
#> 2:              beta                        Beta         FALSE     0.359000
#> 3:     init_cases[1]               Initial cases          TRUE 89321.723361
#> 4:     init_cases[2]         Initial Delta cases          TRUE   219.203386
#> 5:            phi[1] Notification overdispersion         FALSE    92.000000
#> 6:            phi[2]   Sequencing overdispersion         FALSE   189.000000
#> 7:            r_init              Initial growth         FALSE     0.131000
#> 8:           r_noise                 Growth (sd)         FALSE     0.105000
#> 9:           voc_mod        Initial Delta effect          TRUE     1.805793
#>          median       sd     mad           q5          q20          q80
#> 1:     1.805793   0.0283  0.0265     1.724608     1.764734     1.847808
#> 2:     0.457000   0.4700  0.4870    -0.523000    -0.086800     0.798000
#> 3: 89321.723361   0.0713  0.0721 80821.637540 80821.637540 98715.771011
#> 4:   219.203386   0.0975  0.0971   186.792804   202.350228   237.460193
#> 5:    77.000000  62.0000 47.8000    23.600000    43.500000   130.000000
#> 6:   164.000000 119.0000 99.5000    50.400000    92.800000   266.000000
#> 7:     0.130000   0.0740  0.0652     0.015100     0.074400     0.187000
#> 8:     0.092800   0.0597  0.0576     0.031900     0.052500     0.154000
#> 9:     1.805793   0.0283  0.0265     1.724608     1.764734     1.847808
#>           q95 rhat ess_bulk ess_tail
#> 1:     1.8908    1     1880     1410
#> 2:     0.9490    1      616     1090
#> 3: 98715.7710    1     1720     1260
#> 4:   257.2376    1     2220     1480
#> 5:   212.0000    1      809     1100
#> 6:   424.0000    1     2140     1190
#> 7:     0.2510    1     1240      859
#> 8:     0.2200    1      578     1160
#> 9:     1.8908    1     1880     1410